% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{add_nuclear_industry}
\alias{add_nuclear_industry}
\title{Adds a nuclear industry}
\usage{
add_nuclear_industry(
  .tidy_iea_df,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  method = IEATools::iea_cols$method,
  ledger_side = IEATools::iea_cols$ledger_side,
  last_stage = IEATools::iea_cols$last_stage,
  energy_type = IEATools::iea_cols$energy_type,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  nuclear_industry = IEATools::eiou_flows$nuclear_industry,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  autoproducer_elect = IEATools::main_act_plants$autoprod_elect_plants,
  autoproducer_chp = IEATools::transformation_processes$autoproducer_CHP_plants,
  nuclear = "Nuclear",
  electricity = "Electricity",
  heat = "Heat",
  negzeropos = ".negzeropos",
  share_elect_output_From_Func = ".share_elect_output_From_Func",
  Electricity_Nuclear = "Electricity_Nuclear",
  Heat_Nuclear = "Heat_Nuclear",
  ratio_output_to_nuclear_fuel = 0.33
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow_aggregation_point}.}

\item{flow}{The name of the flow column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{e_dot}{The name of the energy column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$e_dot}.}

\item{product}{The name of the product column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$product}.}

\item{method}{The name of the method column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$method}.}

\item{ledger_side}{The name of the ledger side column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$ledger_side}.}

\item{last_stage}{The name of the last stage column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$last_stage}.}

\item{energy_type}{The name of the energy type column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$energy_type}.}

\item{country}{The name of the country column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$country}.}

\item{year}{The name of the year column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$year}.}

\item{unit}{The name of the unit column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$unit}.}

\item{eiou}{A string identifying the energy industry own use in the \code{flow_aggregation_point} column in the \code{.tidy_iea_df}.
Default is \code{IEATools::aggregation_flows$energy_industry_own_use}.}

\item{transformation_processes}{A string identifying the transformation processes in the \code{flow_aggregation_point} column in the \code{.tidy_iea_df}.
Default is \code{IEATools::aggregation_flows$transformation_processes}.}

\item{nuclear_industry}{A string identifying "Nuclear industry" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::eiou_flows$nuclear_industry}.}

\item{main_act_producer_elect}{A string identifying "Main activity producer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_elect_plants}.}

\item{main_act_producer_chp}{A string identifying "Main activity producer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_chp_plants}.}

\item{autoproducer_elect}{A string identifying "Autoproducer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$autoprod_elect_plants}.}

\item{autoproducer_chp}{A string identifying "Autoproducer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::transformation_processes$autoproducer_CHP_plants}.}

\item{nuclear}{A string identifying the "Nuclear" product in the \code{product} column of the \code{tidy_iea_df}.
Default is "Nuclear".}

\item{electricity}{A string identifying the "Electricity" product in the \code{product} column of the \code{tidy_iea_df}.
Default is "Electricity".}

\item{heat}{A string identifying the "Heat" product in the \code{product} column of the \code{tidy_iea_df}.
Default is "Heat".}

\item{negzeropos}{The name of a temporary column added to the data frame.
Default is ".negzeropos".}

\item{share_elect_output_From_Func}{A temporary column added to the data frame.
Default is ".share_elect_output_From_Func".}

\item{Electricity_Nuclear}{A temporary column and product name added to the data frame, which identifies the production of electricity by nuclear plants.
Default is "Electricity_Nuclear".}

\item{Heat_Nuclear}{A temporary column and product name added to the data frame, which identifies the production of heat by nuclear plants.
Default is "Heat_Nuclear".}

\item{ratio_output_to_nuclear_fuel}{A parameter that describes the correspondance between input of nuclear fuel and output of electricity and/or heat.
The IEA World Energy Extended Balances state that the value adopted in the balances is 0.33, which is therefore
the default value of the parameter.}
}
\value{
A modified version of the \code{.tidy_iea_df}, with a nuclear industry added as an additional transformation process.
}
\description{
In the IEA World Energy Extended Balances, there is a "Nuclear industry" Energy industry own use flow,
but there is no "Nuclear industry" in transformation processes flows,
which prevents from defining a nuclear industry in the PSUT.
However, using the World Energy Extended Balances documentation, one can deduce from the amount of nuclear fuel used
by "Main activity producer electricity plants" and "Main activity producer CHP plants"
the energy transformation due to the nuclear industry. This is what this function does.
The function is called within the \code{specify_all()} function.
}
\details{
The World Energy Extended Balances documentation states that "The primary energy equivalent of nuclear electricity is
calculated from the gross generation by assuming a 33\% conversion efficiency. The calculation to be carried out
is the following: gross electricity generation in TWh x0.086 / 0.33 = primary energy equivalent in Mtoe."

Hence this function does the following:
\itemize{
\item the Nuclear fuel consumed by Main activity producer electricity & heat plants is ascribed to nuclear industry plants;
\item the output of Main activity producer electricity and heat plants to be directed to nuclear plants is determined by
multiplying their nuclear fuel consumption per 0.33. In the case of CHP plants, that output is divided into
heat and electricity according to the shares of output of each of these two products;
\item the output ascribed to nuclear plants is subtracted from Main activity producer electricity and heat plants.
}
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  add_nuclear_industry()
}
