% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_to_tidy.R
\name{add_production_details}
\alias{add_production_details}
\title{Add products to production industry names}
\usage{
add_production_details(.iea_df, flow = "Flow",
  production = "Production", product = "Product")
}
\arguments{
\item{.iea_df}{a IEA data frame whose columns have been renamed by \code{\link[=rename_iea_df_cols]{rename_iea_df_cols()}}}

\item{flow}{the name of the flow column in \code{.iea_df}.  Default is "\code{Flow}".}

\item{production}{a string that identifies production in the flow column. Default is "\code{Production}".}

\item{product}{the name of the product column in \code{.iea_df}.  Default is "\code{Product}".}
}
\value{
.iea_df with detailed production industries
}
\description{
The IEA specify all extraction processes as "\code{Production}".
When combined with an input-output view of the world
within the Physical Supply-Use Table (PSUT) framework,
production of any one resource implies production of all other resources.
That's clearly non-physical,
as production of some Crude oil does not imply harvesting of Solar energy.
The solution to this problem is to rename the \code{Flow} values to be
\code{Production (Product)}.
For example, the industry that produces Crude oil will be renamed to
\code{Production (Crude oil)}.
}
\details{
If the suffix of the value of the \code{Flow} column is already "\code{(Product)}",
no changes to the value in the \code{Flow} column will be made.
}
