% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_psut_matnames}
\alias{add_psut_matnames}
\title{Add a column of matrix names to tidy data frame}
\usage{
add_psut_matnames(
  .tidy_iea_df,
  cols = IEATools::iea_cols,
  ledger_side = cols$ledger_side,
  flow_aggregation_point = cols$flow_aggregation_point,
  flow = cols$flow,
  product = cols$product,
  e_dot = cols$e_dot,
  supply = "Supply",
  consumption = "Consumption",
  production = "Production",
  resources = "Resources",
  eiou = "Energy industry own use",
  neg_supply_in_fd = c("Exports", "International aviation bunkers",
    "International marine bunkers", "Losses", "Statistical differences", "Stock changes"),
  matnames = "matnames",
  R = "R",
  U_excl_EIOU = "U_excl_EIOU",
  U_EIOU = "U_EIOU",
  V = "V",
  Y = "Y"
)
}
\arguments{
\item{.tidy_iea_df}{a data frame with \code{ledger_side}, \code{flow_aggregation_point}, \code{flow}, and \code{e_dot} columns.}

\item{cols}{a list containing names of columns. Default is \code{iea_cols}.}

\item{ledger_side}{the name of the column in \code{.tidy_iea_df} that contains ledger side
(a string). Default is \code{cols$ledger_side}.}

\item{flow_aggregation_point}{the name of the column in \code{.tidy_iea_df} that contains flow aggregation point information.
Default is \code{cols$flow_aggregation_point}.}

\item{flow}{the name of the column in \code{.tidy_iea_df} that contains flow information.
Default is \code{cols$flow}.}

\item{product}{the name of the column in \code{.tidy_iea_df} that contains flow information.
Default is \code{cols$product}.}

\item{e_dot}{the name of the column in \code{.tidy_iea_df} that contains energy and exergy values
(a string). Default is \code{cols$e_dot}.}

\item{supply}{the identifier for items on the supply side of the ledger (a string).
Default is "Supply".}

\item{consumption}{the identifier for items on the consumption side
of the ledger (a string). Default is "Consumption".}

\item{production}{a string identifying production in the flow column. Default is \code{Production}.}

\item{resources}{a string identifying resources in the flow column. Default is \code{Resources}.}

\item{eiou}{the identifier for items that are energy industry own use.
Default is "Energy industry own use".}

\item{neg_supply_in_fd}{identifiers for flow items that, when negative,
are entries in the final demand (\code{Y}) matrix.}

\item{matnames}{the name of the output column containing the name of the matrix
to which a row's value belongs (a string). Default is "matnames".}

\item{R}{the name for the resource matrix (a string). Default is "R".}

\item{U_excl_EIOU}{the name for the use matrix that excludes energy industry own use (a string). Default is "U_excl_EIOU".}

\item{U_EIOU}{the name for the energy industry own use matrix. Default is "U_EIOU".}

\item{V}{the name for the make matrix (a string). Default is "V".}

\item{Y}{the name for the final demand matrix (a string). Default is "Y".}
}
\value{
\code{.tidy_iea_df} with an added column \code{matnames}.
}
\description{
This function adds a column of matrix names to a tidy data frame
wherein each row of \code{.tidy_iea_df} is a single value in an energy conversion chain.
The default argument values assume that \code{.tidy_iea_df} uses IEA-style nomenclature
and terminology, although \code{.tidy_iea_df} does not necessarily need to contain IEA data.
}
\details{
In a reasonable workflow, this function would be followed by a call to
\code{add_row_col_meta()} and \code{matsindf::collapse_to_matrices()}.

This function respects groups when identifying entries in the resource matrix (\code{R}).
So be sure to group \code{.tidy_iea_df} before calling this function.

Internally, this function adds a temporary column to \code{.tidy_iea_df} called ".R".
An error will occur if \code{.tidy_iea_df} already has a column named ".R".
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  add_psut_matnames() \%>\%
  glimpse()
}
