% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregate_cols}
\alias{aggregate_cols}
\title{Aggregation columns}
\format{
A string vector with 3 entries.
\describe{
\item{aggregate_primary}{The name of the column containing aggregated primary energy.}
\item{net_aggregate_demand}{The name of the column containing aggregated net final demand energy.}
\item{gross_aggregate_demand}{The name of the column containing aggregated gross final demand energy.}
}
}
\usage{
aggregate_cols
}
\description{
A string vector containing names of aggregation columns.
}
\examples{
aggregate_cols
}
\keyword{datasets}
