% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_flows}
\alias{aggregation_flows}
\title{Aggregation flows}
\format{A string vector with 8 entries}
\usage{
aggregation_flows
}
\description{
A string vector containing names of industries whose purpose in IEA extended energy balance \code{Flow}s is to provide aggregations.
}
\examples{
aggregation_flows
}
\keyword{datasets}
