% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllIEAData_prior_to_fix}
\alias{AllIEAData_prior_to_fix}
\title{Raw International Energy Agency data}
\format{A data frame with 700,111 rows and 9 variables:
\describe{
  \item{Country}{Two-letter ISO country codes or "World" for the entire world.}
  \item{Ledger.side}{Supply or Consumption}
  \item{Flow.aggregation.point}{tells where each row should be aggregated}
  \item{Flow}{the Industry or Sector involved in this flow}
  \item{Product}{the energy product involved in this flow}
  \item{Year}{the year for this flow}
  \item{E.ktoe}{the magnitude of this energy flow}
}}
\source{
\url{http://http://www.iea.org}
}
\usage{
AllIEAData_prior_to_fix
}
\description{
A dataset containing all IEA data before it has been "fixed"
to the best of our ability.  
See \code{IEAData::AllIEAData} for the same data after it has been "fixed."
}
\keyword{datasets}
