% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{arrange_iea_fu_allocation_template}
\alias{arrange_iea_fu_allocation_template}
\title{Arrange the columns of an fu_allocation template}
\usage{
arrange_iea_fu_allocation_template(.fu_allocation_template,
  rowcol = c("both", "row", "col"), ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  ef_product = "Ef.product", machine = "Machine",
  eu_product = "Eu.product", destination = "Destination",
  unit = "Unit", fap_dest_order = IEATools::fap_flow_iea_order,
  ef_product_order = IEATools::product_iea_order,
  quantity = "Quantity", maximum_values = "Maximum.values",
  .temp_sort = ".fap_flow")
}
\arguments{
\item{.fu_allocation_template}{the final-to-useful allocation template created by \code{fu_allocation_template()}}

\item{rowcol}{one of "\code{both}", "\code{row}", or "\code{col}" to indicate whether rows, columns, or both should be arranged. Default is "\code{both}".}

\item{ledger_side}{the ledger side column in \code{.fu_allocation_template}. Default is "\code{Ledger.side}".}

\item{flow_aggregation_point}{the flow aggregation point column in \code{.fu_allocation_template}. Default is "\code{Flow.aggregation.point}".}

\item{ef_product}{the name of the final energy column in \code{.fu_allocation_template}. Default is "\code{Ef.product}".}

\item{machine}{the name of the machine columnin \code{.fu_allocation_template}. Default is "\code{Machine}".}

\item{eu_product}{the name of the useful energy product column in \code{.fu_allocation_template}. Default is "\code{Eu.product}".}

\item{destination}{the name of the destination column in \code{.fu_allocation_template}. Default is "\code{Destination}".}

\item{unit}{the name of the unit in \code{.fu_allocation_template}. Default is "\code{Unit}".}

\item{fap_dest_order}{the desired order for the combination of \code{flow_aggregation_point} and \code{destination} columns. Default is \code{IEATools::fap_flow_iea_order}.}

\item{ef_product_order}{the desired order for final energy products in \code{.fu_allocation_template}. Default is "\code{Ef.product}".}

\item{quantity}{the name of the quantity column in \code{.fu_allocation_template}. Default is "\code{Quantity}".}

\item{maximum_values}{the name of the maximum value column \code{.fu_allocation_template}. Default is "\code{Unit}".}

\item{.temp_sort}{the name of a temporary column to be added to \code{.fu_allocation_template}.
Default is "\code{.fap_flow}".
This column must not be present in \code{.fu_allocation_template}.}
}
\value{
An row- and/or column-ordered version of \code{.fu_allocation_template}
}
\description{
It is helpful to sort columns of a final-to-useful allocation template
in reasonable ways to assist analysts with final-to-useful
allocations.
This function sorts columns (left-to-right) in the order of energy flow through the energy conversion chain:
\itemize{
\item metadata columns,
\item final energy product (\code{Ef.product}).
\item Machine (the final-to-useful transformation process),
\item useful energy product (\code{Eu.product}),
\item destination where the useful energy now flows,
\item years (in columns), and
\item allocations (C_x rows).
}
}
\examples{
Template <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\%
  fu_allocation_template()
Template
Template \%>\% 
  arrange_iea_fu_allocation_template()
}
