% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{arrange_iea_fu_allocation_template_cols}
\alias{arrange_iea_fu_allocation_template_cols}
\title{Arrange the columns of an fu_template}
\usage{
arrange_iea_fu_allocation_template_cols(.fu_allocation_template,
  ef_product = "Ef.product", machine = "Machine",
  eu_product = "Eu.product", destination = "Destination",
  quantity = "Quantity", maximum_values = "Maximum.values")
}
\arguments{
\item{ef_product}{the name of the final energy column. Default is "\code{Ef.product}".}

\item{machine}{the name of the machine column. Default is "\code{Machine}".}

\item{eu_product}{the name of the useful energy product column. Default is "\code{Eu.product}".}

\item{.fu_template}{the final-to-useful allocation template created by \code{fu_allocation_template()}}
}
\value{
An column-ordered data frame with columns and rows in (more-or-less) IEA order
}
\description{
It is helpful to sort rows and columns of a final-to-useful allocation template
in reasonable ways to assist analysts with final-to-useful
allocations.
This function uses a left-to-right approach:
metadata columns proceed from
final energy product (\code{Ef.product}) to
Machine (the final-to-useful transformation process) to
useful energy product (\code{Eu.product}) to
destination where the useful energy now flows.
Further to the right, years (in columns) and
allocations (C_x rows) can be found.
}
