% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arrow_notation}
\alias{arrow_notation}
\title{Arrow notation}
\format{
A character vector with 4 entries
}
\usage{
arrow_notation
}
\description{
A character vector containing the arrow notation used in the IEATools package.
}
\examples{
arrow_notation
}
\keyword{datasets}
