% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{augment_iea_df}
\alias{augment_iea_df}
\title{Augment an IEA data frame}
\usage{
augment_iea_df(.iea_df, ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  country = "COUNTRY", flow = "FLOW", losses = "Losses",
  supply = "Supply", consumption = "Consumption",
  tpes_flows = c("Production", "Imports", "Exports",
  "International marine bunkers", "International aviation bunkers",
  "Stock changes"), tpes = "Total primary energy supply",
  tfc_compare_flows = c("Total primary energy supply", "Transfers",
  "Statistical differences", "Transformation processes",
  "Energy industry own use", "Losses"), tfc_compare = "TFC compare",
  tp_flows_suffix = "(transf.)",
  nstp_flows_suffix = "(transformation)",
  transformation_processes = "Transformation processes",
  eiou_flows_suffix = "(energy)", eiou = "Energy industry own use",
  tfc_flows = c("Industry", "Transport", "Other", "Non-energy use"),
  tfc = "Total final consumption", industry_flows = c("Iron and steel",
  "Chemical and petrochemical", "Non-ferrous metals",
  "Non-metallic minerals", "Transport equipment", "Machinery",
  "Mining and quarrying", "Food and tobacco", "Paper, pulp and print",
  "Wood and wood products", "Construction", "Textile and leather",
  "Non-specified (industry)"), industry = "Industry",
  transport_flows = c("World aviation bunkers", "Domestic aviation",
  "Road", "Rail", "Pipeline transport", "World marine bunkers",
  "Domestic navigation", "Non-specified (transport)"),
  transport = "Transport", other_flows = c("Residential",
  "Commercial and public services", "Agriculture/forestry", "Fishing",
  "Non-specified (other)"), other = "Other",
  non_energy_prefix = "Non-energy use", non_energy = "Non-energy use",
  electricity_output_flows_prefix = "Electricity output (GWh)-",
  electricity_output = "Electricity output (GWh)",
  heat_output_flows_prefix = "Heat output-",
  heat_output = "Heat output", .rownum = ".rownum")
}
\arguments{
\item{.iea_df}{a data frame produced by the \code{\link[=iea_df]{iea_df()}} function}

\item{ledger_side}{the name of the ledger side column. Default is "\code{Ledger.side}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column. Default is "\code{Flow.aggregation.point}".}

\item{country}{the name of the country column in \code{.iea_df}. Default is "\code{COUNTRY}".}

\item{flow}{the name of the flow column in \code{.iea_df}. Default is "\code{FLOW}".}

\item{losses}{the string that indicates losses in the \code{flow} column. Default is "\code{Losses}".}

\item{supply}{the string that indicates supply in the \code{ledger_side} column. Default is "\code{Supply}".}

\item{consumption}{the string that indicates consumption in the \code{ledger_side} column. Default is "\code{Consumption}".}

\item{.rownum}{the name of a column created (and destroyed) internally by this function.
The \code{.rownum} column temporarily holds row numbers for internal calculations.
The \code{.rownum} column is deleted before returning.}

\item{country}{the name of the country column in \code{.iea_df}. Default is "\code{COUNTRY}".}
}
\value{
\code{.iea_df} with additional columns named \code{ledger_side} and \code{flow_aggregation_point}
}
\description{
This function prepares an IEA data frame created by \code{\link[=iea_df]{iea_df()}} for use in R.
}
\details{
This function solves several problems.
The first problem is that metadata in the \code{COUNTRY}, \code{FLOW}, and \code{PRODUCT}
collumns of an IEA data table are not unique.
To solve this problem, two additional columns are added: \code{Ledger.side} and \code{Flow.aggregation.point}.
\code{Ledger.side} can be one of "\code{Supply}" or "\code{Consumption}", corresponding to the top or bottom of the IEA's tables, respectively.
\code{Flow.aggregation.point} indicates the next level of aggregation for these data.
\code{Flow.aggregation.point} can be one of
"\code{Total primary energy supply}", "\code{Transformation processes}", "\code{Energy industry own use}", or "\code{TFC compare}"
on the \code{Supply} side of the ledger.
On the \code{Consumption} side of the ledger, \code{Flow.aggregation.point} can be one of
"\code{Industry}", "\code{Transport}", "\code{Other}", or "\code{Non-energy use}".
The second problem is that the countries are given by their (long) full name.
To solve this problem, the country column is filled with 2-letter ISO abbreviations.
}
\examples{
iea_df(text = ",,TIME,1960,1961\\nCOUNTRY,FLOW,PRODUCT\\nWorld,Production,Hard coal,42,43") \%>\% 
  augment_iea_df()
}
