% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biofuels_and_waste_products}
\alias{biofuels_and_waste_products}
\title{Biofuel and waste products}
\format{
A string vector with 11 entries
\describe{
\item{industrial_waste}{The string identifier for Industrial waste products.}
\item{municipal_waste_renewable}{The string identifier for renewable municipal waste products.}
\item{municipal_waste_nonrenewable}{The string identifier for non-renewable municipal waste products.}
\item{primary_solid_biofuels}{The string identifier for Primary solid biofuels.}
\item{biogases}{The string identifier for Biogases.}
\item{biogasoline}{The string identifier for Biogasoline.}
\item{biodiesels}{The string identifier for Biodiesels.}
\item{bio_jet_kerosene}{The string identifier for Bio jet kerosene.}
\item{other_liquid_biofuels}{The string identifier for Other liquid biofuels.}
\item{non_specified_primary_biofuels_and_waste}{The string identifier for Non-sepcified primary biofuels and waste.}
\item{charcoal}{The string identifier for Charcoal.}
}
}
\usage{
biofuels_and_waste_products
}
\description{
A string vector containing names of products classified by the IEA as biofuel and waste products.
}
\examples{
biofuels_and_waste_products
}
\keyword{datasets}
