% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bracket_notation}
\alias{bracket_notation}
\title{Bracket notation}
\format{
A character vector with 4 entries
}
\usage{
bracket_notation
}
\description{
A character vector containing the "from" notation used in the IEATools package.
}
\examples{
bracket_notation
}
\keyword{datasets}
