% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{calc_tidy_iea_df_balances}
\alias{calc_tidy_iea_df_balances}
\title{Calculate balances on a tidy IEA data frame}
\usage{
calc_tidy_iea_df_balances(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  matnames = IEATools::mat_meta_cols$matnames,
  balancing = "balancing",
  supply_sum = "supply_sum",
  consumption_sum = "consumption_sum",
  supply_minus_consumption = "supply_minus_consumption",
  balance_OK = "balance_OK",
  err = "err",
  tol = 1e-06
)
}
\arguments{
\item{.tidy_iea_df}{an IEA-style data frame containing a \code{ledger_side}, \code{product},
\code{flow}, and energy rate (\code{e_dot}) columns along with
grouping columns, typically \code{Country}, \code{Year}, \code{Product}, etc.
a \code{Ledger.side} column.}

\item{ledger_side, flow_aggregation_point, flow, product, e_dot, unit}{See \code{IEATools::iea_cols}.}

\item{supply, consumption}{See \code{IEATools::ledger_sides}.}

\item{matnames}{See \code{IEATools::mat_meta_cols}.}

\item{balancing}{The ledger side of balancing flows, if any balancing flow has been added to the \code{.tidy_iea_df}.}

\item{supply_sum}{the name of a new column that will contain the sum of all supply for that group.
Default is "supply_sum".}

\item{consumption_sum}{the name of a new column that will contain the sum of all consumption for that group.
Default is "consumption_sum".}

\item{supply_minus_consumption}{the name of a new column that will contain the difference between supply and consumption for that group.
Default is "supply_minus_consumption".}

\item{balance_OK}{the name of a new logical column that tells whether a row's energy balance is OK.
Default is "balance_OK".}

\item{err}{the name of a new column that indicates the energy balance error for each group. Default is "err".}

\item{tol}{if the difference between supply and consumption is greater than \code{tol},
\code{balance_OK} will be set to \code{FALSE}. Default is \code{1e-6}.}
}
\value{
\code{.tidy_iea_df} with additional columns \code{supply_sum}, \code{consumption_sum}, \code{supply_minus_consumption}, \code{balance_OK}, and \code{err}.
}
\description{
It is important to know whether energy flows are balanced before
proceeding with further analyses.
This function calculates energy balances
by groups in \code{.tidy_iea_df}.
So be sure to group \code{.tidy_iea_df} by appropriate variables
before calling this function.
Grouping should \emph{definitely} be done on the \code{Product} column.
Typically, grouping is also done on
\code{Country}, \code{Method}, \code{Year}, \code{EnergyType}, \code{Last.stage}, etc. columns.
Grouping should \emph{not} be done on the \code{Ledger.side} column or the \code{Flow} column.
To test whether all balances are OK,
use the \code{tidy_iea_df_balanced()} function.
}
\details{
Supply side and consumption side energy flows are aggregated to a
\code{supply_sum} and a \code{consumption_sum} column.
There are two possibilities:
\enumerate{
\item A Product appears only on the supply side,
because it is completely transformed before reaching the consumption side of the ledger.
In this case, the \code{consumption_sum} column will have an
\code{NA} value, and the \code{supply_minus_consumption} column
will also have an \code{NA} value.
\item A Product appears on both the supply and the demand sides of the ledger
and, therefore, is not \code{NA} in the \code{consumption_sum} column and the
\code{supply_minus_consumption} column.
The column \code{balance_OK} is calculated as follows:
\item For the first situation, \code{consumption_sum} will be \code{0} (within \code{tol})
if the Product is balanced and
\code{balance_OK} will have a value of \code{TRUE}.
If not, \code{balance_OK} will have a value of \code{FALSE}.
\item In the second situation, the difference between \code{supply_sum} and \code{consumption_sum} is calculated
(\code{supply_minus_consumption}).
If the product is balanced,
\code{supply_minus_consumption} will be \code{0} (within \code{tol})
and \code{balance_OK} will be \code{TRUE}.
If not, \code{balance_OK} will be \code{FALSE}.
}
}
\examples{
library(dplyr)
Ebal <- load_tidy_iea_df() \%>\% 
  calc_tidy_iea_df_balances()
head(Ebal, 5)
}
