% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vif.R
\name{calc_vif}
\alias{calc_vif}
\title{Calculate the variance inflation factors}
\usage{
calc_vif(x)
}
\arguments{
\item{x}{matrix or data frame containing only numerical variables.}
}
\value{
The function returns a matrix with 3 columns:
\describe{
  \item{\code{Rsquared}}{}
  \item{\code{Tolerance}}{}
  \item{\code{VIF}}{}
 }
}
\description{
Calculate the variance inflation factors
}
\examples{
x <- data.frame(a = 1:20, b = 1:20*2 + rnorm(20) )
calc_vif(x)
}
