% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{carnot_efficiency}
\alias{carnot_efficiency}
\title{Calculate Carnot efficiencies from heat types}
\usage{
carnot_efficiency(heat_types, T_0 = 298.15)
}
\arguments{
\item{heat_types}{a string vector of heat types of the form "\code{HTH.600.C}"}

\item{T_0}{dead state temperature in kelvin. Default is \code{298.15} kelvin (25 C).}
}
\value{
a numeric vector of Carnot efficiencies of same length as \code{heat_types}
}
\description{
In societal exergy analysis, converting heat to exergy requires knowledge
of the temperature of that heat and application of the Carnot efficiency.
This function first converts heat types (e.g., "\code{HTH.600.C}")
to temperatures by extracting the temperature from the middle of the string,
in a unit-aware manner.
Then, the Carnot efficiency is calculated from the temperature of the heat
by applying the Carnot efficiency equation: \code{abs(1 - T_0/T)},
where T_0 and T are expected to be in kelvin units.
}
\details{
When the heat temperature is less than \code{T_0},
the Carnot efficiency is calculated as \code{1 - (heat temperature)/T_0}.

\code{T_0} can be supplied as a numeric vector of ambient temperatures of
same length as \code{heat_types}.
}
\examples{
carnot_efficiency(c("HTH.600.C", "MTH.200.C", "MTH.100.C", "LTH.20.C", "LTH.-10.C"))
carnot_efficiency("LTH.-30.F")
}
\seealso{
\code{\link[=extract_TK]{extract_TK()}}
}
