% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{check_fu_allocation_data}
\alias{check_fu_allocation_data}
\title{Check validity of a final-to-useful allocation table}
\usage{
check_fu_allocation_data(
  .fu_allocation_table,
  country = IEATools::iea_cols$country,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  machine = IEATools::template_cols$machine,
  ef_product = IEATools::template_cols$ef_product,
  eu_product = IEATools::template_cols$eu_product,
  destination = IEATools::template_cols$destination,
  non_energy_machine = "Non-energy"
)
}
\arguments{
\item{.fu_allocation_table}{The final-to-useful allocation table you want to check.}

\item{country, flow_aggregation_point}{See \code{IEATools::iea_cols}.}

\item{machine, ef_product, eu_product, destination}{See \code{IEATools::template_cols}.}

\item{non_energy_machine}{The string that identifies a Non-energy machine. Default is "Non-energy".}
}
\value{
\code{TRUE} if the checks are successfuls. An error is emitted if the checks fail.
}
\description{
Check validity of a final-to-useful allocation table
}
\examples{
load_fu_allocation_data() \%>\% 
  check_fu_allocation_data()
}
