% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{check_neu_balance}
\alias{check_neu_balance}
\title{Check Non-energy use memo balance}
\usage{
check_neu_balance(
  .iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  non_energy_flows_industry_transformation_energy =
    IEATools::non_energy_flows$non_energy_use_industry_transformation_energy,
  memo_non_energy_flows_industry = "Memo: Non-energy use in industry",
  memo_non_energy_use_in = "Memo: Non-energy use in ",
  total = IEATools::memo_aggregation_product_prefixes$total,
  .values = IEATools::template_cols$.values,
  .values_summarised = paste0(.values, "_summarised")
)
}
\arguments{
\item{.iea_df}{A data frame of IEA data,
created by \code{rename_iea_df_cols()}.}
}
\description{
For some countries and years, Non-energy use details are supplied
by "Memo:" fields.
We use those details, but only if they are balanced.
This function assess the energy balance of the
Memo: Non-energy use
fields.
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  check_neu_balance()
}
