% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{clean_iea_whitespace}
\alias{clean_iea_whitespace}
\title{Clean whitespace from Flow and Product strings}
\usage{
clean_iea_whitespace(.iea_df, flow = "Flow", product = "Product")
}
\arguments{
\item{.iea_df}{a data frame containing \code{Flow} and \code{Product} columns}

\item{flow}{the name of the flow column in \code{iea_df}. Default is "\code{Flow}".}

\item{product}{the name of the product columns in \code{iea_df}. Default is "\code{Product}".}
}
\value{
\code{.iea_df} with leading and trailing whitespace removed from \code{Flow} and \code{Product} column strings
}
\description{
Occasionally, in the IEA extended energy balance data,
extra whitespace characters are found at the beginning or end of \code{Flow} and \code{Product} strings.
This function removes all leading and trailing whitespece.
}
\examples{
data.frame(Flow = "  a flow   ", Product = "   a product   ", stringsAsFactors = FALSE) \%>\% 
  clean_iea_whitespace()
}
