% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coal_and_coal_products}
\alias{coal_and_coal_products}
\title{Coal and coal products}
\format{
A string vector with 10 entries.
\describe{
\item{primary_coal_products}{The string identifier for \code{primary_coal_products}.}
\item{patent_fuel}{The string identifier for Patent fuel.}
\item{coke_oven_coke}{The string identifier for Coke oven coke.}
\item{gas_coke}{The string identifier for Gas coke.}
\item{coal_tar}{The string identifier for Coal tar.}
\item{bkb}{The string identifier for BKB.}
\item{gas_works_gas}{The string identifier for Gas works gas.}
\item{coke_oven_gas}{The string identifier for Coke oven gas.}
\item{blast_furnace_gas}{The string identifier for Blast furnace gas.}
\item{other_recovered_gases}{The string identifier for Other recovered gases.}
}
}
\usage{
coal_and_coal_products
}
\description{
A string vector containing names of products classified by the IEA as coal and coal products.
}
\examples{
coal_and_coal_products
}
\keyword{datasets}
