% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete_eta_fu_table}
\alias{complete_eta_fu_table}
\title{Title}
\usage{
complete_eta_fu_table(
  eta_fu_table,
  exemplar_eta_fu_tables,
  fu_allocation_table,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  unit = IEATools::iea_cols$unit,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  e_dot = IEATools::iea_cols$e_dot,
  e_dot_perc = IEATools::template_cols$e_dot_perc,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  eta_fu = IEATools::template_cols$eta_fu,
  phi_u = IEATools::template_cols$phi_u,
  year = IEATools::iea_cols$year,
  quantity = IEATools::template_cols$quantity,
  maximum_values = IEATools::template_cols$maximum_values,
  eta_fu_phi_u_source = IEATools::template_cols$eta_fu_phi_u_source,
  phi_source = IEATools::template_cols$phi_source,
  .values = IEATools::template_cols$.values
)
}
\arguments{
\item{eta_fu_table}{}

\item{exemplar_eta_fu_tables}{}

\item{fu_allocation_table}{}

\item{country}{}

\item{method}{}

\item{energy_type}{}

\item{last_stage}{}

\item{unit}{}

\item{machine}{}

\item{eu_product}{}

\item{e_dot}{}

\item{e_dot_perc}{}

\item{year}{}

\item{.values}{}
}
\value{

}
\description{
Title
}
