% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{complete_fu_allocation_table}
\alias{complete_fu_allocation_table}
\title{Complete an FU Allocation table}
\usage{
complete_fu_allocation_table(
  fu_allocation_table,
  exemplar_fu_allocation_tables,
  tidy_specified_iea_data,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  year = IEATools::iea_cols$year,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  e_dot_perc = IEATools::template_cols$e_dot_perc,
  destination = IEATools::template_cols$destination,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  ef_product = IEATools::template_cols$ef_product,
  max_vals = IEATools::template_cols$maximum_values,
  quantity = IEATools::template_cols$quantity,
  .values = "values"
)
}
\arguments{
\item{fu_allocation_table}{The FU allocation table to be completed.
This data frame is probably read by \code{load_fu_allocation_data()}.
If \code{NULL}, the table will be constructed exclusively from
information available in the exemplar country tables.
Only one country is allowed in this data frame.}

\item{exemplar_fu_allocation_tables}{A list of FU Allocation tables, each probably created by \code{load_fu_allocation_data()}.
Note that each exemplar table must contain data for a single country only.
If more than one country is found, an error occurs.}

\item{tidy_specified_iea_data}{A data frame of specified IEA data in tidy format.}

\item{country, method, energy_type, last_stage, ledger_side, flow, product, unit, e_dot, year, flow_aggregation_point}{See \code{IEATools::ieacols}.}

\item{supply, consumption}{See \code{IEATools::ledger_sides}.}

\item{eiou}{See \code{IEATools::tfc_compar_flows}.}

\item{e_dot_perc, destination, machine, eu_product, ef_product, max_vals, quantity}{See \code{IEATools::template_cols}.}

\item{.values}{The name of a values column created internally. Default is "values".}
}
\value{
A completed tidy data frame containing an FU Allocation table to replace argument \code{tidy_fu_allocation_table}.
}
\description{
An FU (final-to-useful) Allocation table
tells how final energy carriers are allocated to final-to-useful machines
in each final demand sector.
A template for an FU Allocation table can be created with
\code{fu_allocation_template()}.
If the analyst does not know some FU allocations for a given country,
this function can be used to build a complete FU allocation table
by supplying allocations from any number of exemplar countries.
}
\details{
\code{fu_allocation_table} is the FU Allocation table to be completed.
Any missing information is obtained from the FU Allocation tables of the exemplar countries,
provided in the \code{exemplar_fu_allocation_tables} argument.
Each exemplar table is interrogated in order,
with data taken from the first exemplar that contains the needed information.

\code{tidy_specified_iea_data} supplies information about which data are needed.
The \code{tidy_specified_iea_data} data frame should be obtained from a call to \code{specify_all()}.

If \code{fu_allocation_table} can't be completed (because not enough information is available in
\code{exemplar_fu_allocation_tables}), an error is emitted.
}
