% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country order}
\format{
A string vector with 249 entries
}
\usage{
countries
}
\description{
A string vector containing 3-letter ISO country codes in alphabetical order.
}
\examples{
countries
}
\keyword{datasets}
