% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_concordance_cols}
\alias{country_concordance_cols}
\title{Country concordance table column names}
\format{
A string list with 2 entries.
\describe{
\item{pfu_code}{The name of a column containing 3-letter Primary-Final-Useful country codes. Normally, these codes should match the ISO 3-letter codes for each country.}
\item{iea_name}{The name of a column containing 3-letter Primary-Final-Useful country names. These names should be the same as the IEA's country names.}
}
}
\usage{
country_concordance_cols
}
\description{
A string list containing named names of columns in country concordance tables.
}
\examples{
country_concordance_cols
}
\keyword{datasets}
