% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_order}
\alias{country_order}
\title{Country order}
\format{
A string vector
}
\usage{
country_order
}
\description{
A string vector containing 3-letter ISO country codes in alphabetical order.
}
\examples{
country_order
}
\keyword{datasets}
