% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{default_aggregation_region_table_path}
\alias{default_aggregation_region_table_path}
\title{The path to the default aggregation table}
\usage{
default_aggregation_region_table_path(version = 2019)
}
\arguments{
\item{version}{the desired version (expressed as the year of IEA data release) of
the aggregation table.
Options are 2019 (default), and 2020.}
}
\value{
the path to the default aggregation table for the desired IEA data version
}
\description{
The path to the default aggregation table
}
\examples{
default_aggregation_region_table_path()     # Assumes 2019
default_aggregation_region_table_path(2019) # Same
# Returns path for default aggregation table appropriate for 2020 IEA data release
default_aggregation_region_table_path(2020) 
}
