% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distribution_industry}
\alias{distribution_industry}
\title{Distribution industry name}
\format{
A character string
\describe{
A character string containing the name of the distribution industry
}
}
\usage{
distribution_industry
}
\description{
A character string containing the name of the distribution industry
}
\examples{
distribution_industry
}
\keyword{datasets}
