% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{do_fix}
\alias{do_fix}
\title{Performs fixes to IEA data}
\usage{
do_fix(.tidy_iea_df, replacement, country, year, e_dot)
}
\arguments{
\item{.tidy_iea_df}{the tidy IEA data frame to be fixed}

\item{replacement}{a data frame containing the data that fixes the IEA data}

\item{country}{the name of the country column in \code{.tidy_iea_df} and \code{replacement}}

\item{year}{the name of the year column in \code{.tidy_iea_df} and \code{replacement}}

\item{e_dot}{the name of the energy flow rate column in \code{.tidy_iea_df} and \code{replacement}}
}
\value{
a modified version of \code{.tidy_iea_df} with \code{replacement} included, if warranted
}
\description{
This is an internal convenience function that performs fixes on IEA data
given a replacement data frame.
It makes use of the \code{replace_join()} function internally.
}
