% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eiou_flows}
\alias{eiou_flows}
\title{Energy industry own use (EIOU) flows}
\format{
A string vector with 25 entries.
\describe{
\item{transformation_processes}{The string vector \code{transformation_processes}.}
\item{coal_mines}{The string that identifies Coal mines.}
\item{oil_and_gas_extraction}{The string that identifies Oil and gas extraction.}
}
}
\usage{
eiou_flows
}
\description{
A string vector containing names of \verb{Energy industry own use} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
eiou_flows
}
\keyword{datasets}
