% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_heat_output}
\alias{elec_heat_output}
\title{Electricity and heat output names}
\format{
A character string
\describe{
\item{electricity_output_prefix}{The string prefix for electricity output.}
\item{heat_output_prefix}{The string prefix for heat output.}
\item{output_machine_delimiter}{The string delimiter between output energy flow and the machine name.}
}
}
\usage{
elec_heat_output
}
\description{
A character vector containing the prefixes for
electricity and heat outputs.
}
\examples{
elec_heat_output
}
\keyword{datasets}
