% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity_heat_output.R
\name{electricity_heat_output}
\alias{electricity_heat_output}
\title{Extract a data frame of electricity and heat output information}
\usage{
electricity_heat_output(
  .iea_file = NULL,
  electricity_output_prefix = IEATools::elec_heat_output$electricity_output_prefix,
  heat_output_prefix = IEATools::elec_heat_output$heat_output_prefix,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  input_colname = IEATools::elec_heat_output$input_product,
  output_colname = IEATools::elec_heat_output$output_product,
  machine_colname = IEATools::template_cols$machine,
  output_machine_delimiter = IEATools::elec_heat_output$output_machine_delimiter,
  total = IEATools::memo_aggregation_product_prefixes$total,
  memo = IEATools::memo_aggregation_flow_prefixes$memo
)
}
\arguments{
\item{.iea_file}{The IEA data file to read}

\item{electricity_output_prefix}{The prefix for electricity output rows.
Default is \code{IEATools::elec_heat_output$electricity_output_prefix}.}

\item{heat_output_prefix}{The prefix for heat output rows.
Default is \code{IEATools::elec_heat_output$heat_output_prefix}.}
}
\value{
A data frame of electricity and heat output data.
}
\description{
Extract a data frame of electricity and heat output information
}
