% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electricity_products}
\alias{electricity_products}
\title{Electricity products}
\format{
A string vector with 1 entries.
\describe{
\item{electricity}{The string identifier for Electricity.}
}
}
\usage{
electricity_products
}
\description{
A string vector containing names of products classified by the IEA as electricity products.
}
\examples{
electricity_products
}
\keyword{datasets}
