% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_type_order}
\alias{energy_type_order}
\title{Energy type order}
\format{A string vector}
\usage{
energy_type_order
}
\description{
A string vector containing the order for energy types.
}
\keyword{datasets}
