% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{eta_fu_table_completed}
\alias{eta_fu_table_completed}
\title{Tell whether a final-to-useful efficiency table has been completed}
\usage{
eta_fu_table_completed(
  eta_fu_table,
  fu_allocation_table,
  which_quantity = c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u)
)
}
\arguments{
\item{eta_fu_table}{The final-to-useful efficiency table whose completeness is to be determined.}

\item{fu_allocation_table}{The final-to-useful allocation table whose final-to-useful machines must be assigned efficiencies.}

\item{which_quantity}{A vector of quantities to be completed in the eta_FU table.
Default is \code{c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u)}.}
}
\value{
A boolean telling whether \code{eta_fu_table} is complete.
If \code{FALSE}, a data frame of fu_allocation_table machines that lack efficiencies
is stored in the "machines_without_efficiencies" attribute of the return value.
Retrieve with \code{attr(done, "machines_without_efficiencies")}
if the result of this function is assigned to the variable \code{done}.
}
\description{
A final-to-useful efficiency table is complete iff all of the machines in a final-to-useful allocation table
have been assigned efficiencies in each year for each country.
}
