% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_product_order}
\alias{eu_product_order}
\title{Useful energy carrier order}
\format{A string vector}
\usage{
eu_product_order
}
\description{
A string vector containing entries for ordered useful energy products.
}
\keyword{datasets}
