% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_na.R
\name{explore_na}
\alias{explore_na}
\title{Create multiple plots to explore gaps in the time series}
\usage{
explore_na(
  x,
  time = NULL,
  cex_x = 1,
  cex_y = 1,
  madj_y = 0,
  hadj_y = 1,
  respect = TRUE
)
}
\arguments{
\item{x}{matrix or data frame containing time series of multiple variables.}

\item{time}{vector of time units that will be used for the x ax}

\item{cex_x}{double; the magnification to be used for the x labels relative to the
current setting of cex.}

\item{cex_y}{double; the magnification to be used for the y labels relative to the
current setting of cex.}

\item{madj_y}{double; adjustment of left margin relative to the
current setting for longer or short variable names in the image plot and right barplot.}

\item{hadj_y}{double; horizontal adjustment of the y labels in the image plot.}

\item{respect}{logical; this argument controls whether a unit column-width is the
same physical measurement on the device as a unit row-height (default is TRUE).}
}
\description{
This function creates an image plot where available values for the different variables and years
are indicated in gray and missing values in white. At the right side and bottom, two barplots
are added showing the frequency of available variables per year and the available years per
variable.
}
\examples{
mat <- matrix(1:200, ncol = 10)
mat[sample(1:200, sample(10:80, 1))] <- NA
colnames(mat) <- paste0("variable", 1:10)
explore_na(x = mat, time = 1981:2000)
colnames(mat) <- LETTERS[1:10]
explore_na(x = mat, time = 1981:2000, madj_y = -2.5, hadj_y = 2, respect = FALSE)
}
\author{
Saskia A. Otto
}
