% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{extend_to_useful_helper}
\alias{extend_to_useful_helper}
\title{A helper function for extending to the useful energy/exergy stage}
\usage{
extend_to_useful_helper(
  dest_mat,
  C_mat,
  eta_fu_vec,
  notation,
  product_type,
  industry_type
)
}
\arguments{
\item{dest_mat}{a \code{Y} or \code{U_eiou} matrix (Product x Industry) that is a destination for final energy in an energy conversion chain}

\item{C_mat}{an allocation matrix (either \code{Y} or \code{U_eiou}, both Product x Industry),
indicating the distribution of
final energy carriers (Products) to
final-to-useful energy conversion machines (Industries).
\code{C_mat} should have been created by \code{form_C_mats()}.}

\item{eta_fu_vec}{an efficiency column vector indicating the efficiency (column)
of final-to-useful energy conversion machines (rows).
\code{eta_fu_vec} should have been created by \code{form_eta_fu_phi_u_vecs()}.}

\item{notation}{a row and column name notation vector. See \code{matsbyname::notation_vec()}.}

\item{product_type}{a string identifying product row or column types}

\item{industry_type}{a string identifying industry row or column types}
}
\value{
a named list containing three items:
\code{add_to_U_f} (a matrix to be added to the \code{U_excl_eiou} matrix),
\code{add_to_V_f} (a matrix to be added to the \code{V} matrix), and
\code{repl_dest_mat} (a matrix to replace either \code{Y_f} or \code{U_eiou}).
}
\description{
The helper function is needed, because moving from final to useful energy
occurs for both the final demand matrix (\code{Y}) and the energy industry own use matrix (\code{U_eiou}).
The calculations are identical, so we factor the calculations into this function.
}
