% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{extract_S_units_from_tidy}
\alias{extract_S_units_from_tidy}
\title{Extract a unit summation matrix from a tidy data frame}
\usage{
extract_S_units_from_tidy(
  .tidy_iea_df,
  cols = IEATools::iea_cols,
  ledger_side = cols$ledger_side,
  flow_aggregation_point = cols$flow_aggregation_point,
  flow = cols$flow,
  product = cols$product,
  e_dot = cols$e_dot,
  unit = cols$unit,
  s_units = "S_units",
  val = ".val",
  rowtype = ".rowtype",
  coltype = ".coltype"
)
}
\arguments{
\item{.tidy_iea_df}{the tidy data frame from which a unit summation \code{S_units} matrix is to be formed.}

\item{cols}{a list containing names of columns. Default is \code{iea_cols}.}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is \code{cols$ledger_side}.}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is \code{cols$ledger_side}.}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is \code{cols$flow}.}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is \code{cols$product}.}

\item{e_dot}{the name of the energy flow rate column in \code{.tidy_iea_df}. Default is \code{cols$e_dot}.}

\item{unit}{the name of the unit column in \code{.tidy_iea_df}. Default is \code{cols$unit}.}

\item{s_units}{the name of the unit summation column to be added to \code{.tidy_iea_df}. Default is "S_unit".}

\item{val}{the name of a temporary column to be created in \code{.tidy_iea_df}. Default is ".val".}

\item{rowtype}{the name of a temporary rowtype column created in \code{.tidy_iea_df}. Default is ".rowtype".}

\item{coltype}{the name of a temporary coltype column created in \code{.tidy_iea_df}. Default is ".coltype".}
}
\value{
a data frame containing grouping variables and a new column of unit summation matrices called \code{s_unit}.
}
\description{
Unit summation matrices have products in rows and units in columns, with
\code{1}s where a product is expressed in the unit and \code{0}s otherwise.
}
\details{
\code{.tidy_iea_df} should be grouped as needed, typically on
\code{Country}, \code{Year}, \code{Energy.type}, \code{Last.stage}, etc., but
\emph{not} on \code{Unit}, \code{Flow} or \code{Product}.
\code{.tidy_iea_df} is typically obtained from \code{\link[=tidy_iea_df]{tidy_iea_df()}}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  extract_S_units_from_tidy()
}
