% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fap_flows}
\alias{fap_flows}
\title{IEA order for combined flow aggregation point and flow}
\format{
A string list with 88
}
\usage{
fap_flows
}
\description{
A string vector containing entries for a united flow aggregation point and flow column in IEA order.
}
\examples{
fap_flows
}
\keyword{datasets}
