% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates}
\alias{finaldemand_aggregates}
\title{Final demand aggregate energy from IEA tables}
\usage{
finaldemand_aggregates(
  .ieadata,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  consumption = IEATools::ledger_sides$consumption,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  diff_colname = ".gross_less_net",
  net_aggregate_demand = IEATools::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand = IEATools::aggregate_cols$gross_aggregate_demand
)
}
\arguments{
\item{.ieadata}{A data frame with columns of IEA data.}

\item{country, method, energy_type, last_stage, year, ledger_side, flow_integration_point, flow, e_dot}{See \code{IEATools::iea_cols}.}

\item{consumption}{See \code{IEATools::ledger_sides}.}

\item{eiou}{See \code{IEATools::tfc_compare_flows}.}

\item{net_aggregate_demand, gross_aggregate_demand}{See \code{IEATools::aggregate_cols}.}
}
\value{
A data frame containing grouping columns of \code{.ieadata} and
two additional columns containing net and gross final demand.
}
\description{
Calculates total aggregate final demand energy from a data frame of IEA data
on both net and gross bases.
}
\details{
This function works similar to \code{dplyr::summarise()}:
it distills \code{.ieadata} to many fewer rows
according to the grouping variables
\code{country}, \code{method}, \code{energy_type}, \code{last_stage}, and \code{year},
and, possibly, additional grouping variables in the input (\code{.ieadata})..
\code{.ieadata} is grouped by those variables internally.
Any grouping variables present in \code{.ieadata} are retained
for calculating primary aggregate energy.
Grouping is removed before output.
}
\examples{
load_tidy_iea_df() \%>\% 
  finaldemand_aggregates()
}
