% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates_IEA}
\alias{finaldemand_aggregates_IEA}
\title{Final demand aggregate energy from IEA tables}
\usage{
finaldemand_aggregates_IEA(
  .ieadata,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  consumption = IEATools::ledger_sides$consumption,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  diff_colname = ".gross_less_net",
  aggregate_net_finaldemand = Recca::aggregate_cols$aggregate_net_finaldemand_iea,
  aggregate_gross_finaldemand = Recca::aggregate_cols$aggregate_gross_finaldemand_iea
)
}
\arguments{
\item{.ieadata}{a data frame with columns of IEA data.}

\item{ledger_side}{the name of the ledger side column in \code{.ieadata}.
Default is "\code{Ledger.side}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.ieadata}.
Default is "\code{Flow.aggregation.point}".}

\item{flow}{the name of the column that contains flow information.
Default is "\code{Flow}".}

\item{consumption}{the identifier for consumption in the \code{flow_aggregation_point} column.
Default is "\code{Consumption}".}

\item{eiou}{the identifier for energy industry own use in the \code{flow_aggregation_point} column.
Default is "\code{Energy industry own use}".}

\item{aggregate_net_finaldemand}{the name of the output column containing aggregates of net final demand.
Default is "\code{EX_fd_net_IEA.ktoe}".}

\item{aggregate_gross_finaldemand}{the name of the output column containing aggregates of gross final demand.
Default is "\code{EX_fd_gross_IEA.ktoe}".}

\item{energy}{the name of the column that contains energy information.
Default is "\code{E.dot}".}
}
\value{
a data frame containing grouping columns of \code{.ieadata} and
two additional columns containing net and gross final demand.
}
\description{
Calculates total aggregate final demand energy from a data frame of IEA data
on both net and gross bases.
}
\details{
This function is named with "_IEA", because it is meant to operate on
tidy, IEA-style data frames.
The function \link{finaldemand_aggregates} does the same thing,
but it is meant to operate on SUT-style data frames.

Note that all items in \code{.ieadata} must be in same units.

This function works similar to \code{\link[dplyr]{summarise}}:
it distills \code{.ieadata} to many fewer rows
according to the grouping variables.
Thus, \code{.ieadata} should be grouped prior to passing into this function.
Grouping columns are preserved on output.
}
\examples{
library(dplyr)
# Works only when all entries are in same units.
# When Last.stage is services, different units are involved.
# Thus, filter to rows in UKEnergy2000tidy where Last.stage is final or useful.
print(names(UKEnergy2000tidy))
UKEnergy2000tidy \%>\%
  group_by(Country, Year, Energy.type, Last.stage) \%>\%
  filter(Last.stage \%in\% c("Final", "Useful")) \%>\%
  finaldemand_aggregates_IEA()
}
