% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{find_allocated_rows}
\alias{find_allocated_rows}
\title{Find rows of final energy consumption already allocated to final-to-useful machines.}
\usage{
find_allocated_rows(
  fu_allocation_table,
  quantity,
  machine,
  eu_product,
  .values
)
}
\arguments{
\item{fu_allocation_table}{A tidy final-to-useful allocation table.}

\item{quantity}{The name of the Quantity column.}

\item{machine}{The name of the Machine column.}

\item{eu_product}{The name of the Eu.product column.}

\item{.values}{The name of the .valuse column.}
}
\value{
A data frame containing unique rows of final energy that are allocated in \code{fu_allocation_table}.
}
\description{
This is convenience function that reduces duplicated code in \code{complete_fu_allocation_table()}.
As such, it is not exported.
}
