% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{find_available_eta_fu_phi_u_info}
\alias{find_available_eta_fu_phi_u_info}
\title{Find available information from an eta_fu_table}
\usage{
find_available_eta_fu_phi_u_info(
  eta_fu_table,
  which_quantity,
  quantity,
  e_dot_machine,
  e_dot_machine_perc,
  maximum_values,
  .values,
  year,
  eta_fu_phi_u_source,
  country
)
}
\arguments{
\item{eta_fu_table}{An efficiency table from which available data are to be obtained.}

\item{which_quantity}{The quantity you want to look for, usually "eta.fu" (for final-to-useful efficiency) or "phi.u" (for the useful exergy-to-energy ratio).}

\item{quantity}{The quantity column in eta_fu_table.}

\item{e_dot_machine}{The row name for energy consumption of the machine.}

\item{e_dot_machine_perc}{The row name for percentage of total energy consumptino by the machine.}

\item{maximum_values}{The name for the Maximum.values column in eta_fu_table.}

\item{.values}{The name for an internally-generated column.}

\item{year}{The name of the year column.}

\item{eta_fu_phi_u_source}{The name of the source column.}

\item{country}{The name of the country column.}
}
\value{
A data frame that containing available data in the eta_fu_table.
}
\description{
This is convenience function that minimizes repeated code.
Thus, it is not exported.
}
