% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_splits.R
\name{find_eiou_end}
\alias{find_eiou_end}
\title{Find the point where Energy industry own use ends}
\usage{
find_eiou_end(.ctry_tbl, flow, non_specified, losses)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{the name of the flow column}

\item{non_specified}{the name of non-specified in the flow column}

\item{losses}{the name for losses in the flow column}
}
\value{
a pair of integers representing the rows that straddle
the split between Energy industry own use and Losses.
The first integer is the last row of Energy industry own use.
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the Energy industry own use and Losses (a TFC compare flow).
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between
Transformation processes and Energy industry own use.
}
