% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_splits.R
\name{find_eiou_start}
\alias{find_eiou_start}
\title{Find the point where Energy industry own use starts}
\usage{
find_eiou_start(
  .ctry_tbl,
  flow = IEATools::iea_cols$flow,
  non_specified = IEATools::non_specified_flows$non_specified,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  coal_mines = "Coal mines"
)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{See \code{IEATools::iea_cols$flow}.}

\item{non_specified}{See \code{IEATools::non_specified_flows}.}

\item{eiou}{See \code{IEATools::tfc_compare_flows}.}

\item{coal_mines}{the name for coal mines in the flow column. Default is "Coal mines".}
}
\value{
a pair of integers representing the rows that straddle
the split between Transformation processes and Energy industry own use
The second integer is the first row of Energy industry own use.
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the Transformation processes and Energy industry own use.
This function simply calls \code{find_transformation_end()}, because
the dividing line between Transformation processes and Energy industry own use
is also the point where Energy industry own use starts.
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between
Transformation processes and Energy industry own use.
}
