% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_splits.R
\name{find_supply_consumption_split}
\alias{find_supply_consumption_split}
\title{Find the split point between Supply and Consumption}
\usage{
find_supply_consumption_split(
  .ctry_tbl,
  flow = IEATools::iea_cols$flow,
  losses = IEATools::tfc_compare_flows$losses,
  iron_and_steel = IEATools::manufacturing_flows$iron_and_steel,
  mining_and_quarrying = IEATools::industry_flows$mining_and_quarrying,
  tfc = IEATools::aggregation_flows$total_final_consumption,
  industry = IEATools::tfc_flows$industry
)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{See \code{IEATools::iea_cols}.}

\item{losses}{See \code{IEATools::tfc_compare_flows}.}

\item{iron_and_steel}{See \code{IEATools::manufactuiring_flows}.}

\item{mining_and_quarrying}{See \code{IEATools::industry_flows}.}

\item{tfc}{See \code{IEATools::aggregation_flows}.}

\item{industry}{See \code{IEATools::tfc_flows}.}
}
\value{
a pair of integers representing the rows that straddle the split between
the Supply and Consumption sides of the ledger
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the Supply and Consumption sides of the ledger.
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between Supply and Consumption.
}
