% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_splits.R
\name{find_transformation_start}
\alias{find_transformation_start}
\title{Find the point where Transformation processes begin}
\usage{
find_transformation_start(
  .ctry_tbl,
  flow = IEATools::iea_cols$flow,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  statistical_differences = IEATools::tfc_compare_flows$statistical_differences,
  mapep = IEATools::transformation_processes$main_activity_producer_electricity_plants
)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{See \code{IEATools::iea_cols}.}

\item{transformation_processes, statistical_differences}{See \code{IEATools::tfc_compare_flows}.}

\item{mapep}{See \code{IEATools::transformation_processes}.}
}
\value{
a pair of integers representing the rows that straddle the split between
the TFC compare and Transformation processes flows.
The second integer is the first row of Transformation processes.
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the first TFC Compare section and Transformation processes.
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between
TFC compare and Transformation processes.
}
