% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_AUS_bfg}
\alias{fix_AUS_bfg}
\title{Fix Australia Blast furnace gas for 2010--2020}
\usage{
fix_AUS_bfg(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  e_dot = IEATools::iea_cols$e_dot
)
}
\arguments{
\item{.tidy_iea_df}{IEA data frame produced by \code{load_tidy_iea_df()}.}

\item{country, year, e_dot}{See \code{IEATools::iea_cols}.}
}
\value{
\code{.tidy_iea_df} with improved Australia Blast furnace gas.
}
\description{
Australia's Blast furnaces have an undesirable characteristic
that leads to singular matrices:
From 2013 onward, the production of Blast furnace gas by Blast furnaces
is consumed only by Blast furnaces.
No other industry or energy production machine consumes
Blast furnace gas.
In fact, the problem is deeper,
starting in 2010, the Iron and steel industry consumes no Blast furnace gas,
in apparent contradiction to the IEA's own policies for reporting
Blast furnace gas consumption.
This function fixes the Blast furnace gas data for Australia
for the 2010--2020 timeframe.
}
\details{
The fix involves ensuring that the Iron and steel industry always
consumes Blast furnace gas, according to the IEA's assumed efficiency
of 40\%.
This function makes use of the \code{Fixed_AUS_bfg} data frame.
}
\examples{
load_tidy_iea_df() |>
  dplyr::filter(Country == "ZAF", Year == 1971) |> 
  dplyr::mutate(
    # Pretend that ZAF is Australia.
    Country = "AUS", 
    # And that 1971 is 2013
    Year = 2013
  ) |> 
  # This will change Blast furnace gas.
  fix_AUS_bfg()
}
