% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_COL_WRLD_electricity}
\alias{fix_COL_WRLD_electricity}
\title{Improve COL electricity generation in the IEA's WEEB.}
\usage{
fix_COL_WRLD_electricity(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  e_dot = IEATools::iea_cols$e_dot
)
}
\arguments{
\item{.tidy_iea_df}{a tidy IEA data frame produced by \code{\link[=load_tidy_iea_df]{load_tidy_iea_df()}}}

\item{country, year, e_dot}{See \code{IEATools::iea_cols}.}
}
\value{
\code{.tidy_iea_df} with improved Ghana Industry Electricity, if warranted.
}
\description{
The 2022 release of the IEA's WEEB data
is different from the 2021 release of the IEA's WEEB data
in terms of Colombia's Electricity generation.
For example:
\itemize{
\item 2021 release:
\itemize{
\item Main activity producer electricity plants 34196.4008 TJ
\item Autoproducer electricity plants: 2671.1993 TJ
}
\item 2022 release:
\itemize{
\item Main activity producer electricity plants 31467.5994 TJ
\item Autoproducer electricity plants: 899.9987 TJ
Similar differences appear in all years 1971 - 1977.
From 1978 onward, the 2021 and 2022 releases agree.
Note that this change leads to overall energy imbalance
for Colombia 1971-1977 and World 1971-1977.
This function reverts to the values from the 2021 release of the IEA WEEB.
}
}
}
\details{
Similarly, World Electricity is different 2021 release to 2022 release.
The 2022 data are unbalanced for 1971 --> 1977.
This function reverts to the value from the 2021 release of the IEA WEEB
for World Electricity in 1971--1977.
}
\examples{
library(dplyr)
example_tidy_iea_df <- load_tidy_iea_df() \%>\% 
  dplyr::filter(Country == "GHA") |> 
  dplyr::mutate(
    # Pretend the GHA is COL.
    Country = "COL"
  )
example_tidy_iea_df
fixed <- example_tidy_iea_df \%>\% 
  fix_COL_WRLD_electricity()
# Compare changed values
example_tidy_iea_df \%>\% 
  dplyr::filter(Flow \%in\% c("Main activity producer electricity plants",
                            "Autoproducer electricity plants"), 
         Product == "Electricity") \%>\% 
  dplyr::select("Year", "Flow", "E.dot", "Unit")
fixed \%>\% 
  dplyr::filter(Flow \%in\% c("Main activity producer electricity plants",
                      "Autoproducer electricity plants"), 
                Product == "Electricity") \%>\% 
  dplyr::select("Year", "Flow", "E.dot", "Unit")
}
