% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_GHA_psb}
\alias{fix_GHA_psb}
\title{Fix Ghana Primary solid biofuels data from 1991 through 1999}
\usage{
fix_GHA_psb(.tidy_iea_df)
}
\arguments{
\item{.tidy_iea_df}{a tidy IEA data frame produced by \code{\link[=load_tidy_iea_df]{load_tidy_iea_df()}}}
}
\value{
\code{.tidy_iea_df} with smoothed Ghana Primary solid biofuels data
}
\description{
Ghana's Primary solid biofuels data show a very large and dramatic decline from 1999 to 2000.
This decline is due to new survey data being used for the 2000 data.
When we look at the PSB data on a per-capita basis, it is clear that
a near-constant PSB/capita value was used to extrapolate per-capita usage in the late 1990s.
When new survey data became available for the 2000 reporting year,
the per-capita consumption of PSB obviously changed.
Our approach to this problem is to smooth out the really big peak in PSB consumption
by reducing the per-capita consumption of PSB, starting in 1991.
}
\details{
See the Supplemental information to
M. K. Heun and P. E. Brockway.
Meeting 2030 primary energy and economic growth goals: Mission impossible?
Applied Energy, 251(112697):1–24, May 2019
for additional details.
}
\examples{

}
