% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{fix_header}
\alias{fix_header}
\title{Fix IEA data header}
\usage{
fix_header(unfixed_header, newlinechar = "\\n",
  incorrect_first_line_lead = ",,TIME",
  correct_first_line_lead = "COUNTRY,FLOW,PRODUCT")
}
\arguments{
\item{unfixed_header}{the header (first two lines) of an IEA data file as a string.}
}
\value{
a fixed version of the header as a string.
}
\description{
The IEA extended energy balance files headers are not conducive to creating data frames.
They contain two-line headers with blanks in the wrong places.
This function fixes the header to ensure that it works properly.
}
\examples{
fix_header(",,TIME,1960,1961\\nCOUNTRY,FLOW,PRODUCT")
}
