% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{fix_iea_df_balance}
\alias{fix_iea_df_balance}
\title{Fix IEA energy balances}
\usage{
fix_iea_df_balance(.tidy_iea_df, flow = "Flow",
  statistical_differences = "Statistical differences", e_dot = "E.dot",
  units = "Units", source = ".source", .iea = "IEA")
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA data}
}
\value{
\code{.tidy_iea_df} with adjusted \code{Statistical differences} flows such that
the data are in perfect energy balance.
}
\description{
IEA data are not quite balanced.
In fact, production and consumption are often wrong by many ktoe
for any given Product in a Country in a Year.
This function ensures that the balance is perfect
by adjusting the \code{Statistical differences} flow
on a per-product basis.
}
\details{
This function assumes that \code{.tidy_iea_df} is grouped appropriately.
Typical grouping variables would include
\code{Country}, \code{Year}, \code{Energy.type}, \code{Last.stage}, \code{Product}, etc.
}
