% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_RUSESTHeat}
\alias{fix_RUSESTHeat}
\title{Fix Russia and Estonia Heat for 1990--1993}
\usage{
fix_RUSESTHeat(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  e_dot = IEATools::iea_cols$e_dot
)
}
\arguments{
\item{.tidy_iea_df}{IEA data frame produced by \code{load_tidy_iea_df()}.}

\item{country, year, e_dot}{See \code{IEATools::iea_cols}.}
}
\value{
\code{.tidy_iea_df} with improved RUS and EST Heat in the early 1990s.
}
\description{
The breakup of the Soviet Union (SUN) caused many irregularities
in the IEA's energy accounting
for SUN, Russia (RUS), and other former-Soviet states.
In particular, the Flow of Heat is assigned to
"Final consumption not elsewhere specified" for
1990--1992 (RUS) and 1990--1993 (EST).
However, for following years, Heat is assigned to specific sectors.
Other FoSUN countries do not exhibit the same problem.
This Heat accounting irregularity
is one of a series of problems that causes a jump
in final-to-useful efficiencies for Europe and World in the CL-PFU database
in the time period 1989--1990.
This function performs a fix wherein
Final consumption not elsewhere specified Heat
is re-assigning
to specific sectors by the proportion found in 1993 (RUS) and 1994 (EST)
for 19990--1992 (RUS) and 1990--1993 (EST).
}
