% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{fix_tidy_iea_df_balances}
\alias{fix_tidy_iea_df_balances}
\title{Fix IEA energy balances}
\usage{
fix_tidy_iea_df_balances(.tidy_iea_df, ledger_side = "Ledger.side",
  supply = "Supply", consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point",
  tfc_compare = "TFC compare", flow = "Flow",
  statistical_differences = "Statistical differences",
  product = "Product", e_dot = "E.dot", err = ".err",
  remove_zeroes = TRUE, grouping_vars = c("Method", "Last.stage",
  "Country", "Year", "Energy.type", "Unit", "Product"))
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA data, typically the output}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is "\code{Ledger.side}".}

\item{supply}{a string indicating the supply side of the ledger in the \code{ledger_side} column. Default is "\code{Supply}".}

\item{consumption}{a string indicating the consumption side of the ledger in the \code{ledger_side} column. Default is "\code{Consumption}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{tfc_compare}{a string indicating the Total final consumption comparison flow aggregation point. Default is "\code{TFC compare}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{statistical_differences}{a string indicating statistical differences in the \code{flow} column. Default is "\code{Statistical differences}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{e_dot}{the name of the energy flow column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}

\item{err}{the name of a temporary error column added to \code{.tidy_iea_df}. Default is "\code{.err}".}

\item{remove_zeroes}{a logical telling whether to remove \code{0}s after balancing.}

\item{grouping_vars}{a string vector of names of columns by which energy balance calculations should be performed.
Default is \code{c("Method", "Last.stage", "Country", "Year", "Energy.type", "Unit", "Product")}.}
}
\value{
\code{.tidy_iea_df} with adjusted \code{Statistical differences} flows such that
the data for each product are in perfect energy balance.
}
\description{
IEA extended energy balance data are sometimes not quite balanced.
In fact, supply and consumption are often wrong by a few ktoe
for any given Product in a Country in a Year.
This function ensures that the balance is perfect
by adjusting the \code{Statistical differences} flow
on a per-product basis.
}
\details{
This function assumes that \code{.tidy_iea_df} is grouped appropriately.
So be sure to group \code{.tidy_iea_df} by appropriate variables
before calling this function.
Grouping should \emph{definitely} be done on the \code{Product} column.
Typically, grouping is also done on
\code{Country}, \code{Year}, \code{Energy.type}, \code{Last.stage}, etc. columns.
Grouping should \emph{not} be done on the \code{flow} column or the \code{ledger_side} column.

Internally, this function calls \code{\link[=calc_tidy_iea_df_balances]{calc_tidy_iea_df_balances()}}
and adjusts \code{Statistical differences} by any imbalances that are present.
}
\examples{
library(dplyr)
unbalanced <- load_tidy_iea_df()
# This will not be balanced, because the IEA data are not in perfect balance
unbalanced \%>\% 
  # group_by(Country, Year, Energy.type, Unit, Product) \%>\% 
  calc_tidy_iea_df_balances() \%>\% 
  tidy_iea_df_balanced()
# This will be balanced, becasue we fix the imbalances.
unbalanced \%>\% 
 #  group_by(Country, Year, Energy.type, Unit, Product) \%>\% 
  fix_tidy_iea_df_balances() \%>\% 
  calc_tidy_iea_df_balances() \%>\% 
  tidy_iea_df_balanced()
}
