% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_AUS_bfg}
\alias{Fixed_AUS_bfg}
\title{Fixed Australia Blast furnace gas for 2010--2020}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_AUS_bfg
}
\description{
Australia's Blast furnaces have an undesirable characteristic
that leads to singular matrices:
From 2013 onward, the production of Blast furnace gas by Blast furnaces
is consumed only by Blast furnaces.
No other industry or energy production machine consumes
Blast furnace gas.
In fact, the problem is deeper,
starting in 2010, the Iron and steel industry consumes no Blast furnace gas,
in apparent contradiction to the IEA's own policies for reporting
Blast furnace gas consumption.
This data frame provides the "fixed" Blast furnace gas data for Australia
for the 2010--2020 timeframe.
}
\details{
The fix involves ensuring that the Iron and steel industry always
consumes Blast furnace gas, according to the IEA's assumed efficiency
of 40\%.
The function \code{fix_AUS_bfg()} makes use of these data.
}
\examples{
Fixed_AUS_bfg
}
\keyword{datasets}
