% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_COL_Electricity_Generation}
\alias{Fixed_COL_Electricity_Generation}
\title{Fixed Colombia Electricty generation 1971--1977}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_COL_Electricity_Generation
}
\description{
Colombia's electricity production changed in the 2022 release of the IEA data.
In the 2022 release, COL is out of balance for 1971--1977 as a result.
Further, the imbalance extends to the World as a whole.
This object contains the (presumably) correct data (obtained from the 2021 release).
The function \code{\link[=fix_COL_electricity_generation]{fix_COL_electricity_generation()}} makes use of these data.
}
\examples{
Fixed_COL_Electricity_Generation
}
\keyword{datasets}
