% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_GHA_Industry_Electricity}
\alias{Fixed_GHA_Industry_Electricity}
\title{Fixed Ghana Industry Electricity data}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_GHA_Industry_Electricity
}
\description{
Ghana's Industry Electricity data have specifics for
Mining and quarrying,
Non-ferrous metals, and
Textile and leather
for the years 1971--1973 only.
However, data to bring more specificity to Industry Electricity consumption
are available from the Ghana Grid Corporation (GridCo) and the Volta River Authority (VRA).
These data have been compiled into this object.
The function fix_GHA_industry_electricity() makes use of these data.
}
\examples{
Fixed_GHA_Industry_Electricity
}
\keyword{datasets}
