% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_GHA_PSB}
\alias{Fixed_GHA_PSB}
\title{Fixed Ghana Primary solid biofuels data}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_GHA_PSB
}
\description{
Ghana's Primary solid biofuels data show a very large and dramatic decline from 1999 to 2000
in IEA WEEB data.
This decline is due to new survey data being used for the 2000 data.
When we look at the PSB data on a per-capita basis, it is clear that
a near-constant PSB/capita value was used to extrapolate per-capita usage in the late 1990s.
When new survey data became available for the 2000 reporting year,
the per-capita consumption of PSB obviously changed.
Our approach to this problem is to smooth out the really big peak in PSB consumption
by reducing the per-capita consumption of PSB, starting in 1991.
This data frame contains the "fixed" data.
The function \code{\link[=fix_GHA_psb]{fix_GHA_psb()}} makes use of these data.
}
\examples{
Fixed_GHA_PSB
}
\keyword{datasets}
