% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_OAMR_Cpp}
\alias{Fixed_OAMR_Cpp}
\title{Fixed Other non-OECD Americas Charcoal production 1971--2010}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_OAMR_Cpp
}
\description{
Other Non-OECD Americas has several years where Charcoal is produced
but no Primary solid biofuels are consumed to
create the Charcoal.
This object contains (presumably) correct data.
In particular, Charcoal production plants
now consume Primary solid biofuels, and
Primary solid biofuels production is boosted accordingly.
The function \code{fix_OAMR_Cpp()} makes use of these data.
}
\examples{
Fixed_OAMR_Cpp
}
\keyword{datasets}
