% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_RUSEST_heat}
\alias{Fixed_RUSEST_heat}
\title{Fixed Russia and Estonia Heat 1990--1993}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_RUSEST_heat
}
\description{
Fixed Russia and Estonia Heat 1990--1993
}
\examples{
Fixed_RUSEST_heat
}
\keyword{datasets}
