% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{form_C_mats}
\alias{form_C_mats}
\title{Create allocation matrices (\code{C}) from an allocation table}
\usage{
form_C_mats(
  .fu_allocations,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  e_dot = IEATools::iea_cols$e_dot,
  year = IEATools::iea_cols$year,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  quantity = IEATools::template_cols$quantity,
  machine = IEATools::template_cols$machine,
  ef_product = IEATools::template_cols$ef_product,
  eu_product = IEATools::template_cols$eu_product,
  destination = IEATools::template_cols$destination,
  e_dot_perc = IEATools::template_cols$e_dot_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  product = IEATools::row_col_types$product,
  industry = IEATools::row_col_types$industry,
  sep = " -> ",
  C_eiou = IEATools::template_cols$C_eiou,
  C_Y = IEATools::template_cols$C_Y,
  .should_be_1_vector = ".should_be_1_vector",
  .is_1_vector = ".is_1_vector"
)
}
\arguments{
\item{.fu_allocations}{a final-to-useful allocation table read by \code{load_fu_allocation_data()}.}

\item{country, method, energy_type, last_stage, ledger_side, flow_aggregation_point, e_dot, year}{See \code{IEATools::iea_cols}.}

\item{supply, consumption}{See \code{IEATools::ledger_sides}.}

\item{quantity, machine, ef_product, eu_product, destination, e_dot_perc, maximum_values, C_eiou, C_Y}{See \code{IEATools::template_cols}.}

\item{matnames, matvals, rownames, colnames, rowtypes, coltypes}{See \code{IEATools::mat_meta_cols}.}

\item{product, industry}{See \code{IEATools::row_col_types}.}

\item{sep}{The string separator between prefix and suffix of compound row and column names. Default is " -> ".
The default value matches the default value for the \code{sep} argument of \code{matsbyname::vectorize_byname()}, because
\code{matsbyname::vectorize_byname()} will be used for further manipulations.}

\item{.should_be_1_vector}{a temporary column created internally (and not returned) for error checking.
This column should contain 1 vectors (i.e., vectors filled with 1's).}

\item{.is_1_vector}{a temporary column created internally (and not returned)
that contains \code{TRUE} or \code{FALSE} depending on whether a 1 vector was created by row sums.
This column is tested internally to ensure all values are \code{TRUE} before returning.}
}
\value{
a tidy data frame with metadata columns (and year) along with \code{matnames} and \code{matvals} columns
indicating and containing \code{C_eiou} and \code{C_Y} matrices, respectively.
}
\description{
This function uses information in a filled allocation template (created by \code{write_fu_allocation_template()})
to create allocation matrices (\code{C}).
rownames of the \code{C} matricds are taken from the \code{Ef.product} and \code{Destination} columns, and
colnames are taken from the \code{Machine} and \code{Eu.product} columns.
\code{C} matrices are created for both energy industry own use
and final demand (\code{C_eiou} and \code{C_Y}, respectively).
}
\details{
Rows of the output \code{C} matrices should sum to 1.
An error is thrown if this is not so.
Such errors probably indicate the template was not filled correctly.
}
\examples{
load_fu_allocation_data() \%>\% 
  form_C_mats()
}
