% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{form_eta_fu_phi_u_vecs}
\alias{form_eta_fu_phi_u_vecs}
\title{Create final-to-useful efficiency vectors (\code{eta_fu}) and exergy-to-energy ratio vectors (\code{phi_u}) from a final-to-useful efficiency table}
\usage{
form_eta_fu_phi_u_vecs(
  .eta_fu_table,
  unit = IEATools::iea_cols$unit,
  year = IEATools::iea_cols$year,
  quantity = IEATools::template_cols$quantity,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  eta_fu = IEATools::template_cols$eta_fu,
  phi_u = IEATools::template_cols$phi_u,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  sep = " -> "
)
}
\arguments{
\item{.eta_fu_table}{a final-to-useful efficiency table read by \code{load_eta_fu_allocation_data()}.
A template for this table should have been created by \code{eta_fu_table()} and
\code{write_eta_fu_table()}.}

\item{unit, year}{See \code{IEATools::iea_cols}.}

\item{quantity, machine, eu_product, e_dot_machine, e_dot_machine_perc, maximum_values, eta_fu, phi_u}{See \code{IEATools::template_cols}.}

\item{matnames, matvals, rownames, colnames, rowtypes, coltypes}{See \code{IEATools::mat_meta_cols}.}

\item{sep}{The string separator between prefix and suffix of compound row and column names. Default is " -> ".
The default value matches the default value for the \code{sep} argument of \code{matsbyname::vectorize_byname()}, because
\code{matsbyname::vectorize_byname()} will be used for further manipulations.}
}
\value{
a tidy data frame with metadata columns (and year) along with \code{matnames} and \code{matvals} columns
indicating and containing \code{eta_fu} and \code{phi_u} vectors, respectively.
}
\description{
Create final-to-useful efficiency vectors (\code{eta_fu}) and exergy-to-energy ratio vectors (\code{phi_u}) from a final-to-useful efficiency table
}
\examples{
load_eta_fu_data() \%>\% 
  form_eta_fu_phi_u_vecs()
}
