% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{form_eta_fu_phi_vecs}
\alias{form_eta_fu_phi_vecs}
\title{Title}
\usage{
form_eta_fu_phi_vecs(
  .eta_fu_table,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  e_dot = IEATools::iea_cols$e_dot,
  year = IEATools::iea_cols$year,
  quantity = IEATools::template_cols$quantity,
  machine = IEATools::template_cols$machine,
  ef_product = IEATools::template_cols$ef_product,
  eu_product = IEATools::template_cols$eu_product,
  destination = IEATools::template_cols$destination,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_perc = IEATools::template_cols$e_dot_perc,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes
)
}
\arguments{
\item{.eta_fu_table}{}
}
\value{

}
\description{
Title
}
\examples{
load_eta_fu_data() \%>\% 
  form_eta_fu_vecs()
}
