% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{form_eta_fu_vecs}
\alias{form_eta_fu_vecs}
\title{Title}
\usage{
form_eta_fu_vecs(.eta_fu_table)
}
\arguments{
\item{.eta_fu_table}{}
}
\value{

}
\description{
Title
}
\examples{
load_eta_fu_data() \%>\% 
  form_eta_fu_vecs()
}
