% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{from_notation}
\alias{from_notation}
\title{From notation}
\format{
A character vector with 4 entries.
}
\usage{
from_notation
}
\description{
A character vector containing the "from" notation used in the IEATools package.
}
\examples{
from_notation
}
\keyword{datasets}
