% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{fu_allocation_table_completed}
\alias{fu_allocation_table_completed}
\title{Tell whether a final-to-useful allocation table has been completed}
\usage{
fu_allocation_table_completed(
  fu_allocation_table,
  specified_iea_data,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  consumption = IEATools::ledger_sides$consumption,
  e_dot = IEATools::iea_cols$e_dot,
  maximum_values = IEATools::template_cols$maximum_values,
  quantity = IEATools::template_cols$quantity,
  C_perc = IEATools::template_cols$C_perc,
  machine = IEATools::template_cols$machine,
  e_u_product = IEATools::template_cols$eu_product,
  e_dot_perc = IEATools::template_cols$e_dot_perc
)
}
\arguments{
\item{fu_allocation_table}{The final-to-useful allocation table whose completeness is to be determined.}

\item{specified_iea_data}{An IEA data frame from which final energy flows are gleaned.}
}
\value{
A boolean telling whether \code{fu_allocation_table} is complete.
}
\description{
A final-to-useful allocation table is complete iff all of the final energy flows for a country
are routed to a final-to-useful machine for each year in which those final energy flows exist.
Also, all routes need to add to 100\%.
}
