% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{fu_allocations_available}
\alias{fu_allocations_available}
\title{Determine which final energy uses are covered in an FU Allocation table}
\usage{
fu_allocations_available(
  .fu_allocation_table,
  max_vals,
  year,
  .values,
  quantity,
  e_dot,
  e_dot_perc,
  country,
  source
)
}
\arguments{
\item{.fu_allocation_table}{The FU Allocation table from which allocations are to be determined}

\item{max_vals}{The name of maximum value rows in the \code{fu_allocation_table}.}

\item{year}{The Year column in a tidy version of the \code{fu_allocation_table}.}

\item{.values}{The name of the values column in a tidy versino of the \code{fu_allocation_table}.}

\item{quantity}{The name of the quantity column in the \code{fu_allocation_table}.}

\item{e_dot}{The name of the E.dot rows in the \code{fu_allocation_table}.}

\item{e_dot_perc}{The name of the E.dot percentage rows in the \code{fu_allocation_table}.}

\item{country}{The name of the Country column in the \code{fu_allocation_table}.}

\item{source}{The name of the C_source column in the tidy \code{fu_allocation_table}.}
}
\value{
A data frame containing final energy consummption that has been allocated in the \code{fu_allocation_table}
}
\description{
This is convenience function that reduces duplicated code in \code{complete_fu_allocation_table()}.
As such, it is not exported.
}
