% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fu_analysis_file_info}
\alias{fu_analysis_file_info}
\title{Final-to-useful analysis file information}
\format{
A string list with 3 entries
\describe{
\item{fu_analysis_file_suffix}{The suffix for final-to-useful analysis file names.}
\item{fu_allocation_tab_name}{The string name of final-to-useful allocation tabs.}
\item{eta_fu_tab_name}{The string name of final-to-useful efficiency tabs.}
}
}
\usage{
fu_analysis_file_info
}
\description{
A string list containing named expected filename suffixes and tab names in Excel files
associated with final-to-useful analyses.
Items in the list provide default values for final-to-useful analysis files
throughout the \code{IEATools} package.
}
\examples{
fu_analysis_file_info
}
\keyword{datasets}
