% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{gather_producer_autoproducer}
\alias{gather_producer_autoproducer}
\title{Gather main activity producer and autoproducer industries}
\usage{
gather_producer_autoproducer(
  .tidy_iea_df,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  negzeropos = ".negzeropos",
  autoproducer_elect = IEATools::main_act_plants$autoprod_elect_plants,
  autoproducer_chp = IEATools::transformation_processes$autoproducer_CHP_plants,
  autoproducer_heat = IEATools::transformation_processes$autoproducer_heat_plants,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow_aggregation_point}.}

\item{flow}{The name of the flow column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{e_dot}{The name of the energy column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{transformation_processes}{A string identifying transformation processes in the \code{flow_aggregation_point} column of the \code{.tidy_iea_df}
Default is \code{IEATools::aggregation_flows$flow_aggregation_point}.}

\item{negzeropos}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".negzeropos".}

\item{autoproducer_elect}{A string identifying "Autoproducer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$autoprod_elect_plants}.}

\item{autoproducer_chp}{A string identifying "Autoproducer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::transformation_processes$autoproducer_CHP_plants}.}

\item{autoproducer_heat}{A string identifying "Autoproducer heat plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::transformation_processes$autoproducer_heat_plants}.}

\item{main_act_producer_elect}{A string identifying "Main activity producer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_elect_plants}.}

\item{main_act_producer_heat}{A string identifying "Main activity producer heat plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_heat_plants}.}

\item{main_act_producer_chp}{A string identifying "Main activity producer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_chp_plants}.}
}
\value{
The \code{tidy_iea_df} with autoproducer plants merged with main activity producer plants.
}
\description{
The IEA extended energy balances include both main activity producer
and autoproducer industries for electricity, heat, and CHP plants.
See details for an explication of each.
This function gathers main activity producer and autoproducer, for each
of the three types of plants: electricity, heat, and CHP plants.
This function is called within the \code{specify_all()} function.
}
\details{
Autoproducer plants are those that consume in-situ the energy they produce.
For instance, an iron and steel plant that produces electricity
and directly consumes it would be classified as an autoproducer electricity plant.
Conversely, main activity producer plants are those that produce
a product, be it electricity, heat, or both (CHP plants) and sell it
to the market.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  gather_producer_autoproducer()
}
