% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_industries}
\alias{grid_industries}
\title{Grid industries names}
\format{
A string list with 1 entries.
\describe{
\item{electricity_grid}{The string name of the electricity grid industry.}
}
}
\usage{
grid_industries
}
\description{
A string list containing the names of the grid industries that can be added.
}
\examples{
grid_industries
}
\keyword{datasets}
