% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heat_products}
\alias{heat_products}
\title{Heat products}
\format{
A string vector with 1 entries.
\describe{
\item{heat}{The string identifier for Heat.}
}
}
\usage{
heat_products
}
\description{
A string vector containing names of products classified by the IEA as heat products.
}
\examples{
heat_products
}
\keyword{datasets}
