% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{iea_df}
\alias{iea_df}
\title{Load IEA data from csv file}
\usage{
iea_df(.iea_file = NULL, text = NULL,
  expected_1st_line_start = ",,TIME",
  expected_2nd_line_start = "COUNTRY,FLOW,PRODUCT",
  year_colname_pattern = "^\\\\d*$", missing_data = c("..", "x"))
}
\arguments{
\item{.iea_file}{a string containing the path to a .csv file of extended energy balances from the IEA}

\item{text}{a character string that can be parsed as IEA extended energy balances. 
(This argument is useful for testing.)}

\item{expected_1st_line_start}{the expected start of the first line of \code{iea_file}. Default is "\code{,,TIME}".}

\item{expected_2nd_line_start}{the expected start of the second line of \code{iea_file}. Default is "\code{COUNTRY,FLOW,PRODUCT}".}

\item{year_colname_pattern}{a regex that identifies columns with year titles. 
Default is "\code{^\\d*$}" which identifies columns whose names are exclusively digits.}

\item{missing_data}{a vector containing strings that identify missing data. Default is \code{c("..", "x")}.}
}
\value{
a data frame containing the as-read IEA data
}
\description{
This function reads an IEA extended energy balances file and
converts it to a data frame with appropriately-labelled columns.
One of \code{iea_file} or \code{text} must be specified, but not both.
The first line of \code{iea_file} or \code{text} 
is expected to start with \code{expected_start_1st_line}, and
the second line is expected to start with \code{expected_2nd_line_start}, and
it may have any number of commas appended.
(The extra commas might come from opening and re-saving the file in Excel.)
If those conditions are not met, execution is halted, and
an error message is given.
}
\details{
Missing data (by default "\code{..}" and "\code{x}") are converted to \code{0}.

This function is designed to work even as more years are added
as columns at the right of \code{iea_file}, 
because column names in the output are constructed from the first two lines of \code{iea_file} 
(which contain years and country, flow, product information).

The data frame returned from this function is not ready to be used in R, 
because rows are not unique.
To further prepre the data frame for use, call \code{augment_iea_data()},
passing the output of this function in the \code{.iea_df} argument \code{augment_iea_data()}.
}
\examples{
# Original file format
iea_df(text = ",,TIME,1960,1961\\nCOUNTRY,FLOW,PRODUCT\\nWorld,Production,Hard coal,42,43")
# With extra commas on the 2nd line
iea_df(text = ",,TIME,1960,1961\\nCOUNTRY,FLOW,PRODUCT,,\\nWorld,Production,Hard coal,42,43")
}
