% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Imputation of missing values by local mean substitution}
\usage{
impute(x, n = 2)
}
\arguments{
\item{x}{numeric vector containing NAs.}

\item{n}{integer; the number of previous and following values to be included in the mean.
The default is 2 so the mean is based on a 5yr-period (including the year of the missing
value).}
}
\value{
The function returns the same numeric input vector, but with replaced missing values.
}
\description{
The function substitutes missing values in time series with a mean of the \emph{n} previous and following years
(or less depending on the position of NAs, whether it is at the beginning or end of the time series, and
the presence of further NAs in the selected time period).
}
\examples{
x <- c(NA, 1, 4, 2, 5, NA, 9, NA, 12, 11, NA)
impute(x)
}
\author{
Saskia A. Otto
}
