% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interface_industries}
\alias{interface_industries}
\title{Interface industries}
\format{
A string vector with 7 entries.
\describe{
\item{imports}{The string identifier for Imports.}
\item{exports}{The string identifier for Exports.}
\item{international_aviation_bunkers}{The string identifier for International aviation bunkers.}
\item{international_marine_bunkers}{The string identifier for International marine bunkers.}
\item{stock_changes}{The string identifier for Stock changes.}
}
}
\usage{
interface_industries
}
\description{
A string vector containing names of \code{Flow}s that interface with the world outside of the economy.
}
\examples{
interface_industries
}
\keyword{datasets}
