% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_stage_order}
\alias{last_stage_order}
\title{Last stage order}
\format{A string vector}
\usage{
last_stage_order
}
\description{
A string vector containing options for the last stage of energy conversion chain analysis.
}
\keyword{datasets}
