% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ledger_sides}
\alias{ledger_sides}
\title{IEA order for ledger side}
\format{
A string vector
}
\usage{
ledger_sides
}
\description{
A string vector containing \code{Ledger.side} entries in IEA order.
}
\examples{
ledger_sides
}
\keyword{datasets}
