% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electricity_heat_output.R
\name{load_electricity_heat_output}
\alias{load_electricity_heat_output}
\title{Extract a data frame of electricity and heat output information}
\usage{
load_electricity_heat_output(
  .iea_file = NULL,
  electricity_output_prefix = IEATools::elec_heat_output$electricity_output_prefix,
  heat_output_prefix = IEATools::elec_heat_output$heat_output_prefix,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  input_colname = IEATools::elec_heat_output$input_product,
  output_colname = IEATools::elec_heat_output$output_product,
  machine_colname = IEATools::template_cols$machine,
  unit_colname = IEATools::iea_cols$unit,
  unit = "TJ",
  electricity = "Electricity",
  output_machine_delimiter = IEATools::elec_heat_output$output_machine_delimiter,
  total = IEATools::memo_aggregation_product_prefixes$total,
  memo = IEATools::memo_aggregation_flow_prefixes$memo
)
}
\arguments{
\item{.iea_file}{The IEA data file to read}

\item{electricity_output_prefix}{The prefix for electricity output rows.
Default is \code{IEATools::elec_heat_output$electricity_output_prefix}.}

\item{heat_output_prefix}{The prefix for heat output rows.
Default is \code{IEATools::elec_heat_output$heat_output_prefix}.}

\item{country, year, flow, product, e_dot, unit_colname}{See \code{IEATools::iea_cols}.}

\item{input_colname, output_colname, output_machine_delimiter}{See \code{IEATools::elec_heat_output}.}

\item{machine_colname}{See \code{IEATools::template_cols}.}

\item{unit}{The desired output unit. Default is "TJ". Best not to change this.}

\item{electricity}{A string that defines electricity output.
Default is "Electricity".}

\item{total}{See \code{IEATools::memo_aggregation_product_prefixes}.}

\item{memo}{See \code{IEATools::memo_aggregation_flow_prefixes}.}
}
\value{
A data frame of electricity and heat output data.
}
\description{
Extract a data frame of electricity and heat output information
}
\examples{
sample_iea_data_path() |> 
  load_electricity_heat_output()
}
