% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{load_eta_fu_data}
\alias{load_eta_fu_data}
\title{Load final-to-useful machine efficiency data}
\usage{
load_eta_fu_data(
  path = sample_eta_fu_table_path(),
  eta_fu_tab_name = IEATools::fu_analysis_file_info$eta_fu_tab_name
)
}
\arguments{
\item{path}{the path from which final-to-useful machine efficiency data will be loaded. Default is the path to sample efficiency data supplied with this package.}

\item{eta_fu_tab_name}{the tab in \code{path} that contains the final-to-useful machine efficiency data. Default is "FU etas".}
}
\value{
the \code{eta_fu_tab_name} tab in \code{path} as a data frame.
}
\description{
When performing extending an energy conversion chain from useful energy to final energy,
efficiencies of final-to-useful energy conversion machines are defined by the analyst.
The Excel file at \code{path} contains those allocations.
}
\details{
A final-to-useful machine efficiencies template can be
generated using \code{eta_fu_template()} and \code{write_eta_fu_template()}.

A filled example can be loaded with the default value of \code{path}.
}
\examples{
load_eta_fu_data()
}
