% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memo_aggregation_flow_prefixes}
\alias{memo_aggregation_flow_prefixes}
\title{Memo and aggregation flow prefixes}
\format{A string vector with 3 entries}
\usage{
memo_aggregation_flow_prefixes
}
\description{
A string vector containing names of \code{Flow}s that provide memos and aggregations.
}
\keyword{datasets}
