% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memo_aggregation_product_prefixes}
\alias{memo_aggregation_product_prefixes}
\title{Memo and aggregation product prefixes}
\format{
A string vector with 2 entries.
\describe{
\item{memo}{The string identifier for "Memo: ".}
\item{total}{The string identifier for "Total".}
}
}
\usage{
memo_aggregation_product_prefixes
}
\description{
A string vector containing names of \code{Product}s that provide memos and aggregations.
}
\examples{
memo_aggregation_product_prefixes
}
\keyword{datasets}
