% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{memo_non_energy_flows}
\alias{memo_non_energy_flows}
\title{Memo Non-energy flows}
\format{
A string vector with 14 entries.
\describe{
\item{memo_non_energy_use_in_industry}{The string identifier for the "Memo: Non-energy use in industry" sector.}
\item{memo_non_energy_use_in_constructon}{The string identifier for the "Memo: Non-energy use in construction" sector.}
\item{memo_non_energy_use_in_mining_quarrying}{The string identifier for the "Memo: Non-energy use in mining and quarrying" sector.}
\item{memo_non_energy_use_in_iron_steel}{The string identifier for the "Memo: Non-energy use in iron and steel" sector.}
\item{memo_non_energy_use_in_chemical_petrochemical}{The string identifier for the "Memo: Non-energy use in chemical/petrochemical" sector.}
\item{memo_non_energy_use_in_non_ferrous_metals}{The string identifier for the "Memo: Non-energy use in non-ferrous metals" sector.}
\item{memo_non_energy_use_in_non_metallic_minerals}{The string identifier for the "Memo: Non-energy use in non-metallic minerals" sector.}
\item{memo_non_energy_use_in_transport_equipment}{The string identifier for the "Memo: Non-energy use in transport equipment" sector.}
\item{memo_non_energy_use_in_machinery}{The string identifier for the "Memo: Non-energy use in machinery" sector.}
\item{memo_non_energy_use_in_food_beverages_tobacco}{The string identifier for the "Memo: Non-energy use in food/beverages/tobacco" sector.}
\item{memo_non_energy_use_in_paper_pulp_printing}{The string identifier for the "Memo: Non-energy use in paper/pulp and printing" sector.}
\item{memo_non_energy_use_in_wood_and_wood_products}{The string identifier for the "Memo: Non-energy use in wood and wood products" sector.}
\item{memo_non_energy_use_in_textiles_leather}{The string identifier for the "Memo: Non-energy use in textiles and leather" sector.}
\item{memo_non_energy_use_in_industry_not_elsewhere_specified}{The string identifier for the "Memo: Non-energy use in industry not elsewhere specified" sector.}
}
}
\usage{
memo_non_energy_flows
}
\description{
A string vector containing names of \verb{Memo: Non-energy use in ...} \code{Flow}s
in the IEA extended energy balances database.
These flows are not always specified for a given country.
}
\examples{
non_energy_flows
}
\keyword{datasets}
