% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_order}
\alias{method_order}
\title{Method order}
\format{A string vector}
\usage{
method_order
}
\description{
A string vector containing types of methods for quantifying the primary equivalent of renewable electricity.
}
\keyword{datasets}
