% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods}
\alias{methods}
\title{Method order}
\format{
A string vector
}
\usage{
methods
}
\description{
A string vector containing types of methods for quantifying the primary equivalent of renewable electricity.
}
\examples{
methods
}
\keyword{datasets}
