% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{move_to_useful_last_stage}
\alias{move_to_useful_last_stage}
\title{Move an ECC from final to useful as its last stage}
\usage{
move_to_useful_last_stage(
  .tidy_psut_data,
  tidy_C_data,
  tidy_eta_fu_data,
  last_stage = IEATools::iea_cols$last_stage,
  final = IEATools::last_stages$final,
  useful = IEATools::last_stages$useful,
  R = IEATools::psut_cols$R,
  U_eiou = IEATools::psut_cols$U_eiou,
  U_excl_eiou = IEATools::psut_cols$U_excl_eiou,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  C_eiou = IEATools::template_cols$C_eiou,
  C_Y = IEATools::template_cols$C_Y,
  eta_fu = IEATools::template_cols$eta_fu,
  .useful = "_useful"
)
}
\arguments{
\item{.tidy_psut_data}{A tidy data frame of PSUT matrices that represent an energy conversion chain.
Matrix names should be in the \code{matnames} column, and
matrices themselves should be in the \code{matvals} column.
The last stage of these ECCs should be final (not useful).
\code{.tidy_psut_data} is likely the result of calling (in sequence)
\code{load_tidy_iea_df()} \verb{\%>\%} \code{specify_all()} \verb{\%>\%} \code{prep_psut()}}

\item{tidy_C_data}{a tidy data frame of final-to-useful allocation matrices, probably the result of calling \code{form_C_mats()}.}

\item{tidy_eta_fu_data}{a tidy data frame of final-to-useful machine efficiency matrices, probably the result of calling \code{form_eta_fu_phi_u_vecs}.}

\item{last_stage}{See \code{IEATools::iea_cols$last_stage}.
\code{last_stage} should be a column in all of \code{.tidy_psut_data}, \code{C_data}, and \code{eta_fu_data}.}

\item{final, useful}{See \code{IEATools::last_stages}.}

\item{R, U_eiou, U_excl_eiou, V, Y}{See \code{IEATools::psut_cols}.
These matrices should be found in the \code{matvals} column of the \code{.tidy_psut_data} data frame.}

\item{matnames, matvals}{See \code{IEATools::mat_meta_cols}.}

\item{C_eiou, C_Y, eta_fu}{See \code{IEATools::template_cols}.
\code{C_eiou} and \code{C_Y} matrices should be found in the \code{matvals} column of the \code{C_Y_data} data frame.
\code{eta_fu} should be found in the \code{matvals} column of the \code{eta_fu_data} data frame.}

\item{.useful}{A suffix applied to versions of PSUT matrices where useful is the last stage. Default is "_useful".}
}
\value{
a version of \code{.tidy_sut_data} that contains additional rows with useful final stage ECC matrices
}
\description{
Move an ECC from final to useful as its last stage
}
