% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_energy_flows}
\alias{non_energy_flows}
\title{Non-energy flows}
\format{
A string vector with 3 entries.
}
\usage{
non_energy_flows
}
\description{
A string vector containing names of \code{Non-energy} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
non_energy_flows
}
\keyword{datasets}
