% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notation}
\alias{notation}
\title{Notation used in IEATools}
\format{
A character vector with 2 entries
}
\usage{
notation
}
\description{
A character vector containing the notation used in the IEATools package.
Notation for delimiting specification strings is included.
}
\examples{
notation
}
\keyword{datasets}
