% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuclear_products}
\alias{nuclear_products}
\title{Nuclear products}
\format{
A string vector with 1 entries.
\describe{
\item{nuclear}{The string identifier for Nuclear}
}
}
\usage{
nuclear_products
}
\description{
A string vector containing names of products classified by the IEA as nuclear products.
}
\examples{
nuclear_products
}
\keyword{datasets}
