% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{of_notation}
\alias{of_notation}
\title{From notation}
\format{
A character vector with 4 entries
}
\usage{
of_notation
}
\description{
A character vector containing the "of" notation used in the IEATools package.
}
\examples{
of_notation
}
\keyword{datasets}
