% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other_flows}
\alias{other_flows}
\title{Other flows}
\format{
A string vector with 6 entries.
\describe{
\item{residential}{The string identifier for the Residential sector.}
\item{commercial_and_public_services}{The string identifier for the Commercial and public services sector.}
\item{agriculture_forestry}{The string identifier for the Agriculture/forestry sector.}
\item{fishing}{The string identifier for the Fishing sector.}
\item{non_specified_other}{The string identifier for the Non-specified (other) sector.}
\item{final_consumption_not_elsewhere_specified}{The string identifier for Final consumption not elsewhere specified sector.}
}
}
\usage{
other_flows
}
\description{
A string vector containing names of \code{Other} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
other_flows
}
\keyword{datasets}
