% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_funcs.R
\name{panel_funcs}
\alias{panel_funcs}
\alias{panel_cor}
\alias{panel_lines}
\alias{panel_hist}
\title{Show correlation coefficient, histogram or regression lines in pairs plot}
\usage{
panel_cor(x, y)

panel_lines(x, y)

panel_hist(x)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
\code{panel_cor}, \code{panel_lines}  and \code{panel_hist} are utility functions
for the pairs plot. \code{panel_cor} shows the Pearson product-moment correlation
coefficient in the selected panels. \code{panel_lines} shows the regression line and
\code{panel_hist} histograms.
}
\examples{
z <- matrix(rnorm(40), ncol = 4)
pairs(z, upper.panel = panel_cor,
lower.panel = panel_lines, diag.panel = panel_hist)
}
\seealso{
\code{\link{pairs}}
}
