% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peat_and_peat_products}
\alias{peat_and_peat_products}
\title{Peat and peat products}
\format{
A string vector with 2 entries.
\describe{
\item{primary_peat_products}{The string identifier for \code{primary_peat_products}.}
\item{peat_products}{The string identifier for Peat products.}
}
}
\usage{
peat_and_peat_products
}
\description{
A string vector containing names of products classified by the IEA as peat and peat products.
}
\examples{
peat_and_peat_products
}
\keyword{datasets}
