% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phi_constants_names}
\alias{phi_constants_names}
\title{Constant phi table names}
\format{
A string list with 5 entries.
\describe{
\item{phi_constants_names}{The string name of the tab in the Excel file containing the constant phi values table.}
\item{product_colname}{The string name of the energy product column in the constant phi values table.}
\item{phi_colname}{The string name of the constant phi value column in the constant phi values table.}
\item{is_useful_colname}{The string name of the IsUseful column in the constant phi values table.}
\item{phi_source_colname}{The string name of the PhiSource column in the completed phi values table.}
}
}
\usage{
phi_constants_names
}
\description{
A string list containing named names of columns and tabs for constant phi (exergy-to-energy ratio) tables.
Items in the list provide default values for column name arguments
throughout the \code{IEATools} package.
}
\examples{
phi_constants_names
}
\keyword{datasets}
