% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_statespace_ch.R
\name{plot_statespace_ch}
\alias{plot_statespace_ch}
\title{Convex hull plot}
\usage{
plot_statespace_ch(x, col_ch_ref = "red", col_ch_cur = "blue", size_time = 4)
}
\arguments{
\item{x}{An object from the \code{\link{statespace_ch}} function.}

\item{col_ch_ref}{Color of reference period (for points, path, and labels).}

\item{col_ch_cur}{Color of current period (for points, path, and labels).}

\item{size_time}{Text size of the time labels (both periods).}
}
\value{
The function returns a \code{\link[ggplot2]{ggplot}} object.
}
\description{
\code{plot_statespace_ch} generates a scatter plot of all observed combinations
in the 2-dimensional indicator space including the convex hull from defined
reference conditions and the current period.
}
\examples{
x <- rnorm(25)
y <- rnorm(25)
ch <- statespace_ch(x, y, time = 1981:2005, period_ref = 1981:1986, period_current = 1995:2005)
plot_statespace_ch(ch)
}
\author{
Saskia A. Otto
}
