% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_statespace_ed.R
\name{plot_statespace_ed}
\alias{plot_statespace_ed}
\title{Time series plot of Euclidean distance}
\usage{
plot_statespace_ed(x)
}
\arguments{
\item{x}{The output tibble from the \code{\link{statespace_ed}} function.}
}
\value{
The function returns a \code{\link[ggplot2]{ggplot}} object.
}
\description{
\code{plot_statespace_ed} generates a time series plot of the Euclidean
distance in indicator state space from a defined reference conditions.
}
\examples{
x <- as.data.frame(matrix(rnorm(250), ncol = 10))
ed <- statespace_ed(x, time = 1976:2000, ref_time = 1976)
plot_statespace_ed(ed)
}
\author{
Saskia A. Otto
}
