% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{prep_psut}
\alias{prep_psut}
\title{Prepare for PSUT analysis}
\usage{
prep_psut(
  .tidy_iea_df,
  cols = IEATools::iea_cols,
  ledger_side = cols$ledger_side,
  flow_aggregation_point = cols$flow_aggregation_point,
  flow = cols$flow,
  product = cols$product,
  unit = cols$unit,
  e_dot = cols$e_dot,
  supply = "Supply",
  consumption = "Consumption",
  matnames = "matnames",
  matvals = "matvals",
  rownames = "rownames",
  colnames = "colnames",
  rowtypes = "rowtypes",
  coltypes = "coltypes"
)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame that has been specified with \code{specify_all()}.}

\item{cols}{a list containing names of columns. Default is \code{IEATools::iea_cols}.}

\item{ledger_side}{the name of the ledger side column. Default is \code{cols$ledger_side}.}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is \code{cols$flow_aggregation_point}.}

\item{flow}{the name of the flow column. Default is \code{cols$flow}.}

\item{product}{the name of the product column. Default is \code{cols$product}.}

\item{unit}{the name of the unit column. Default is \code{cols$unit}.}

\item{e_dot}{the name of the energy rate column. Default is \code{cols$e_dot}.}

\item{supply}{the string identifying the supply side of the ledger. Default is "Supply".}

\item{consumption}{the string identifying the consumption side of the ledger. Default is "Consumption".}

\item{matnames}{the name of the matrix names column added by this function. Default is "matnames".}

\item{matvals}{the name of the matrix value column added by this function. Default is "matvals".}

\item{rownames}{the name of the rownames column passed to \code{collapse_to_tidy_psut()}. Default is "rownames".}

\item{colnames}{the name of the colnames column passed to \code{collapse_to_tidy_psut()}. Default is "colnames".}

\item{rowtypes}{the name of the rowtypes column passed to \code{collapse_to_tidy_psut()}. Default is "rowtypes".}

\item{coltypes}{the name of the coltypes column passed to \code{collapse_to_tidy_psut()}. Default is "coltypes".}
}
\value{
a tidy PSUT data frame
}
\description{
Converts a tidy IEA data frame into a PSUT data frame
by collapsing the IEA data into PSUT matrices (R, U, V, and Y).
}
\details{
This function bundles several others:
\enumerate{
\item \code{add_psut_matnames()}
\item \code{add_row_col_meta()}
\item \code{collapse_to_tidy_psut()}
}

Furthermore, it extracts \code{S_units} matrices using \code{extract_S_units_from_tidy()}
and adds those matrices to the data frame.
}
\examples{
library(dplyr)
library(tidyr)
Simple <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  prep_psut() \%>\% 
  rename(matval_simple = matvals)
S_units <- load_tidy_iea_df() \%>\% 
  extract_S_units_from_tidy()
Complicated <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  add_psut_matnames() \%>\% 
  add_row_col_meta() \%>\% 
  collapse_to_tidy_psut() \%>\% 
  spread(key = matnames, value = matvals) \%>\% 
  full_join(S_units, by = c("Method", "Energy.type", "Last.stage", 
                            "Country", "Year")) \%>\% 
  gather(key = matnames, value = matvals, R, U_EIOU, U_excl_EIOU, 
                                        V, Y, S_units) \%>\% 
  rename(matval_complicated = matvals)
# Simple and Complicated are same.
full_join(Simple, Complicated, by = c("Method", "Energy.type", 
                                      "Last.stage", "Country", 
                                      "Year", "matnames")) \%>\% 
  dplyr::mutate(
    same = matsbyname::equal_byname(matval_simple, matval_complicated)
  ) \%>\% 
  magrittr::extract2("same") \%>\% 
  as.logical() \%>\% 
  all()
}
