% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prim_agg_flows}
\alias{prim_agg_flows}
\title{Primary aggregates flows}
\format{
A string vector with 3 entries.
\describe{
\item{resources}{The string identifier for Resource flows.}
\item{imports}{The string identifier for Import flows.}
\item{stock_changes}{The string identifier for Stock changes flows.}
}
}
\usage{
prim_agg_flows
}
\description{
A string vector containing a limited set of names of \verb{Total primary energy supply} \code{Flow}s in
the IEA extended energy balances database. Used for calculating domestic primary energy consumption
using \code{Recca::primary_aggregates()}
}
\examples{
prim_agg_flows
}
\keyword{datasets}
