% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_coal_products}
\alias{primary_coal_products}
\title{Primary coal products}
\format{
A string vector with 7 entries.
\describe{
\item{hard_coal_if_no_detail}{The string identifier for Hard coal (if no detail).}
\item{brown_coal_if_no_detail}{The string identifier for Brown coal (if no detail).}
\item{anthracite}{The string identifier for Anthracite.}
\item{coking_coal}{The string identifier for Coking coal.}
\item{other_bituminous_coal}{The string identifier for Other bituminous coal.}
\item{sub_bituminous_coal}{The string identifier for Sub-bituminous coal.}
\item{lignite}{The string identifier for Lignite.}
}
}
\usage{
primary_coal_products
}
\description{
A string vector containing names of products classified by the IEA as primary coal products.
}
\examples{
primary_coal_products
}
\keyword{datasets}
