% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_gas_products}
\alias{primary_gas_products}
\title{Primary gas products}
\format{
A string vector with 1 entries.
}
\usage{
primary_gas_products
}
\description{
A string vector containing names of products classified by the IEA as primary gas products.
}
\examples{
primary_gas_products
}
\keyword{datasets}
