% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_oil_products}
\alias{primary_oil_products}
\title{Primary oil products}
\format{
A string vector with 6 entries.
\describe{
\item{crude_ngl_feedstocks_if_no_detail}{The string identifier for Crude/NGL/feedstocks (if no detail).}
\item{crude_oil}{The string identifier for Crude oil.}
\item{natural_gas_liquids}{The string identifier for Natural gas liquids.}
\item{additives_blending_components}{The string identifier for Additives/blending components.}
\item{other_hydrocarbons}{The string identifier for Other hydrocarbons.}
\item{oil_shales_and_oil_sands}{The string identifier for Oil shale and oil sands.}
}
}
\usage{
primary_oil_products
}
\description{
A string vector containing names of products classified by the IEA as primary oil products.
}
\examples{
primary_oil_products
}
\keyword{datasets}
