% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{products}
\alias{products}
\title{IEA order for products}
\format{
A string vector
}
\usage{
products
}
\description{
A string vector containing entries for the product column in IEA order.
}
\examples{
products
}
\keyword{datasets}
