% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocate.R
\name{reallocate_industry_nes}
\alias{reallocate_industry_nes}
\title{Reallocate Industry not elsewhere specified}
\usage{
reallocate_industry_nes(
  .sutmats,
  industry_nes = c(IEATools::industry_flows$non_specified_industry,
    IEATools::industry_flows$industry_not_elsewhere_specified),
  industry_sectors = IEATools::industry_flows,
  Y = IEATools::psut_cols$Y,
  Y_prime = "Y_prime"
)
}
\arguments{
\item{.sutmats}{A data frame of PSUT matrices,
most likely the reult of \code{\link[=prep_psut]{prep_psut()}}.}

\item{industry_nes}{A vector of industry names (columns in the \strong{Y} matrix)
that indicate unspecified industries.
Default is \code{c(IEATools::industry_flows$non_specified_industry, IEATools::industry_flows$industry_not_elsewhere_specified)}
to account for the IEA's nomenclature change in 2019.}

\item{industry_sectors}{A vector of industry names (columns in the \strong{Y} matrix)
that consume energy,
including the "Industry not elsewhere specified"
columns included in \code{industry_nes}.
Default is \code{IEATools::industry_flows}.}

\item{Y}{The name of the \code{Y} column in \code{.sutmats}.
Default is \link{psut_cols}\verb{$Y}.}

\item{Y_prime}{The name of the modified \code{Y} column in which
"Industry not elsewhere specified" has been reallocated.
Default is "Y_prime".}
}
\value{
A version of \code{.sutmats} in which energy consumption by "Industry not elsewhere specified"
is reallocated to other Industries in proportion to their energy consumption.
}
\description{
The IEA data include a final demand sector called
"Industry not elsewhere specified".
Sometimes, such as when performing decomposition analysis,
it is desirable to reallocate "Industry not elsewhere specified"
to other industries in proportion to the consumption of each
energy carrier in specified industries.
This function performs that reallocation.
}
\details{
This function calls \code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}
internally to perform the reallocation.
}
