% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocate.R
\name{reallocate_statistical_differences}
\alias{reallocate_statistical_differences}
\title{Reallocate Statistical differences}
\usage{
reallocate_statistical_differences(
  .sutmats,
  stat_diffs = IEATools::tfc_compare_flows$statistical_differences,
  R = IEATools::psut_cols$R,
  U_feed = IEATools::psut_cols$U_feed,
  U = IEATools::psut_cols$U,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  prime_suffix = "_prime",
  R_prime = paste0(R, prime_suffix),
  U_feed_prime = paste0(U_feed, prime_suffix),
  U_prime = paste0(U, prime_suffix),
  Y_prime = paste0(Y, prime_suffix)
)
}
\arguments{
\item{.sutmats}{A data frame of PSUT matrices,
most likely the result of \code{\link[=prep_psut]{prep_psut()}}.}

\item{stat_diffs}{The name of the row in \strong{R} and columns in \strong{U_feed} and \strong{Y}
for Statistical differences.
Default is \code{IEATools::tfc_compare_flows$statistical_differences}.}

\item{R, U_feed, U, V, Y}{Names of columns of matrices in \code{.sutmats}.
See \link{psut_cols}.}

\item{prime_suffix}{The string suffix for new versions of matrices with reallocated
statistical differences.
Default is "_prime".}

\item{R_prime, U_feed_prime, U_prime}{}
}
\value{
A version of \code{.sutmats} in which energy consumption by "Statistical differences"
is reallocated to other Industries in proportion to their energy consumption.
}
\description{
The IEA data include "Statistical differences".
In some cases, it is desirable to reallocate Statistical differences
in proportion to the non-zero consumption of each
energy carrier in other industries.
This function performs that reallocation
and should be called after the PSUT matrices have been formed,
most likely by calling \code{\link[=prep_psut]{prep_psut()}}.
}
\details{
Statistical differences can be found in either the \strong{R} or \strong{Y} matrix.
Both are reallocated to the \strong{Y} and \strong{U} matrices in proportion.
The steps are:
\itemize{
\item Move Statistical differences found in the \strong{R} matrix to the \strong{Y} matrix by subtraction.
\item Reallocate Statistical differences in the \strong{Y} matrix to the \strong{Y} and \strong{U}
matrices using \code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
}

Internally, the \strong{Y} and \strong{U_feed} matrices are added before calling
\code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
The matrices are split again prior to returning.
}
