% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{remove_agg_regions}
\alias{remove_agg_regions}
\title{Remove aggregation regions from an IEA data frame}
\usage{
remove_agg_regions(
  .iea_df,
  country = IEATools::iea_cols$country,
  agg_regions = IEATools::aggregation_regions
)
}
\arguments{
\item{.iea_df}{The IEA data frame from which you want to remove aggregation regions.}

\item{country}{The name of the Country column in \code{.iea_df}.
Default is \code{IEATools::iea_cols$country}.}

\item{agg_regions}{A list of aggregation regions in the \code{country} column of \code{.iea_df}.
Default is \code{IEATools::aggregation_regions}.}
}
\value{
A version of \code{.iea_df} with aggregation regions removed.
}
\description{
The IEA extended energy balances contain several aggregation regions
by default.
In some situations, it may be desirable to remove those aggregation regions.
This function performs that task.
}
\examples{
tibble::tibble(Year = c(1967, 1995), 
               Country = c("World", "Spain")) \%>\%
 remove_agg_regions()
}
