% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renewable_industries}
\alias{renewable_industries}
\title{Renewable energy industries names}
\format{
A string list with 5 entries.
\describe{
\item{geothermal_plants}{The string name of geothermal plants.}
\item{hydro_plants}{The string name of hydropower plants.}
\item{solar_pv_plants}{The string name of solar photovoltaics plants.}
\item{solar_th_plants}{The string name of solar thermal plants.}
\item{oceanic_plants}{The string name of oceanic power plants.}
\item{wind_power_plants}{The string name of wind power plants.}
}
}
\usage{
renewable_industries
}
\description{
A string list containing the names of renewable industries added with the \code{specify_renewable_plants()} function.
}
\examples{
renewable_industries
}
\keyword{datasets}
