% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{replace_join}
\alias{replace_join}
\title{\code{left_join} with replacement}
\usage{
replace_join(
  x,
  y,
  replace_col,
  by = dplyr::intersect(names(x), names(y)) \%>\% dplyr::setdiff(replace_col),
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{by}{}

\item{copy}{}

\item{suffix}{}

\item{...}{}
}
\value{

}
\description{
This function adds a replacement join type to the various \verb{*_join} functions in the dplyr package.
\code{replace_join()} returns all rows from \code{x} and all columns from \code{x}.
}
