% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{replace_null_RUV}
\alias{replace_null_RUV}
\title{Fill missing \strong{R}, \strong{U}, and \strong{V} matrices}
\usage{
replace_null_RUV(
  .sutmats = NULL,
  R = IEATools::psut_cols$R,
  U_feed = IEATools::psut_cols$U_feed,
  U_eiou = IEATools::psut_cols$U_eiou,
  U = IEATools::psut_cols$U,
  r_eiou = IEATools::psut_cols$r_eiou,
  Y = IEATools::psut_cols$Y,
  V = IEATools::psut_cols$V,
  resources = IEATools::tpes_flows$resources,
  .R_temp_name = ".R_temp",
  .U_temp_name = ".U_temp",
  .U_feed_temp_name = ".U_feed_temp",
  .U_eiou_temp_name = ".U_EIOU_temp",
  .r_eiou_temp_name = ".r_EIOU_temp",
  .V_temp_name = ".V_temp",
  R_name = IEATools::psut_cols$R,
  U_name = IEATools::psut_cols$U,
  U_feed_name = IEATools::psut_cols$U_feed,
  U_eiou_name = IEATools::psut_cols$U_eiou,
  r_eiou_name = IEATools::psut_cols$r_eiou,
  V_name = IEATools::psut_cols$V
)
}
\arguments{
\item{.sutmats}{A data frame of metadata columns and matrix name columns}

\item{R, U_feed, U_eiou, U, r_eiou, Y, V}{See \code{IEATools::psutcols}. Default values are names for variables incoming with \code{.sutmats}. Can be overridden with actual matrices.}

\item{resources}{See \code{IEATools::tpes_flows}. The name of the only row of the output \strong{0} \strong{R} matrix.}

\item{.R_temp_name, .U_temp_name, .U_feed_temp_name, .U_eiou_temp_name, .r_eiou_temp_name, .V_temp_name}{Names of temporary variables unused internally to the function.}

\item{R_name, U_name, U_feed_name, U_eiou_name, r_eiou_name, V_name}{See \code{IEATools::psutcols}. The final names for matrices in the output.}
}
\value{
A version of \code{.sutmats} with \strong{R}, \strong{U}, \strong{U_feed}, \strong{U_EIOU}, or \strong{V} filled with \strong{0} matrices if they were missing.
}
\description{
In some cases (e.g., bunkers where \code{Last.stage} is "final"),
\strong{R}, \strong{U}, \strong{U_feed}, \strong{U_EIOU}, or \strong{V} matrices can be missing, because
imports which appear in the \strong{V} matrix (or \strong{R} matrix) are consumed in final demand (\strong{Y}) matrix,
without any intermediate processing.
When a data frame is pivoted wider by matrices,
the \strong{R}, \strong{U_feed}, and \strong{U_EIOU} columns will contain \code{NULL} entries.
This function fills those \code{NULL} entries with reasonable defaults.
}
\details{
Reasonable defaults arise from the following thought processes.
If all energy is supplied by imports (in the \strong{V} matrix),
there are no resources.
Thus, we can replace the missing \strong{R} matrix with a \strong{0} matrix with a generic
"Natural resources" row and the same products as the rows of the \strong{Y} matrix.

Similarly, missing values for \strong{U}, \strong{U_feed}, \strong{U_EIOU}, or \strong{r_EIOU} can be replaced by a \strong{0} matrix
with row and column names same as a transposed \strong{V} matrix when it exists.
If neither \strong{U} nor \strong{V} exist, the \strong{R} matrix can supply row and column names.
}
\examples{
# Set up a PSUT data frame with NULL for
# R, U_feed, and U_EIOU in 1971 for GHA.
psut <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  prep_psut() \%>\% 
  tidyr::pivot_longer(cols = c("R", "U_EIOU", "U_feed", "U", "r_EIOU", "V", "Y", "S_units"),
                      names_to = "matnames", values_to = "matvals") \%>\% 
  dplyr::filter(!(Country == "GHA" & Year == 1971 & matnames == "R")) \%>\% 
  dplyr::filter(!(Country == "GHA" & Year == 1971 & matnames == "U_feed")) \%>\% 
  dplyr::filter(!(Country == "GHA" & Year == 1971 & matnames == "U_EIOU")) \%>\% 
  tidyr::pivot_wider(names_from = "matnames", values_from = "matvals")
# Replace the `NULL` matrices in the first row.
res <- psut \%>\% 
  replace_null_RUV()
res$R[[1]]
res$U_feed[[1]]
res$U_EIOU[[1]]
}
