% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{replace_null_UV}
\alias{replace_null_UV}
\title{Fill missing \strong{U} and \strong{V} matrices}
\usage{
replace_null_UV(
  .sutmats = NULL,
  R = IEATools::psut_cols$R,
  U_feed = IEATools::psut_cols$U_feed,
  U_eiou = IEATools::psut_cols$U_eiou,
  U = IEATools::psut_cols$U,
  r_eiou = IEATools::psut_cols$r_eiou,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  manufacture = IEATools::transformation_processes$manufacture,
  manufacture_notation = RCLabels::of_notation,
  resources_ind_notation = RCLabels::of_notation,
  raw_product_notation = RCLabels::from_notation,
  .R_temp_name = ".R_temp",
  .U_temp_name = ".U_temp",
  .U_feed_temp_name = ".U_feed_temp",
  .U_eiou_temp_name = ".U_EIOU_temp",
  .r_eiou_temp_name = ".r_EIOU_temp",
  .V_temp_name = ".V_temp",
  R_name = IEATools::psut_cols$R,
  U_name = IEATools::psut_cols$U,
  U_feed_name = IEATools::psut_cols$U_feed,
  U_eiou_name = IEATools::psut_cols$U_eiou,
  r_eiou_name = IEATools::psut_cols$r_eiou,
  V_name = IEATools::psut_cols$V,
  tol = 1e-06
)
}
\arguments{
\item{.sutmats}{A data frame of metadata columns and named matrix columns}

\item{R, U_feed, U_eiou, U, r_eiou, V, Y}{See \code{IEATools::psutcols}.
Default values are names for variables
incoming with \code{.sutmats}.
Can be overridden with actual matrices.}

\item{manufacture}{The string name of the "Manufacture" industry.
Default is \code{IEATools::transformation_processes$manufacture}.}

\item{manufacture_notation, resources_ind_notation, raw_product_notation}{Row and column
label notations for the manufacture industry, the resources industry,
and raw material products.
Defaults are \code{RCLabels::of_notation}, \code{RCLabels::of_notation}, and
\code{RCLabels::from_notation}.}

\item{.R_temp_name, .U_temp_name, .U_feed_temp_name, .U_eiou_temp_name, .r_eiou_temp_name, .V_temp_name}{Names of temporary variables unused internally to the function.}

\item{R_name, U_name, U_feed_name, U_eiou_name, r_eiou_name, V_name}{See \code{IEATools::psutcols}.
The final names
for matrices in the output.}

\item{tol}{The tolerance for deciding whether energy flows are in balance.
Default is \code{1e-6}.}
}
\value{
A version of \code{.sutmats}
with \strong{R}, \strong{U}, \strong{U_feed}, \strong{U_EIOU}, or \strong{V}
filled with \strong{0} matrices if they were missing.
}
\description{
In some cases (e.g., bunkers where \code{Last.stage} is "final"),
\strong{U}, \strong{U_feed}, \strong{U_EIOU}, and \strong{V} matrices can be missing, because
Imports (which appear in the \strong{R} matrix) are consumed directly
in the final demand (\strong{Y}) matrix,
without any intermediate processing.
When a data frame is pivoted wider by matrices,
the \strong{U_feed}, \strong{U_EIOU}, \strong{r_EIOU}, and \strong{V} columns
will contain \code{NULL} entries.
This function fills those \code{NULL} entries with reasonable defaults.
}
\details{
Reasonable defaults are obtained by introducing one new
industry for each product in the \strong{R} and \strong{Y} matrices.
The new industry is called
"Manufacture [of XYZ]", where "XYZ" is the name of the product.
The new "Manufacture" industries have 100\% efficiency.
}
\examples{
# Set up a PSUT data frame with NULL for
# U_feed, U_EIOU, and V in 1971 for GHA.
  R <- matrix(c(10, 0, 
                0, 200), nrow = 2, ncol = 2, byrow = TRUE, 
              dimnames = list(c("Imports [of Fuel oil]", 
                                "Imports [of Aviation gasoline]"), 
                              c("Fuel oil", "Aviation gasoline")))
  Y <- matrix(c(10, 0, 
                0, 200), nrow = 2, ncol = 2, byrow = TRUE, 
              dimnames = list(c( "Fuel oil", "Aviation gasoline"), 
                              c("International marine bunkers", 
                                "International aviation bunkers")))
# Replace the missing **U** and **V** matrices.
replace_null_UV(R = R, Y = Y)
}
