% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{route_non_specified_flows}
\alias{route_non_specified_flows}
\title{Routes non specified flows}
\usage{
route_non_specified_flows(
  .tidy_iea_df,
  route_non_specified_eiou = TRUE,
  route_non_specified_tp = TRUE
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{is_non_specified_eiou_routed}{A boolean that indicates whether non specified EIOU flows should be routed to existing
industries or kept as non specified.}

\item{is_non_specified_tp_routed}{A boolean that indicates whether non specified transformation processes flows should be routed to existing
industries or kept as non specified.}
}
\value{
A modified version of the \code{.tidy_iea_df} with non specified flows routed to existing industries.
}
\description{
This function is a wrapper of the functions
\code{route_non_specified_eiou} and \code{route_non_specified_tp}.
It is called within the \code{specify_all()} function.
}
\details{
See \code{route_non_specified_eiou} and \code{route_non_specified_tp} functions documentations for additional details.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  route_non_specified_flows()
}
