% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{route_non_specified_tp}
\alias{route_non_specified_tp}
\title{Routes non-specified transformation processes flows to existing industries}
\usage{
route_non_specified_tp(
  .tidy_iea_df,
  routing_non_specified_tp = TRUE,
  country = IEATools::iea_cols$country,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  ledger_side = IEATools::iea_cols$ledger_side,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  non_spec = "Non-specified",
  negzeropos = ".negzeropos",
  n_counting = ".n_counting",
  Total_input_output_by_prod_excl_nonspec = ".Total_input_output_by_prod_excl_nonspec",
  Input_output_by_prod_per_tp = ".Input_output_by_prod_per_tp",
  Share_input_output_by_prod_per_tp = ".Share_input_output_by_prod_per_tp",
  destination_flow = ".destination_flow"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{routing_non_specified_tp}{A boolean indicating whether non-specified EIOU flows should be redirected to other existing industries.
If FALSE, the function returns the input data frame.
Default is TRUE.}

\item{country}{The name of the country column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$country}.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow_aggregation_point}.}

\item{flow}{The name of the flow column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{ledger_side}{The name of the ledger side column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$ledger_side}.}

\item{method}{The name of the method column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$method}.}

\item{energy_type}{The name of the energy type column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$energy_type}.}

\item{last_stage}{The name of the last stage column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$last_stage}.}

\item{year}{The name of the country year in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$year}.}

\item{product}{The name of the product column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$product}.}

\item{unit}{The name of the unit column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$unit}.}

\item{e_dot}{The name of the energy column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$energy}.}

\item{transformation_processes}{A string that identifies "Transformation processes" in the \code{flow_aggregation_point} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::aggregation_flows$transformation_processes}.}

\item{non_spec}{A string that identifies "Non-specified" flows in the \code{flow} column of the \code{.tidy_iea_df}.
Default is "Non-specified".}

\item{negzeropos}{The name of a temporary column added to the data frame.
Default is ".negzeropos".}

\item{n_counting}{The name of a temporary column added to the data frame.
Default is ".n_counting".}

\item{Total_input_output_by_prod_excl_nonspec}{The name of a temporary column added to the data frame.
Default is ".Total_input_output_by_prod_excl_nonspec_From_Func".}

\item{Input_output_by_prod_per_tp}{The name of a temporary column added to the data frame.
Default is ".Input_output_by_prod_per_tp_From_Func".}

\item{Share_input_output_by_prod_per_tp}{The name of a temporary column added to the data frame.
Default is ".Share_input_output_by_prod_per_tp_From_Func".}

\item{destination_flow}{The name of a temporary column added to the data frame.
Default is ".destination_flow".}
}
\value{
A modified version of the \code{.tidy_iea_df} with non-specified transformation processes flows routes to existing industries.
}
\description{
This function routes non-specified transformation processes flows to existing industries.
It does so using the shares of product use and supply of the other transformation processes.
If no transformation processes consume or supply a product that is present in the non-specified flows,
then the flow remains non-specified.
The function is called within the \code{route_non_specified_flows()} function.
Note that the \code{routing_non_specified_eiou} parameter enables to switch on and off the routing of the non-specified EIOU flow.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  route_non_specified_tp()
}
