% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{route_own_use_elect_chp_heat}
\alias{route_own_use_elect_chp_heat}
\title{Routes own use in electricity, chp, and heat plants EIOU flow to main activity producer flows}
\usage{
route_own_use_elect_chp_heat(
  .tidy_iea_df,
  split_using_shares_of = c("input", "output"),
  country = IEATools::iea_cols$country,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  ledger_side = IEATools::iea_cols$ledger_side,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  supply = IEATools::ledger_sides$supply,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  own_use_elect_chp_heat = IEATools::eiou_flows$own_use_elect_chp_heat_plants,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  main_act_producer_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  negzeropos = ".negzeropos",
  n_counting = ".n_counting",
  destination_flow = ".destination_flow",
  Total_main_activity_From_Func = ".Total_main_activity_From_Func",
  Total_per_main_activity_From_Func = ".Total_per_main_activity_From_Func",
  Share_per_main_activity_From_Func = ".Share_per_main_activity_From_Func"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{split_using_shares_of}{A string that identifies which method is to be used for splitting the \code{own_use_elect_chp_heat} flow.
Default is "input". The other valid value is "output". See details for more information.}

\item{country}{The name of the country column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$country}.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow_aggregation_point}.}

\item{flow}{The name of the flow column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{ledger_side}{The name of the ledger side column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$ledger_side}.}

\item{method}{The name of the method column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$method}.}

\item{energy_type}{The name of the energy_type column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$energy_type}.}

\item{last_stage}{The name of the last stage column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$last_stage}.}

\item{year}{The name of the year column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{product}{The name of the product column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$product}.}

\item{unit}{The name of the unit column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$unit}.}

\item{e_dot}{The name of the energy column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{supply}{A string identifying "Supply" in the \code{ledger_side} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::ledger_sides$supply}.}

\item{eiou}{A string identifying "Energy industry own use" in the \code{flow_aggregation_point} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::aggregation_flows$energy_industry_own_use}.}

\item{transformation_processes}{A string identifying transformation processes in the \code{flow_aggregation_point} column of the \code{.tidy_iea_df}
Default is \code{IEATools::aggregation_flows$flow_aggregation_point}.}

\item{own_use_elect_chp_heat}{A string identifying "Own use in electricity, CHP and heat plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::eiou_flows$own_use_elect_chp_heat_plants}.}

\item{main_act_producer_elect}{A string identifying "Main activity producer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_elect_plants}.}

\item{main_act_producer_chp}{A string identifying "Main activity producer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_heat_plants}.}

\item{main_act_producer_heat}{A string identifying "Main activity producer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_heat_plants}.}

\item{negzeropos}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".negzeropos".}

\item{n_counting}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".n_counting".}

\item{destination_flow}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".destination_flow".}

\item{Total_main_activity_From_Func}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".Total_main_activity_From_Func".}

\item{Total_per_main_activity_From_Func}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".Total_per_main_activity_From_Func".}

\item{Share_per_main_activity_From_Func}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".Share_per_main_activity_From_Func".}
}
\value{
A modified version of the \code{.tidy_iea_df}, in which the \code{own_use_elect_chp_heat} flow has been routed to the main activity producer industries.
}
\description{
This function routes the "Own use in electricity, CHP and heat plants"
Energy Industry Own Use flow to each of the three electricity, CHP and heat
main activity producer plants.
The function is called within the \code{specify_all()} function.
}
\details{
The function either performs the routing using the share of outputs or of inputs of each
of the three main activity producer industries.

The method is selected using the \code{split_using_shares_of} argument,
to which either the have the "input" or "output" value can be passed.

When none of the main activity producer industries is present in the data frame,
the "Own use in electricity, CHP and heat plants" flow is ascribed by default to the
"Main activity producer electricity plant".
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  route_own_use_elect_chp_heat()
}
