% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{route_pumped_storage}
\alias{route_pumped_storage}
\title{Route pumped storage to Main activity electricity producer plant}
\usage{
route_pumped_storage(
  .tidy_iea_df,
  specify_renewable_plants = FALSE,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  year = IEATools::iea_cols$year,
  last_stage = IEATools::iea_cols$last_stage,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  pumped_storage = IEATools::eiou_flows$pumped_storage_plants,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  hydro_plants = IEATools::renewable_industries$hydro_plants,
  hydro = IEATools::renewable_products$hydro,
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{specify_renewable_plants}{A boolean indicating whether renewable energy plants should be specified or not.
Default is FALSE.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow_aggregation_point}.}

\item{country, method, energy_type, last_stage, unit, year, product}{See \code{IEATools::iea_cols}.}

\item{flow}{The name of the flow column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{e_dot}{The name of the energy column in the \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{eiou}{A string identifying "Energy industry own use" in the \code{flow_aggregation_point} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::aggregation_flows$energy_industry_own_use}.}

\item{pumped_storage}{A string identifying "Pumped storage plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::eiou_flows$pumped_storage_plants}.}

\item{main_act_producer_elect}{A string identifying "Main activity producer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_elect_plants}.}

\item{hydro_plants}{The name of the newly created hydropower industry.
Default is \code{IEATools::renewable_industries$hydro_plants}.}

\item{hydro}{The name of the "Hydro" product.
Default is \code{IEATools::renewable_products$hydro}.}

\item{negzeropos}{The name of a temporary column created in \code{.tidy_iea_df}.
Default is ".negzeropos".}
}
\value{
A modified \code{.tidy_iea_df} with "Pumped storage plants" industry routed
to the "Main activity producer electricity plant" industry.
}
\description{
The function routes Energy industry own use by Pumped storage plants
to Energy industry own use by Main activity producer electricity plants
when the value is negative, as it should be.
However, for Japan, there are a few years where
Energy industry own use by Pumped storage plants is positive.
In those instances, the Flow is changed from
Energy industry own use to
Main activity producer electricity plants \emph{and}
the Flow.aggregation.point is changed to "Transformation processes"
the positive value is retained.
This approach preserves the overall energy balance.
This approach implicitly assumes that Japan's reported
Energy industry own use for Pumped storage plants
is a net value, not a total value.
All other countries seemingly report total values for
Energy industry own use by Pumped Storage plants,
as those values are all negative.
This function is called within the \code{specify_all()} function.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  gather_producer_autoproducer()
}
