% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_fu_allocation_table_path}
\alias{sample_fu_allocation_table_path}
\title{The path to a filled final-to-useful allocation table}
\usage{
sample_fu_allocation_table_path(version = 2021)
}
\arguments{
\item{version}{the desired version (expressed as the year of IEA data release) of
the sample final-to-useful allocation table.
Options are 2021 (default), 2019, and 2018.}
}
\value{
the path to a final-to-useful allocation table
}
\description{
The path to a filled final-to-useful allocation table
}
\examples{
sample_fu_allocation_table_path()     # Assumes 2021
sample_fu_allocation_table_path(2021) # Same
# Returns path for sample allocation table appropriate for other IEA data releases
sample_fu_allocation_table_path(2019) 
sample_fu_allocation_table_path(2018) 
}
