% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_iea_data_path}
\alias{sample_iea_data_path}
\title{The path to a sample data frame}
\usage{
sample_iea_data_path()
}
\value{
the path to a sample data file.
}
\description{
The sample data file is in .csv format and contains for Ghana and South Africa for 1971 and 2000.
}
\details{
Permission to include this data was provided in a phone conversation between
Nick Johnstone (IEA Chief Statistician) and Matthew Kuperus Heun (IEATools developer)
on Monday, 3 June 2019.
}
\examples{
sample_iea_data_path()
}
