% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_to_exergy.R
\name{sample_phi_constants_path}
\alias{sample_phi_constants_path}
\title{Gives a file path to a sample phi constants table}
\usage{
sample_phi_constants_path()
}
\value{
A path to a sample phi constants table bundled with this package.
}
\description{
Gives a file path to a sample phi constants table
}
\examples{
sample_phi_constants_path()
}
