% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secondary_oil_products}
\alias{secondary_oil_products}
\title{Secondary oil products}
\format{
A string vector with 18 entries.
\describe{
\item{refinery_feedstocks}{The string identifier for Refinery feedstocks.}
\item{refinery_gas}{The string identifier for Refinery gas.}
\item{ethane}{The string identifier for Ethane.}
\item{liquefied_petroleum_gases_lpg}{The string identifier for Liquefied petroleum gases (LPG).}
\item{motor_gasoline_excl_biofuels}{The string identifier for Motor gasoline excl. biofuels.}
\item{aviation_gasoline}{The string identifier for Aviation gasoline.}
\item{gasoline_type_jet_fuel}{The string identifier for Gasoline type jet fuel.}
\item{kerosene_type_jet_fuel_excl_biofuels}{The string identifier for Kerosene type jet fuel excl. biofuels.}
\item{other_kerosene}{The string identifier for Other kerosene.}
\item{gas_diesel_oil_excl_biofuels}{The string identifier for Gas/diesel oil excl. biofuels.}
\item{fuel_oil}{The string identifier for Fuel oil.}
\item{naptha}{The string identifier for Naphtha.}
\item{white_spirit_SBP}{The string identifier for White spirit & SBP.}
\item{lubricants}{The string identifier for Lubricants.}
\item{bitumen}{The string identifier for Bitumen.}
\item{paraffin_waxes}{The string identifier for Paraffin waxes.}
\item{petroleum_coke}{The string identifier for Petroleum coke.}
\item{other_oil_products}{The string identifier for Other oil products.}
}
}
\usage{
secondary_oil_products
}
\description{
A string vector containing names of products classified by the IEA as \code{oil_and_oil_products} that are not \code{primary_coal_products}.
}
\examples{
secondary_oil_products
}
\keyword{datasets}
