% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secondary_peat_products}
\alias{secondary_peat_products}
\title{Secondary peat products}
\format{
A string vector with 1 entries.
\describe{
\item{peat_products}{The string identifier for Peat products.}
}
}
\usage{
secondary_peat_products
}
\description{
A string vector containing names of products classified by the IEA as "Peat and peat products" that are not \code{primary_peat_products}.
}
\examples{
secondary_peat_products
}
\keyword{datasets}
