% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sort_iea_df}
\alias{sort_iea_df}
\title{Sort an IEA data frame in IEA row order}
\usage{
sort_iea_df(
  .iea_df,
  col_names = IEATools::iea_cols,
  country = col_names$country,
  method = col_names$method,
  energy_type = col_names$energy_type,
  last_stage = col_names$last_stage,
  year = col_names$year,
  ledger_side = col_names$ledger_side,
  flow_aggregation_point = col_names$flow_aggregation_point,
  flow = col_names$flow,
  sep = "_",
  fap_flow = paste0(flow_aggregation_point, sep, flow),
  product = col_names$product,
  country_order = IEATools::countries,
  method_order = IEATools::methods,
  energy_type_order = IEATools::energy_types,
  last_stage_order = IEATools::last_stages,
  ledger_side_iea_order = IEATools::ledger_sides,
  fap_flow_iea_order = IEATools::fap_flows,
  product_iea_order = IEATools::products,
  .clean_flow = ".clean_flow",
  .clean_product = ".clean_product"
)
}
\arguments{
\item{.iea_df}{the IEA data frame to be sorted}

\item{col_names}{a list of column names in IEA data frames. Default is \code{IEATools::iea_cols}.}

\item{country}{the name of the country column in \code{.tidy_iea_df}. Default is "Country".}

\item{method}{the name of the method column in \code{.tidy_iea_df}. Default is "Method".}

\item{energy_type}{the name of the energy type column in \code{.tidy_iea_df}. Default is "Energy.type".}

\item{last_stage}{the name of the last stage column in \code{.tidy_iea_df}. Default is "Last.stage".}

\item{year}{the name of the year column in \code{.tidy_iea_df}. Default is "Year".}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is "Ledger.side".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "Flow.aggregation.point".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "Flow".}

\item{sep}{a separator between the flow aggregation point column and the flow column. Used when uniting those two columns internally. Default is "_".}

\item{fap_flow}{the name of the united flow aggregation point and flow column to be created internally in \code{.tidy_iea_df}. Default is "Flow.aggregation.point_Flow".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "Product".}

\item{country_order}{the order in which to sort the \code{country} column of \code{.tidy_iea_df}. Default is \link{countries}.}

\item{method_order}{the order in which to sort the \code{method} column of \code{.tidy_iea_df}. Default is \link{methods}.}

\item{energy_type_order}{the order in which to sort the \code{energy_type} column of \code{.tidy_iea_df}. Default is \link{energy_types}.}

\item{last_stage_order}{the order in which to sort the \code{last_stage} column of \code{.tidy_iea_df}. Default is \link{last_stages}.}

\item{ledger_side_iea_order}{the order in which to sort the \code{ledger_side} column of \code{.tidy_iea_df}. Default is \link{ledger_sides}.}

\item{fap_flow_iea_order}{the order in which to sort the united \code{flow_aggregation_point} and \code{flow} columns of \code{.tidy_iea_df}. Default is \link{fap_flows}.}

\item{product_iea_order}{the order in which to sort the \code{product} column of \code{.tidy_iea_df}. Default is \link{products}.}

\item{.clean_flow}{the name of an internally-generated column in \code{.iea_df} that stores
a de-specified version of the \code{flow} column. Default is ".clean_flow".}

\item{.clean_product}{the name of an internally-generated column in \code{.iea_df} that stores
a de-specified version of the \code{product} column. Default is ".clean_product".}
}
\value{
a version of \code{.tidy_iea_df} sorted in IEA order
}
\description{
The IEA data frame to be sorted can be either
(a) tidy (long) where each observation is on its own row and
there is a \code{year} column present or
(b) wide where year columns are spread to the right.
}
\details{
Sorting is accomplished (by default) using
the values of the arguments \code{countries}, \code{methods},
\code{energy_type}, \code{last_stage}, \code{year} (if present),
\code{ledger_side}, \code{fap_flows}, and
\code{product}
(in that order of precedence).

Years are sorted if the \code{year} column is present (a tidy data frame).
If years are not present, they are assumed to be spread to the right
to create a wide data frame.
Wide data frames are sorted in the same order.
}
\examples{
library(dplyr)
tidy <- load_tidy_iea_df()
# See first and last rows
head(tidy)
tail(tidy)
# Move the first row to the bottom to put everything out of order
unsorted <- tidy[-1, ] \%>\% 
  bind_rows(tidy[1, ])
head(unsorted)
tail(unsorted)
# Now sort it
sorted <- sort_iea_df(unsorted)
head(sorted)
tail(sorted)
}
