% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bunkers.R
\name{specify_bunkers}
\alias{specify_bunkers}
\title{Change bunker flow names to reflect their purpose}
\usage{
specify_bunkers(
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  imb = IEATools::tpes_flows$international_marine_bunkers,
  iab = IEATools::tpes_flows$international_aviation_bunkers,
  etwmb = IEATools::tpes_flows$exports_to_world_marine_bunkers,
  etwab = IEATools::tpes_flows$exports_to_world_aviation_bunkers,
  imports = IEATools::tpes_flows$imports,
  wmb_country = "WMB",
  wab_country = "WAB"
)
}
\arguments{
\item{.tidy_iea_df}{An IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}}

\item{country}{See \code{IEATools::iea_cols$country}.}

\item{flow}{See \code{IEATools::iea_cols$flow}.}

\item{product}{See \code{IEATools::iea_cols$product}.}

\item{imb}{\verb{See IEATools::tpes_flows$international_marine_bunkers}.}

\item{iab}{\verb{See IEATools::tpes_flows$international_aviation_bunkers}.}

\item{etwmb}{\verb{See IEATools::tpes_flows$exports_to_world_marine_bunkers}.}

\item{etwab}{\verb{See IEATools::tpes_flows$exports_to_world_aviation_bunkers}.}

\item{imports}{See \code{IEATools::tpes_flows$imports}.}

\item{wmb_country}{The 3-letter code for the world marine bunkers country. Default is "WMB".}

\item{wab_country}{The 3-letter code for the world aviation bunkers country. Default is "WAB".}
}
\value{
A \code{.tidy_iea_df} with bunker flows specified.
}
\description{
The IEA extended energy balance data includes
flows "International marine bunkers" and "International aviation bunkers"
which are insufficiently specific for our purposes.
These flows are actually exports to and imports at the regions
"World marine bunkers" and "World aviation bunkers"
(which we treat as a countries).
This function changes the names of the flows to
"Exports to World marine bunkers" and "Exports to World aviation bunkers".
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_tp_eiou() \%>\% 
 specify_bunkers()   
}
