% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{specify_distribution_losses}
\alias{specify_distribution_losses}
\title{Specifies distribution industries}
\usage{
specify_distribution_losses(
  .tidy_iea_df,
  specify_distribution_industries = FALSE,
  supplying_industry_notation = RCLabels::from_notation,
  distribution_industry_notation = RCLabels::of_notation,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  losses = IEATools::tfc_compare_flows$losses,
  distribution_industry = IEATools::distribution_industry,
  supply = IEATools::ledger_sides$supply,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy__iea_df} for which an electricity grid industry should be added.}

\item{specify_distribution_industries}{A boolean stating whether distribution industries should be added or not.
Default is FALSE.}

\item{supplying_industry_notation}{Notation to use to specify the supplying industries.
Default is \code{RCLabels::from_notation}.}

\item{distribution_industry_notation}{Notation to use to specify the distribution industries.
Default is \code{RCLabels::of_notation}.}

\item{flow_aggregation_point, flow, e_dot, product, method, ledger_side, last_stage, energy_type, country, year, unit}{See \code{IEATools::iea_cols}.}

\item{losses}{The name of the "Losses" flows in the input data frame.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{distribution_industry}{The name of the distribution industries to be added.
Default is \code{IEATools::distribution_industry}.}

\item{supply}{The name of the supply ledger side.
Default is \code{IEATools::ledger_sides$supply}.}

\item{transformation_processes}{The name of transformation processes in the flow aggregation point column.
Default is \code{IEATools::tfc_compare_flows$transformation_processes}.}

\item{negzeropos}{The name of a temporary column added to the data frame.
Default is ".negzeropos".}
}
\value{
The \code{.tidy__iea_df} to which distribution industries have been added.
}
\description{
For each product where losses are reported, creates a distribution industry which
is specified by product (e.g., "Distribution [of Heat]"),
and converts a product from a specific origin (e.g., "Heat [from Oil refineries]")
into the final demand product (e.g., "Heat").
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_distribution_losses()
}
