% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiou.R
\name{specify_eiou}
\alias{specify_eiou}
\title{Specify destinations for inter-industry energy industry own use flows}
\usage{
specify_eiou(.tidy_iea_df,
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use", flow = "Flow",
  own_use_elect_chp_heat = "Own use in electricity, CHP and heat plants",
  pumped_storage = "Pumped storage plants",
  main_act_producer_elect = "Main activity producer electricity plants",
  product = "Product", electricity = "Electricity")
}
\arguments{
\item{.tidy_iea_df}{}

\item{flow}{}

\item{own_use_elect_chp_heat}{}

\item{main_act_producer_elect}{}

\item{product}{}

\item{electricity}{}
}
\description{
Specify destinations for inter-industry energy industry own use flows
}
