% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiou.R
\name{specify_eiou_tp}
\alias{specify_eiou_tp}
\title{Specify destinations for energy industry own use flows into transformation processes}
\usage{
specify_eiou_tp(.tidy_iea_df,
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use", flow = "Flow",
  own_use_elect_chp_heat = "Own use in electricity, CHP and heat plants",
  pumped_storage = "Pumped storage plants",
  nuclear_industry = "Nuclear industry",
  main_act_producer_elect = "Main activity producer electricity plants",
  non_spec_energy = "Non-specified (energy)",
  nonspecenergy_reclassify = "Oil and gas extraction")
}
\arguments{
\item{.tidy_iea_df}{an IEA data frame whose columns have been renamed by \code{\link[=rename_iea_df_cols]{rename_iea_df_cols()}}}

\item{flow_aggregation_point}{}

\item{eiou}{}

\item{flow}{}

\item{own_use_elect_chp_heat}{}

\item{pumped_storage}{}

\item{nuclear_industry}{}

\item{main_act_producer_elect}{}

\item{non_spec_energy}{}

\item{nonspecenergy_reclassify}{}
}
\value{
a modified version of \code{.tidy_iea_df}
}
\description{
The extended energy balance data from the IEA includes
Energy industry own use (EIOU) for many transformation processes.
Unfortunately, the EIOU flows into industries that aren't included in transformation processes.
For example, \code{Electricity} is consumed by
\code{Own use in electricity, CHP and heat plants},
which is not transformation process.
We have to make some decisions to ensure that
EIOU is routed to actual transformation processes.
See details for a list of actual changes made to the \code{.tidy_iea_df} data frame.
}
\details{
The following changes are made to the \code{.tidy_iea_df} data frame:
\enumerate{
\item EIOU classified as \code{own_use_elect_chp_heat} is sent to \code{main_act_producer_elect}.
\item EIOU classified as \code{pumped_storage} is sent to \code{main_act_producer_elect}.
\item EIOU classified as \code{nuclear_industry} is sent to \code{main_act_producer_elect}.
\item EIOU classified as \code{non_spec_energy} is sent to \code{nonspecenergy_reclassify}.
}
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_eiou_tp() \%>\% 
  filter(Flow.aggregation.point == "Energy industry own use" & 
           Flow == "Main activity producer electricity plants")
}
