% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{specify_electricity_grid}
\alias{specify_electricity_grid}
\title{Specifies electricity grid}
\usage{
specify_electricity_grid(
  .tidy_iea_df,
  specify_electricity_grid = FALSE,
  supplying_industry_notation = RCLabels::of_notation,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  e_dot = IEATools::iea_cols$e_dot,
  losses = IEATools::tfc_compare_flows$losses,
  grid_industry = IEATools::grid_industries$electricity_grid,
  supply = IEATools::ledger_sides$supply,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  negzeropos = ".negzeropos"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy__iea_df} for which an electricity grid industry should be added.}

\item{supplying_industry_notation}{Notation to use to specify the electricity supplying industry.
Default is \code{RCLabels::of_notation}.}

\item{flow_aggregation_point, flow, e_dot, product, method, ledger_side, last_stage, energy_type, country, year, unit}{See \code{IEATools::iea_cols}.}

\item{losses}{The name of the "Losses" flows in the input data frame.
Default is \code{IEATools::tfc_compare_flows$losses}.}

\item{grid_industry}{The name of the electricity grid industry to be added.
Default is \code{IEATools::grid_industries$electricity_grid}.}

\item{supply}{The name of the supply ledger side.
Default is \code{IEATools::ledger_sides$supply}.}

\item{transformation_processes}{The name of transformation processes in the flow aggregation point column.
Default is \code{IEATools::tfc_compare_flows$transformation_processes}.}

\item{negzeropos}{The name of a temporary column added to the data frame.
Default is ".negzeropos".}

\item{specify_grid}{A boolean stating whether an electricity grid industry should be created or not.
Default is FALSE.}
}
\value{
The \code{.tidy__iea_df} to which an electricity grid industry has been added.
}
\description{
Adds an electricity grid industry that takes as input all electricity produced by any industry,
which is now specified by producing industry (e.g., "Electricity [from Wind power plants]"),
and converts it into Electricity.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_electricity_grid()
}
