% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_interface_industries}
\alias{specify_interface_industries}
\title{Specify interface industries}
\usage{
specify_interface_industries(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  int_industries = c(IEATools::interface_industries, IEATools::tpes_flows$resources,
    manufacture = "Manufacture"),
  product = IEATools::iea_cols$product,
  flow_notation = RCLabels::of_notation,
  product_notation = RCLabels::bracket_notation
)
}
\arguments{
\item{.tidy_iea_df}{A tidy data frame containing IEA extended energy balance data.}

\item{flow}{The name of the flow column in \code{.tidy_iea_df}.  Default is "\code{Flow}".}

\item{int_industries}{A string vector of industries involved in exchanges with other countries,
bunkers, or stock changes.
Default is \code{c(IEATools::interface_industries, IEATools::tpes_flows$resources, manufacture = "Manufacture")}.}

\item{product}{The name of the product column in \code{.tidy_iea_df}.
Default is "\code{Product}".}

\item{flow_notation}{The notation for the \code{flow} column.
Default is \code{RCLabels::of_notation}.}

\item{product_notation}{The notation for the \code{product} column.
Default is \code{RCLabels::bracket_notation}, meaning that
any type of "X [from Resources]" or "X [to Y]"
will be parsed correctly for its prefix.}
}
\value{
A modified version of \code{.tidy_iea_df} with specified interface industries.
}
\description{
An interface industry is one that moves energy carriers into or out of a country.
When \code{flow} is any of the interface industries, we need to be more specific.
If we don't separate these \code{flow}s by \code{product}, we run into trouble with
upstream swims (e.g., all \code{product}s are produced even if only one is needed) and
embodied energy calculations (many types of energy are embodied, even if only one should be).
This function adds a suffix "[of Product]" to each of these interface industries.
}
\details{
Note that "Production" also needs to be specified,
but that is accomplished in the \code{specify_primary_production()} and
\code{specify_production_to_resources()} functions.

Resource flows and manufacture flows are included by default,
because they need a place where they are specified, too.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_interface_industries()
}
