% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{specify_non_energy_use}
\alias{specify_non_energy_use}
\title{Specify Non-energy use when possible}
\usage{
specify_non_energy_use(
  .iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  non_energy_flows_industry_transformation_energy =
    IEATools::non_energy_flows$non_energy_use_industry_transformation_energy,
  memo = IEATools::memo_aggregation_flow_prefixes$memo,
  memo_non_energy_flows_industry =
    IEATools::memo_non_energy_flows$memo_non_energy_use_in_industry,
  memo_non_energy_use_in = "Memo: Non-energy use in ",
  total = IEATools::memo_aggregation_product_prefixes$total,
  .values = IEATools::template_cols$.values,
  .values_summarised = paste0(.values, "_summarised"),
  .diff = ".diff",
  tol = 1e-06
)
}
\arguments{
\item{.iea_df}{A data frame of IEA data,
created by \code{rename_iea_df_cols()}.}

\item{country, year, flow, product}{See \code{IEATools::iea_cols}.}

\item{non_energy_flows_industry_transformation_energy}{See \code{IEATools::non_energy_flows}.}

\item{memo}{A string prefix for memo flows.
Default is \code{IEATools::memo_aggregation_flow_prefixes$memo}.}

\item{memo_non_energy_flows_industry}{See \code{IEATools::memo_non_energy_flows}.}

\item{memo_non_energy_use_in}{A prefix for specific Non-energy use flows.
Default is "Memo: Non-energy use in ".}

\item{total}{See \code{IEATools::memo_aggregation_product_prefixes}.}

\item{.values}{An internal column name.
Default is \code{IEATools::template_cols$.values}.}

\item{.values_summarised}{An internal column name.
Default is \code{paste0(.values, "_summarised")}.}

\item{.diff}{An internal column name.
Default is ".diff".}

\item{tol}{The tolerance for differences from \code{0}.
Default is \code{1e-6}.}
}
\value{
\code{.iea_df} with specified \verb{Non-energy use}s, where possible.
}
\description{
For some countries and years, Non-energy use details are supplied
by "Memo:" fields.
We use those details, but only if they are balanced.
This function assess the energy balance of the
"Memo: Non-energy use in xxxxx"
fields and replaces
"Non-energy use in industry"
with the details in the "Memo:" entries when possible.
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  clean_iea_whitespace() \%>\%
  augment_iea_df() \%>\%
  specify_non_energy_use()
}
