% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specify_notation}
\alias{specify_notation}
\title{Specification notation}
\format{
A character vector with 15 entries
}
\usage{
specify_notation
}
\description{
A character vector containing the specification notation used in the IEATools package.
Notation for delimiting specification strings is included.
}
\examples{
specify_notation
}
\keyword{datasets}
