% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_primary_production}
\alias{specify_primary_production}
\title{Specify primary production industries}
\usage{
specify_primary_production(
  .tidy_iea_df,
  eiou_destinations = c("Coal mines", "Oil and gas extraction"),
  liquefaction_regas = "Liquefaction (LNG) / regasification plants",
  liquefaction_regas_reassign = "Oil and gas extraction",
  production_products = list(IEATools::primary_coal_products,
    c(IEATools::primary_oil_products, "Natural gas")),
  production_products_short_names = c("Coal", "Oil and natural gas"),
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  transformation_processes = "Transformation processes",
  flow = "Flow",
  resources = "Resources",
  production = "Production",
  e_dot = "E.dot",
  product = "Product",
  notation = IEATools::specify_notation,
  .resources_open = notation$resources_open,
  .resources_close = notation$resources_close
)
}
\arguments{
\item{.tidy_iea_df}{an IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}}

\item{eiou_destinations}{a vector of destinations for EIOU for primary production of coal and coal products and oil and natural gas.
Default is \code{c("Coal mines", "Oil and gas extraction")}.}

\item{liquefaction_regas}{a string identifying liquefaction and regasification plants. Default is "Liquefaction (LNG) / regasification plants".}

\item{liquefaction_regas_reassign}{a string identifying the industry to which EIOU into \code{liquefaction_regas} will be reassigned.
Default is "\verb{Oil and gas extraction}".}

\item{production_products}{a list of products for which we want to specify primary industries.
Default is \code{list(primary_coal_products, c(primary_oil_products, "Natural gas"))}.}

\item{production_products_short_names}{a vector of short names for primary industries.
Default is \code{c("Coal", "Oil and natural gas")}.}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}.
Default is "\code{Flow.aggregation.point}".}

\item{eiou}{a string identifying energy industry own use in the \code{Flow.aggregation.point} column of \code{.tidy_iea_df}.
Default is "\verb{Energy industry own use}".}

\item{transformation_processes}{a string identifying transformation processes in the flow column of \code{.tidy_iea_df}.
Default is "\verb{Transformation processes}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}.  Default is "\code{Flow}".}

\item{resources}{a string identifying resource industries to be added to \code{.tidy_iea_df}.
Default is "\code{Resources}".}

\item{production}{a string identifying production in the flow column. Default is "\code{Production}".}

\item{e_dot}{the name of the energy column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}.  Default is "\code{Product}".}

\item{notation}{a list of specification notations. Default is \code{IEATools::specify_notation}.}

\item{.resources_open}{a string that identifies the start of the specification portion of flows. Default is \code{specify_notation$resources_open}.}

\item{.resources_close}{a string that identifies the end of specification portion of  flows. Default is \code{specify_notation$resources_close}.}
}
\value{
\code{.tidy_iea_df} with adjusted production information for primary energy
for both coal and coal products and oil and gas extraction
}
\description{
The IEA extended energy balances include
some \code{Flow}s that identify \verb{Energy industry own use} (EIOU),
the consumption of energy by energy-producing industries.
But some primary production industries that receive EIOU do not produce anything.
For example, \verb{Coal mines} receive electricity
but there are no \verb{Coal mines} that produce coal.
Rather, the generic \code{Production} industry produces coal.
This function solves that problem by
replacing the generic \code{Production} industry with
specific industries.
}
\details{
By default, the following changes are made to \code{.tidy_iea_df}:
\enumerate{
\item Energy industry own use for \verb{Liquefaction (LNG) / regasification plants} is
reassigned to \verb{Oil and gas extraction}.
\item The \code{Production} industry for \code{coal_and_coal_products}
is replaced by \verb{Coal mines}.
\item A \code{Resources (Coal)} industry is created which becomes the
source of all primary \code{coal_and_coal_products} in the energy conversion chain.
\item Flows from \code{Resources (Coal)} to \verb{Coal mines} are added.
\item The \verb{Coal mines (energy)} EIOU \code{Flow} is replaced by \verb{Coal mines}.
\item The \code{Production} industry for \code{oil_and_oil_products} and \verb{Natural gas}
is replaced by \verb{Oil and gas extraction}.
\item A \verb{Resources (Oil and natural gas)} industry is created which becomes the
source of all primary \code{oil_and_oil_products} and \verb{Natural gas} in the energy conversion chain.
\item Flows from \verb{Resources (Oil and natural gas)} to \verb{Oil and gas extraction} are added.
\item The \verb{Oil and gas extraction (energy)} EIOU \code{Flow} is replaced by \verb{Oil and gas extraction}.
}

Users can specify other changes by adjusting the default argument values.

Be sure to call this function \emph{after} calling \code{augment_iea_df()} or
\code{load_tidy_iea_df()}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  add_psut_matnames() \%>\% 
  filter(Flow \%in\% c("Resources (Coal)", "Coal mines")) \%>\%
  select(-Method, -Last.stage, -Ledger.side, -Unit)
# EIOU by "Liquefaction (LNG) / regasification plants" is reassigned to "Oil and gas extraction"
data.frame(
  Flow.aggregation.point = c("Energy industry own use"),
  Flow = c("Liquefaction (LNG) / regasification plants"), 
  Product = c("Natural gas"),
  E.dot = c(-42), 
  stringsAsFactors = FALSE
) \%>\% 
  specify_primary_production()
}
