% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_production_to_resources}
\alias{specify_production_to_resources}
\title{Convert Production Flows to Resource Flows}
\usage{
specify_production_to_resources(
  .tidy_iea_df,
  flow = "Flow",
  product = "Product",
  production = "Production",
  resources = "Resources",
  .open = notation$specify_open,
  .close = notation$specify_close
)
}
\arguments{
\item{.tidy_iea_df}{an IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}.  Default is "Flow".}

\item{product}{the name of the product column in \code{.tidy_iea_df}.  Default is "Product".}

\item{production}{a string identifying production in the flow column. Default is "Production".}

\item{resources}{a string identifying resource industries to be added to \code{.tidy_iea_df}.
Default is "Resources".}

\item{.open}{a string that identifies the start of the specification portion of flows. Default is \code{notation$specify_open}.}

\item{.close}{a string that identifies the end of specification portion of  flows. Default is \code{notation$specify_close}.}
}
\value{
\code{.tidy_iea_df} with \code{Production} changed to \code{Resources (product)}
}
\description{
The IEA gives resource extraction in rows where the \code{Flow} is "\code{Production}".
This function changes the "\code{Production}" string
to "\code{Resources (product)}",
where \code{product} is the name of the energy carrier for this resource.
}
\details{
This function should be called \emph{after} \code{specify_primary_production()},
which adjusts for energy industry own use
of some primary energy producing industries.
If this function is called first,
EIOU will not be accounted correctly.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_production_to_resources()
}
