% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{specify_renewable_plants}
\alias{specify_renewable_plants}
\title{Specifies renewable electricity and heat}
\usage{
specify_renewable_plants(
  .tidy_iea_df,
  specify_renewable_plants = FALSE,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  product = IEATools::iea_cols$product,
  method = IEATools::iea_cols$method,
  ledger_side = IEATools::iea_cols$ledger_side,
  last_stage = IEATools::iea_cols$last_stage,
  energy_type = IEATools::iea_cols$energy_type,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  unit = IEATools::iea_cols$unit,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  main_act_producer_elect = IEATools::main_act_plants$main_act_prod_elect_plants,
  main_act_producer_chp = IEATools::main_act_plants$main_act_prod_chp_plants,
  main_act_producer_heat = IEATools::main_act_plants$main_act_prod_heat_plants,
  autoproducer_elect = IEATools::main_act_plants$autoprod_elect_plants,
  autoproducer_chp = IEATools::transformation_processes$autoproducer_CHP_plants,
  autoproducer_heat = IEATools::main_act_plants$autoprod_heat_plants,
  geothermal = IEATools::renewable_products$geothermal,
  hydro = IEATools::renewable_products$hydro,
  solar_pv = IEATools::renewable_products$solar_photovoltaics,
  solar_th = IEATools::renewable_products$solar_thermal,
  oceanic = IEATools::renewable_products$tide_wave_and_ocean,
  wind = IEATools::renewable_products$wind,
  electricity = IEATools::electricity_products$electricity,
  heat = "Heat",
  ratio_solar_th_elec = 0.33,
  ratio_solar_th_heat = 1,
  ratio_geothermal_elec = 0.1,
  ratio_geothermal_heat = 0.5,
  ratio_other_renewable_elec = 1,
  geothermal_plants = IEATools::renewable_industries$geothermal_plants,
  hydro_plants = IEATools::renewable_industries$hydro_plants,
  solar_pv_plants = IEATools::renewable_industries$solar_pv_plants,
  solar_th_plants = IEATools::renewable_industries$solar_th_plants,
  oceanic_plants = IEATools::renewable_industries$oceanic_plants,
  wind_power_plants = IEATools::renewable_industries$wind_power_plants,
  negzeropos = ".negzeropos",
  ratio_elec_to_heat = ".ratio_elec_to_heat"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} which flows need to be specified.}

\item{specify_renewable_plants}{A boolean indicating whether renewable energy plants should be specified or not.
Default is FALSE.}

\item{flow_aggregation_point, flow, e_dot, product, method, ledger_side, last_stage, energy_type, country, year, unit}{See \code{IEATools::iea_cols}.}

\item{transformation_processes}{A string identifying the transformation processes in the \code{flow_aggregation_point} column in the \code{.tidy_iea_df}.
Default is \code{IEATools::aggregation_flows$transformation_processes}.}

\item{main_act_producer_elect}{A string identifying "Main activity producer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_elect_plants}.}

\item{main_act_producer_chp}{A string identifying "Main activity producer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_chp_plants}.}

\item{main_act_producer_heat}{A string identifying "Main activity producer heat plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$main_act_prod_heat_plants}.}

\item{autoproducer_elect}{A string identifying "Autoproducer electricity plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::main_act_plants$autoprod_elect_plants}.}

\item{autoproducer_chp}{A string identifying "Autoproducer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::transformation_processes$autoproducer_CHP_plants}.}

\item{autoproducer_heat}{A string identifying "Autoproducer CHP plants" in the \code{flow} column of the \code{.tidy_iea_df}.
Default is \code{IEATools::transformation_processes$autoprod_heat_plants}.}

\item{geothermal, hydro, solar_pv, solar_th, oceanic, wind}{Renewable energy product names. See \code{IEATools::renewable_products}.}

\item{electricity}{The name of the electricity product.
Default is \code{IEATools::electricity_products$electricity}.}

\item{heat}{The name of the heat product.
Default is "Heat".}

\item{ratio_solar_th_elec}{The ratio of primary energy to electricity to use for solar thermal.
Default is 0.33 as this is the value assumed in the IEA's energy balances.}

\item{ratio_solar_th_heat}{The ratio of primary energy to heat to use for solar thermal.
Default is 1 as this is the value assumed in the IEA's energy balances.}

\item{ratio_geothermal_elec}{The ratio of primary energy to electricity to use for geothermal.
Default is 0.1 as this is the value assumed in the IEA's energy balances.}

\item{ratio_geothermal_heat}{The ratio of primary energy to heat to use for geothermal.
Default is 0.5 as this is the value assumed in the IEA's energy balances.}

\item{ratio_other_renewable_elec}{The ratio of primary energy to electricity to use for hydropower, solar photovoltaic, oceanic and wind power.
Default is 1 as this is the value assumed in the IEA's energy balances.}

\item{geothermal_plants, hydro_plants, solar_pv_plants, solar_th_plants, oceanic_plants, wind_power_plants}{Names of renewable industries added. See \code{IEATools::renewable_industries}.}

\item{negzeropos}{The name of a temporary column added to the data frame.
Default is ".negzeropos".}

\item{ratio_elec_to_heat}{A temporary column added to the data frame.
Default is ".ratio_elec_to_heat".}
}
\value{
Returns a .tidy_iea_df with renewable electricity and heat from geothermal, hydropower, solar thermal, solar photovoltaic, wind power, and oceanic power specified.
}
\description{
This function specifies hydropower, geothermal, solar photovoltaic, solar thermal, oceanic, and wind power industries.
}
\details{
The primary energy use of hydro, geothermal, solar photovoltaic, solar thermal, oceanic, and wind power energy by main activity and autoproducer plants are used
to create new renewable industries that produce electricity and heat (heat only in the case of geothermal and solar thermal). The physical content method is used
to derive the electricity produced by renewable industries, except in the case of geothermal and solar thermal, for which the IEA uses other factors in its balances.
In the case of CHP plants (which can be relevant for geothermal and solar thermal), the output of the new renewable industry follows the same heat vs electricity
breakdown as the main industry from which it is derived.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_renewable_plants()
}
