% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_tp_eiou}
\alias{specify_tp_eiou}
\title{Specify destinations for energy industry own use flows into transformation processes}
\usage{
specify_tp_eiou(
  .tidy_iea_df,
  split_own_use_elect_chp_heat_using_shares_of = c("input", "output"),
  route_non_specified_eiou = TRUE,
  route_non_specified_tp = TRUE,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  flow = IEATools::iea_cols$flow,
  own_use_elect_chp_heat = IEATools::eiou_flows$own_use_elect_chp_heat_plants,
  pumped_storage = IEATools::eiou_flows$pumped_storage_plants,
  nuclear_industry = IEATools::eiou_flows$nuclear_industry,
  e_dot = IEATools::iea_cols$e_dot,
  negzeropos = ".negzeropos",
  main_act_producer_elect =
    IEATools::eiou_flows$main_activity_producer_electricity_plants
)
}
\arguments{
\item{.tidy_iea_df}{An IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}.}

\item{split_own_use_elect_chp_heat_using_shares_of}{Indicates whether the input or outputs to
Main activity producer plants should be use for
splitting the Own use in electricity, chp and heat plants
EIOU flow. Default is "input".}

\item{route_non_specified_eiou}{Boolean stating whether non-specified EIOU flows should be routed to existing industries.
Default is TRUE.}

\item{route_non_specified_tp}{Boolean stating whether non-specified transformation processes flows should be routed to existing industries.
Default is TRUE.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in \code{.tidy_iea_df}. Default is FlowAggregationPoint.}

\item{eiou}{A string identifying energy industry own use in the flow aggregation point column. Default is "Energy industry own use".}

\item{transformation_processes}{A string identifying transformation processes in the flow aggregation point column. Default is "Transformation processes".}

\item{flow}{The name of the flow column in \code{.tidy_iea_df}. Default is "Flow".}

\item{own_use_elect_chp_heat}{A string identifying own use in electricity, CHP and heat plants in the flow column. Default is "Own use in electricity, CHP and heat plants".}

\item{pumped_storage}{A string identifying pumped storage plants in the flow column. Default is "Pumped storage plants".}

\item{nuclear_industry}{A string identifying nuclear plants in the flow column. Default is "Nuclear industry".}

\item{e_dot}{The name of the energy flow column in \code{.tidy_iea_df}. Default is "Edot".}

\item{negzeropos}{The name of a temporary column created in \code{.tidy_iea_df}. Default is ".negzeropos".}

\item{main_act_producer_elect}{A string identifying main activity producer electricity plants. Default is "Main activity producer electricity plants".}
}
\value{
A modified version of \code{.tidy_iea_df}.
}
\description{
The extended energy balance data from the IEA includes
"Energy industry own use" (EIOU) for many transformation processes.
Unfortunately, in some cases
the EIOU flows into industries that aren't included in transformation processes.
For example, "Electricity" is consumed by
"Own use in electricity, CHP and heat plants",
which is not a transformation process.
We have to make some decisions to ensure that
EIOU is routed to actual transformation processes.
See details for a list of changes made to the \code{.tidy_iea_df} data frame.
}
\details{
The following changes are made to the \code{.tidy_iea_df} data frame:
\enumerate{
\item EIOU classified as \code{own_use_elect_chp_heat} is sent to \code{main_act_producer_elect}.
\item EIOU classified as \code{pumped_storage} is sent to \code{main_act_producer_elect}.
\item EIOU classified as \code{nuclear_industry} is sent to \code{main_act_producer_elect}.
\item EIOU classified as \code{non_spec_energy} is sent to \code{nonspecenergy_reclassify}.
}

After the changes are made, reassigned EIOU may double-up pre-existing EIOU.
For example, a country may already have "Electricity" EIOU flowing into "Main activity producer electricity plants".
It may also have EIOU flowing into "Nuclear industry".
When we switch the EIOU flow into "Nuclear industry" into "Main activity producer electricity plants",
we now have two rows of electricity EIOU into "Main activity producer electricity plants".
To avoid double rows, all like rows are summed before returning.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_tp_eiou() \%>\% 
  filter(FlowAggregationPoint == "Energy industry own use" & 
           Flow == "Main activity producer electricity plants")
}
