% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_tp_eiou.R
\name{split_oil_gas_extraction_eiou}
\alias{split_oil_gas_extraction_eiou}
\title{Separates EIOU flows of oil and gas extraction}
\usage{
split_oil_gas_extraction_eiou(
  .tidy_iea_df,
  eiou = IEATools::aggregation_flows$energy_industry_own_use,
  country = IEATools::iea_cols$country,
  energy_type = IEATools::iea_cols$energy_type,
  method = IEATools::iea_cols$method,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  year = IEATools::iea_cols$year,
  flow = IEATools::iea_cols$flow,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  e_dot = IEATools::iea_cols$e_dot,
  oil_gas_extraction = IEATools::eiou_flows$oil_and_gas_extraction,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  oil_extraction = IEATools::industry_flows$oil_extraction,
  gas_extraction = IEATools::industry_flows$natural_gas_extraction,
  share = "Share"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} on which oil and gas extraction EIOU flows need to be separated.}

\item{eiou}{The name of the Energy industry own use flow aggregation point.
Default is \code{IEATools::aggregation_flows$energy_industry_own_use}.}

\item{country, energy_type, method, last_stage, ledger_side, year, flow, flow_aggregation_point, e_dot}{See \code{IEATools::iea_cols}.}

\item{oil_gas_extraction}{The name of the Oil and gas extraction EIOU flow.
Default is \code{IEATools::eiou_flows$oil_and_gas_extraction}.}

\item{transformation_processes}{The name of the flow aggregation point referring to transformation processes.
Default is \code{IEATools::aggregation_flows$transformation_processes}.}

\item{oil_extraction}{The name of the Oil extraction industry.
Default is \code{IEATools::industry_flows$oil_extraction}.}

\item{gas_extraction}{The name of the Natural gas extraction industry.
Default is \code{IEATools::industry_flows$natural_gas_extraction}.}

\item{share}{The name of a temporary column that is added to the data frame.
Default is "Share".}
}
\value{

}
\description{
This function separates the EIOU flows of the Oil and gas extraction industry into EIOU flows
for the Oil extraction industry and EIOU flows for the Natural gas extraction industry.
It uses the shares of production of each of these two industries to separate EIOU flows.
As such, each EIOU flow will be equivalent in the two industries, once reported to each industry's output.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  split_oil_gas_extraction_eiou()
}
