% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{spread_fu_allocation_table}
\alias{spread_fu_allocation_table}
\title{Spread an FU Allocation table}
\usage{
spread_fu_allocation_table(
  .tidy_fu_allocation_table,
  year = IEATools::iea_cols$year,
  max_vals = IEATools::template_cols$maximum_values,
  val_col = "values"
)
}
\arguments{
\item{.tidy_fu_allocation_table}{The FU Allocation table to be widened.}

\item{year}{See \code{IEATools::iea_cols}.}

\item{max_vals}{See \code{IEATools::template_cols}.}

\item{val_col}{The name of the values column to be spread. Default is "values".}
}
\value{
a wide version of \code{.tidy_fu_allocation_table}.
}
\description{
FU (final-to-useful) allocation tables are not tidy data structures,
but they can be tidied by calling \code{tidy_fu_allocation_table()}.
This function reverses that process to widen a tidy FU Allocation table
by spreading the years (and the \code{max_vals} column) into new columns with names taken from the \code{year} column.
}
\examples{
load_fu_allocation_data() \%>\% 
  tidy_fu_allocation_table() \%>\% 
  spread_fu_allocation_table()
}
