% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{starts_with_any_of}
\alias{starts_with_any_of}
\title{Tell if a string starts with any of a vector of strings}
\usage{
starts_with_any_of(x, target)
}
\arguments{
\item{x}{a string (or vector or list of strings)}

\item{target}{a vector or list of strings}
}
\value{
\code{TRUE} if \code{x} starts with any of the strings in \code{target},
\code{FALSE} otherwise.
If \code{x} is a vector or list of strings, the return value is the same length as \code{x}
and contains the result of applying the test to each item in \code{x}.
}
\description{
This function returns \code{TRUE} if \code{x}
starts with any of the strings in \code{target} and \code{FALSE} otherwise.
}
\details{
This function is vectorized. If \code{x} is a vector or list of strings,
the return value has the same length as \code{x} and contains the result
of applying the test (does \code{x} start with any of \code{target})
for each item in \code{x}.
}
\examples{
starts_with_any_of(x = "prefix - suffix", target = c("a", "b", "prefix"))
starts_with_any_of(x = "prefix - suffix", target = c("a", "b", "c"))
starts_with_any_of(x = "prefix - suffix", target = "suffix")
starts_with_any_of(x = c("Production - Crude", "Production - NG",
                         "Exports - Oil", "Exports - Crude"),
                   target = c("Production", "Imports"))
}
