% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sut_meta_cols}
\alias{sut_meta_cols}
\title{PSUT matrix meta information column names}
\format{
A string list with 5 entries.
\describe{
\item{country}{The name of a column containing countries.}
\item{method}{The name of a column containing methods for calculating primary energy equivalent of renewable electricity. See \code{IEATools::methods}.}
\item{energy_type}{The name of a column containing energy types. See \code{IEATools::energy_types}.}
\item{last_stage}{The name of a column containing last stages of energy conversion chain analysis. See \code{IEATools::last_stages}.}
\item{year}{The name of a column containing years in wide data frames.}
}
}
\usage{
sut_meta_cols
}
\description{
After forming PSUT matrices, meta information is provided in columns.
This list provides the typical names for the meta information columns
throughout the \code{IEATools} package.
}
\examples{
sut_meta_cols
}
\keyword{datasets}
