% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{switch-notation-byname}
\alias{switch-notation-byname}
\alias{arrow_to_paren_byname}
\alias{paren_to_arrow_byname}
\title{Change specification notation on matrix row or column names}
\usage{
arrow_to_paren_byname(m, margin = c(1, 2))

paren_to_arrow_byname(m, margin = c(1, 2))
}
\arguments{
\item{m}{a single matrix or a list of matrices.}

\item{margin}{the margin over which the notation switch should be made:
\code{1} for rows, \code{2} for columns, or \code{c(1, 2)} (the default) for both rows and columns.}
}
\value{
a version of \code{m} with switched notation
}
\description{
Switch notation style
in row or column names of matrices or a list of matrices.
Notation style is either arrow ("a -> b") or parenthetical ("b [a]")
}
\details{
The suffix "_byname" indicates that this function behaves like functions in the \code{byname} package.
Specifically, it can work with a single matrix or a list of matrices supplied to the \code{m} argument.
\itemize{
\item \code{arrow_to_paren_byname()} switches from arrow to parenthetical specification notation for a matrix or list of matrices.
\item \code{paren_to_arrow_byname()} switches from parenthetical to arrow specification notation for a matrix or list of matrices.
}
}
\examples{

}
\seealso{
\code{arrow_to_paren()}, \code{paren_to_arrow()}, and \code{switch_notation()}
}
