% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ternary_xor}
\alias{ternary_xor}
\title{Perform a ternary xor calculation}
\usage{
ternary_xor(a, b, c)
}
\arguments{
\item{a}{the first argument (a logical)}

\item{b}{the second argument (a logical)}

\item{c}{the third argument (a logical)}
}
\value{
\code{TRUE} if only one argument is true, \code{FALSE} otherwise.
}
\description{
Binary xor returns TRUE iff only one of two arguments is \code{TRUE}, but not both.
Ternary xor returns TRUE iff only one of three arguments is \code{TRUE}, but not both.
}
\details{
Internally, this function converts all arguments to numeric and sums.
Only when the sum is \code{1}, the function returns \code{TRUE}.
}
\examples{
ternary_xor(TRUE, FALSE, FALSE)
ternary_xor(FALSE, TRUE, FALSE)
ternary_xor(FALSE, FALSE, TRUE)
ternary_xor(TRUE, TRUE, FALSE)
ternary_xor(FALSE, TRUE, TRUE)
ternary_xor(TRUE, FALSE, TRUE)
ternary_xor(TRUE, TRUE, TRUE)
ternary_xor(FALSE, FALSE, FALSE)
\dontrun{
  ternary_xor(1, 0, 0) # error: all arguments need to be logical
  ternary_xor(1, FALSE, FALSE) # error: all arguments need to be logical
}
}
