% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tac.R
\name{test_tac}
\alias{test_tac}
\title{Compute (partial) autocorrelation functions and test for significance}
\usage{
test_tac(x)
}
\arguments{
\item{x}{numeric; an evenly spaced time vector which should be tested for temporal
autocorrelation. Temporal gaps should be included as NAs.}
}
\value{
The function returns a list with 2 components:
\describe{
  \item{\code{tac}}{TRUE if any lag in the (partial) autocorrelation functions
    is significantly correlated (if correlation value > 0.4), else FALSE.}
  \item{\code{max_lag}}{The maximal lag that is correlated.}
 }
}
\description{
Compute (partial) autocorrelation functions and test for significance
}
\examples{
test_tac(x = 1:20)
test_tac(x = rnorm(20))

}
