% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tidy_eta_fu_table}
\alias{tidy_eta_fu_table}
\title{Tidy a final-to-useful efficiency table}
\usage{
tidy_eta_fu_table(
  .eta_fu_table,
  year = IEATools::iea_cols$year,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  quantity = IEATools::template_cols$quantity,
  maximum_values = IEATools::template_cols$maximum_values,
  .values = IEATools::template_cols$.values
)
}
\arguments{
\item{.eta_fu_table}{The final-to-useful efficiency table to be tidied.}
}
\value{
A tidy version of \code{.eta_fu_table}.
}
\description{
Analysts fill final-to-useful (FU) efficiency tables in a human-readable format
provided by \code{eta_fu_template()} and \code{write_eta_fu_template()}.
The templates are not tidy.
However, most code uses in \code{IEATools} requires tidy data frames.
This function converts an FU efficiency table with years in columns
to a tidy data frame with years in a \code{year} column and \code{eta_fu} and \code{phi_u} values in a \code{.values} column.
Identifiers for the variables are in the \code{quantity} column.
}
\details{
If \code{.eta_fu_table} is already tidy, it is returned unmodified.
}
\examples{
load_eta_fu_data() \%>\% 
  tidy_eta_fu_table()
}
