% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{tidy_fu_allocation_table}
\alias{tidy_fu_allocation_table}
\title{Tidy an FU Allocation table}
\usage{
tidy_fu_allocation_table(
  .fu_allocation_table,
  year = IEATools::iea_cols$year,
  max_vals = IEATools::template_cols$maximum_values,
  values = ".values"
)
}
\arguments{
\item{.fu_allocation_table}{The FU Allocation table to be tidied.}

\item{year}{See \code{IEATools::iea_cols}.}

\item{max_vals}{See \code{IEATools::template_cols}.}

\item{values}{The name of the values column created. Default is "values".}
}
\value{
a tidy version of \code{.fu_allocation_table}.
}
\description{
FU (final-to-useful) allocation tables are not tidy data structures.
This function converts an FU Allocation table to a tidy data frame
by pulling the years (and the \code{max_vals} column) into a new column with the name \code{values}.
}
\examples{
load_fu_allocation_data() \%>\% 
  tidy_fu_allocation_table()
}
