% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_to_tidy.R
\name{tidy_iea}
\alias{tidy_iea}
\title{Title}
\usage{
tidy_iea(.aug_iea_df, country = "Country", year = "Year",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point", flow = "Flow",
  product = "Product", energy_type = "Energy.type", units = "Units",
  ex = "EX")
}
\arguments{
\item{.aug_iea_df}{}

\item{country}{}

\item{year}{}

\item{ledger_side}{}

\item{flow_aggregation_point}{}

\item{flow}{}

\item{product}{}

\item{energy}{}
}
\description{
Title
}
\examples{
file.path("extdata", "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") \%>\% 
  system.file(package = "IEAData") \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  remove_agg_memo_flows() \%>\% 
  use_iso_countries() \%>\% 
  tidy_iea()
}
