% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{tidy_iea_df}
\alias{tidy_iea_df}
\title{Creates a tidy IEA data frame}
\usage{
tidy_iea_df(
  .iea_df,
  col_names = IEATools::iea_cols,
  year = col_names$year,
  e_dot = col_names$e_dot,
  method = col_names$method,
  last_stage = col_names$last_stage,
  country = col_names$country,
  ledger_side = col_names$ledger_side,
  flow_aggregation_point = col_names$flow_aggregation_point,
  energy_type = col_names$energy_type,
  unit = col_names$unit,
  flow = col_names$flow,
  product = col_names$product,
  remove_zeroes = TRUE
)
}
\arguments{
\item{.iea_df}{A IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}.}

\item{col_names}{A list of column names in IEA data frames.
Default is \code{IEATools::iea_cols}.}

\item{year}{The name of the year column created in \code{.iea_df} by this function.
Default is \code{col_names$year}.}

\item{e_dot}{The name of the energy/exergy value column created in \code{.iea_df} by this function.
Default is \code{col_names$e_dot}.}

\item{method}{The name of the method column created in \code{.iea_df} by this function.
Default is \code{col_names$method}.}

\item{last_stage}{The name of the last stage column created in \code{.iea_df} by this function.
Default is \code{col_names$last_stage}.}

\item{country}{The name of the country column in \code{.iea_df}.
Default is \code{col_names$country}.}

\item{ledger_side}{The name of the ledger side in \code{.iea_df}.
Default is \code{col_names$ledger_side}.}

\item{flow_aggregation_point}{The name of the flow aggregation point column in \code{.iea_df}.
Default is \code{col_names$flow_aggregation_point}.}

\item{energy_type}{The name of the energy type column in \code{.iea_df}.
Default is \code{col_names$energy_type}.}

\item{unit}{The name of the unit column in \code{.iea_df}.
Default is \code{col_names$unit}.}

\item{flow}{The name of the flow column in \code{.iea_df}.
Default is \code{col_names$flow}.}

\item{product}{The name of the product column in \code{.iea_df}.
Default is \code{col_names$product}.}

\item{remove_zeroes}{A logical indicating whether data points with the value \code{0} are to be removed from the output.
Default is \code{TRUE}.}
}
\value{
A tidy version of \code{.iea_df} containing new columns \code{year} and \code{e_dot} and, optionally, \code{0} values removed.
}
\description{
Data from the IEA have years in columns,
but the \href{\doi{10.18637/jss.v059.i10}}{tidy data format}
requires one row for each datum.
This function uses \code{tidyr::pivot_longer()} to
make an IEA data frame tidy.
}
\details{
Default argument values assume that \code{rename_iea_df_cols()} has been called on \code{.iea_df}.
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  remove_agg_memo_flows() \%>\% 
  use_iso_countries() \%>\% 
  augment_iea_df() \%>\% 
  tidy_iea_df()
}
