% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{tidy_iea_df_balanced}
\alias{tidy_iea_df_balanced}
\title{Tell whether all rows of a tidy IEA data frame conserve energy.}
\usage{
tidy_iea_df_balanced(.tidy_iea_df_balances, balance_OK = "balance_OK")
}
\arguments{
\item{balance_OK}{the name of a new logical column that tells whether a row's energy balance is OK.
Default is "\code{balance_OK}".}

\item{.tidy_iea_df_balance}{an IEA-style data frame containing a column that indicates whether
each row is in balance.}
}
\value{
\code{TRUE} if all rows of \code{.tidy_iea_df_balance} are balanceed, \code{FALSE} otherwise.
}
\description{
This function provides a handy way to tell if all rows of \code{.tidy_iea_df_balance}
are in balance.
Argument \code{.tidy_iea_df_balance} should be set to the value of a call to
\code{\link[=calc_tidy_iea_df_balance]{calc_tidy_iea_df_balance()}}.
}
\examples{
library(dplyr)
file.path("extdata", "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") \%>\% 
  system.file(package = "IEAData") \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  remove_agg_memo_flows() \%>\% 
  augment_iea_df() \%>\% 
  tidy_iea_df() \%>\% 
  group_by(Country, Year, Energy.type, Units, Product) \%>\% 
  calc_tidy_iea_df_balance() \%>\% 
  tidy_iea_df_balanced()
}
