% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{tp_make_use_rows}
\alias{tp_make_use_rows}
\title{Transformation process make and use rows}
\usage{
tp_make_use_rows(.tidy_iea_df,
  flow_aggregation_point = "Flow.aggregation.point",
  transformation_processes = "Transformation processes",
  eiou = "Energy industry own use", flow = "Flow",
  product = "Product", e_dot = "E.dot", grouping_vars = c("Method",
  "Last.stage", "Country", "Year", "Energy.type"))
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{transformation_processes}{a string that identifies transformation processes in the \code{flow_aggregation_point} column. Default is "\code{Transformation processes}".}

\item{eiou}{a string that identifies energy industry own use in the \code{flow_aggregation_point} column. Default is "\code{Energy industry own use}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{e_dot}{the name of the energy rate column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}

\item{grouping_vars}{a string vector of column names by which \code{.tidy_iea_df} will be grouped before finding transformation sinks. Default is \code{c("Method", "Last.stage", "Country", "Year", "Energy.type")}.}
}
\value{
a list containing two data frames. "\code{make}" contains making Transformation process rows of \code{.tidy_iea_df}.
"\code{use}" contains using Transformation process rows of \code{.tidy_iea_df}.
}
\description{
An internal function that identifies making and using transformation process rows from \code{.tidy_iea_df}.
}
