% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tp_products}
\alias{tp_products}
\title{Identify all energy types supplied or consumed by transformation process machines}
\usage{
tp_products(file_path = file.path("extdata",
  "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") \%>\%
  system.file(package = "IEATools"),
  iea_df = IEATools::load_tidy_iea_df(file_path),
  type = c("Consumption", "Supply"),
  flow_aggregation_point = "Flow.aggregation.point",
  transformation_processes = "Transformation processes",
  e_dot = "E.dot", flow = "Flow", product = "Product")
}
\arguments{
\item{file_path}{the path to the IEA data file (optional)}

\item{iea_df}{a data frame containing IEA data. Default is \code{IEATools::load_tidy_iea_df(file_path)}.}
}
\value{
a list of string vectors
}
\description{
Sometimes, it is helpful to know all types of energy supplied or consumed by transformation processes.
This function (optionally) reads an IEA data file or loads an IEA data frame and builds a named list of energy types
supplied or consumed by transformation processes.
}
\details{
The names in the returned list are the transformation processes in \code{iea_df}.
The items in the returned list are vectors of energy types consumed by the corresponding transformation process.
}
\examples{
tp_products()
}
\seealso{
\code{tp_machine_production()}
}
