% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{tp_sinks_sources}
\alias{tp_sinks_sources}
\title{Find transformation sinks and sources}
\usage{
tp_sinks_sources(
  .tidy_iea_df,
  type = c("sinks", "sources"),
  flow_aggregation_point = "Flow.aggregation.point",
  transformation_processes = "Transformation processes",
  eiou = "Energy industry own use",
  flow = "Flow",
  product = "Product",
  e_dot = "E.dot"
)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{type}{one of "sinks" or "sources"}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{transformation_processes}{a string that identifies transformation processes in the \code{flow_aggregation_point} column. Default is "\verb{Transformation processes}".}

\item{eiou}{a string that identifies energy industry own use in the \code{flow_aggregation_point} column. Default is "\verb{Energy industry own use}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{e_dot}{the name of the energy rate column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}
}
\value{
The \code{grouping_vars} and the \code{flow} column,
with one row for each industry that is a transformation sink or source.
Industries that are transformation sinks or sources are named in the \code{flow} column.
}
\description{
In the IEA extended energy balance data,
transformation processes (tp) ought to both consume and produce energy.
But some transformation processes consume energy without producing any energy;
others produce without consuming.
Such transformation processes can be called "transformation sinks" and
"transformation sources," respectively.
This function finds and identifies transformation processes that act as sinks or sources.
}
\details{
It is important to identify transformation sinks,
because they cause two problems for physical supply-use table (PSUT) analysis.
First, when swimming upstream, a PSUT analysis cannot "see" the sunk energy carriers,
because they have no downstream effects.
Thus, upstream swims cannot conserve energy.
Second, when calculating embodied energy for each downstream energy carrier,
the sunk energy carriers cannot be embodied in any final demand energy carriers.
Thus, embodied energy calculations cannot conserve energy.

Transformation sources can also cause problems for physical supply-use table (PSUT) analysis.
In particular, when swimming upstream, a PSUT analysis will "see" the final energy sources,
but cannot see any associated primary energy carriers.

Transformation sinks and sources are identified by the following algorithm:
\enumerate{
\item Identify (per group in \code{.tidy_iea_df}) all \verb{Transformation processes} that consume energy (negative value for \code{E.dot}).
Energy consumption can be for the transformation process itself or for Energy industry own use.
\item Identify (per group in \code{.tidy_iea_df}) all \verb{Transformation processes} that produce energy (positive value for \code{E.dot}).
\item Take the set difference between the two (consumers less producers for sinks and producers less consumers for sources).
The set difference is the list of transformation sinks or sources, respectively.
}

\code{\link[=tp_sinks_sources]{tp_sinks_sources()}} is a function not unlike \code{\link[dplyr:summarise]{dplyr::summarise()}};
it returns a summary containing grouping variables and industries that are transformation sinks or sources.
So be sure to specify (or accept defaults for)
the \code{grouping_vars} argument.
Typical grouping variables are \code{Method}, \code{Last.stage}, \code{Country}, \code{Year}, \code{Energy.type}.
Don't group on \code{Flow.aggregation.point}, because energy from different aggregation points
(\verb{Energy industry own use} and \verb{Transformation processes}) flows into each machine.
Don't group on \code{Flow}, \code{Product}, or \code{E.dot}, either.
If groups are not set,
\code{flow}s will be analyzed together, possibly leading to missed transformation sinks or sources.

The various \verb{specify_*()} functions should also be called \emph{before} calling \code{\link[=tp_sinks_sources]{tp_sinks_sources()}}.
The \verb{specify_*()} functions clean up the IEA data, ensuring that energy is routed to the right places.

Note that this function only identifies transformation sinks or sources;
it does not fix the problem.
To solve the problem of transformation sinks,
see the \code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}} function.
\code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}} uses the output of \code{\link[=tp_sinks_sources]{tp_sinks_sources()}}
to route energy consumed by transformation sinks to \verb{Non-energy use industry/transformation/energy}.
There is no function to solve the problem of transformation sources at this time.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_interface_industries() \%>\% 
  specify_tp_eiou() \%>\% 
  tp_sinks_sources()
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_interface_industries() \%>\% 
  specify_tp_eiou() \%>\% 
  tp_sinks_sources(type = "sources")
}
