% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafficlight.R
\name{trafficlight}
\alias{trafficlight}
\title{Create a traffic light plot}
\usage{
trafficlight(
  x,
  time = NULL,
  sort_5yrmean = TRUE,
  sort_vec = NULL,
  method = "quantiles",
  probs = seq(0, 1, 0.2),
  quantile_type = 7,
  intervals = 5,
  cols = c("green3", "greenyellow", "yellow", "gold", "red"),
  main = "",
  xlab = "",
  ylab = "",
  adj_xlab = NULL,
  adj_ylab = NULL,
  hadj_x = 0.5,
  vadj_x = -0.7,
  hadj_y = -0.7,
  vadj_y = 0.3,
  madj_bottom = 0,
  madj_left = 0,
  madj_top = 0,
  madj_right = 0,
  legend_pos = "top",
  legend_intersp_x = 1,
  legend_intersp_y = 1,
  tick_x = FALSE,
  tick_y = FALSE,
  cex_x = 0.8,
  cex_y = 0.8,
  cex_xlab = 1,
  cex_ylab = 1,
  cex_legend = 0.8,
  cex_main = 1,
  respect = TRUE
)
}
\arguments{
\item{x}{matrix or data frame containing time series of multiple variables.}

\item{time}{vector of time units that will be used for the x axis.}

\item{sort_5yrmean}{logical; should the variables be sorted by the first 5yr mean? Default
is set to TRUE.}

\item{sort_vec}{integer vector; if specific order of variable is desired the sorting index should be
provided here.}

\item{method}{character; the type of methods to create the color code. Choose
between using "quantiles" (default) or "intervals".}

\item{probs}{a vector of probabilities to calculate the quantiles using the function
\code{quantile}. Should include the probabilities 0 and 1.
numeric vector of probabilities with values in [0,1]. (Values up to
2e-14 outside that range are accepted and moved to the nearby endpoint.)}

\item{quantile_type}{an integer between 1 and 9 selecting one of the nine quantile
algorithms detailed below to be used. Default is 7 (see also \code{\link{quantile}}).}

\item{intervals}{logical; number of evenly spaced intervals. Default is 5.}

\item{cols}{a character vector with colors for each quantile or interval.}

\item{main}{a title.}

\item{xlab}{The x axis title. Default is none.}

\item{ylab}{The y axis title. Default is none.}

\item{adj_xlab}{integer; vertical adjustment of the x axis title.}

\item{adj_ylab}{integer; horizontal adjustment of the y axis title.}

\item{hadj_x}{double; horizontal adjustment of the x labels.}

\item{vadj_x}{double; vertical adjustment of the x labels.}

\item{hadj_y}{double; horizontal adjustment of the y labels.}

\item{vadj_y}{double; vertical adjustment of the y labels.}

\item{madj_bottom}{double; adjustment of bottom margin relative to the
current setting.}

\item{madj_left}{double; adjustment of left margin relative to the
current setting.}

\item{madj_top}{double; adjustment of top margin relative to the
current setting.}

\item{madj_right}{double; adjustment of right margin relative to the
current setting.}

\item{legend_pos}{character; the legend is shown on the "top" (default),
"center", or "bottom" right of the plot.}

\item{legend_intersp_x}{double; character interspacing factor for horizontal (x) spacing in legend.}

\item{legend_intersp_y}{double; interspacing factor for vertical (y) line distances in legend.}

\item{tick_x}{logical; set to TRUE if ticks on x-axis should be displayed.}

\item{tick_y}{logical; set to TRUE if ticks on x-axis should be displayed.}

\item{cex_x}{double; the magnification to be used for the x labels relative to the
current setting of cex.}

\item{cex_y}{double; the magnification to be used for the y labels relative to the
current setting of cex.}

\item{cex_xlab}{double; the magnification to be used for the x axis title relative
to the current setting of cex.}

\item{cex_ylab}{double; the magnification to be used for the y axis title  relative
to the current setting of cex.}

\item{cex_legend}{double; the magnification to be used for the legend relative to
the current setting of cex.}

\item{cex_main}{double; the magnification to be used for the title relative to the
current setting of cex.}

\item{respect}{logical; this argument controls whether a unit column-width is the
same physical measurement on the device as a unit row-height (default is TRUE).}
}
\description{
This function creates for multiple time series an image plot where the color
code is based on selected quantiles or evenly spaced intervals.
}
\examples{
df <- data.frame(variable1 = 1:20, variable2= rnorm(20, 100, 30), variable3 = 1:20 + rnorm(20))
trafficlight(x = df, time = 1981:2000)
df <- matrix(rnorm(100), ncol = 5)
colnames(df) <- letters[1:5]
trafficlight(x = df, time = 1981:2000, legend_pos = "bottom", method = "intervals")
}
\author{
Saskia A. Otto
}
