% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transformation_processes}
\alias{transformation_processes}
\title{Transformation processes}
\format{
A string vector with 6 entries
\describe{
\item{main_activity_producer_electricity_plants}{The string that identifies main activity producer electricity plants.}
\item{autoproducer_electricity_plants}{The string that identifies autoproducer electricity plants.}
\item{main_activity_producer_CHP_plants}{The string that identifies main activity producer combined heat and power plants.}
\item{autoproducer_CHP_plants}{The string that identifies autoproducer combined heat and power plants.}
\item{main_activity_producer_heat_plants}{The string that identifies main activity producer heat plants.}
\item{autoproducer_heat_plants}{The string that identifies autoproducer heat plants.}
\item{heat_pumps}{The string that identifies heat pumps.}
\item{electric_boilers}{The string that identifies electric boilers.}
\item{chemical_heat_for_electricity_production}{The string that identifies chemical heat for electricity production.}
\item{blast_furnaces}{The string that identifies blast furnaces.}
\item{gas_works}{The string that identifies gas works.}
\item{coke_ovens}{The string that identifies coke ovens.}
\item{patent_fuel_plants}{The string that identifies patent fuel plants.}
\item{bkb_peat_briquette_plants}{The string that identifies BKB/peat briquette plants.}
\item{oil_refineries}{The string that identifies oil refineries.}
\item{petrochemical_plants}{The string that identifies petrochemical plants.}
\item{coal_liquefaction_plants}{The string that identifies coal liquefaction plants.}
\item{gas_to_liquid_gtl_plants}{The string that identifies gas to liquid (GTL) plants.}
\item{for_blended_natural_gas}{The string that identifies for blended natural gas.}
\item{charcoal_production_plants}{The string that identifies charcoal production plants.}
\item{non_specified_transformation}{The string that identifies non-specified transformation.}
}
}
\usage{
transformation_processes
}
\description{
A string vector containing names of \verb{Transformation processes} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
transformation_processes
}
\keyword{datasets}
