% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{transformation_sinks}
\alias{transformation_sinks}
\title{Find transformation sinks}
\usage{
transformation_sinks(.tidy_iea_data)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}
}
\value{
the rows of \code{.tidy_iea_df} that represent energy consumed by transformation sinks
}
\description{
In the IEA extended energy balance data,
transformation processes are expected to both consume and make energy.
But some transformation processes consume energy without making any energy.
Those transformation processes are called "transformation sinks."
This function finds and identifies transformation sinks.
}
\details{
It is important to identify transformation sinks,
because they cause two problems for input-output analysis.
First, when swimming upstream, we cannot "see" the sunk energy carriers,
because tjhey have no downstream effects.
Thus, upstream swims cannot conserve energy.
Second, when calculating embodied energy for each downstream energy carrier,
the Fuel oil, LPG, or Gas/diesel oil excl. biofuels cannot be embodied.
Thus, embodied energy calculations cannot conserve energy.

Note that this function only identifies the problem, it does not fix the problem.
\code{\link[=transformation_sinks]{transformation_sinks()}} is a function not unlike \code{\link[dplyr:filter]{dplyr::filter()}}:
it returns rows that represent energy consumed by transformation sinks.
The \code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}} function uses the output of \code{\link[=transformation_sinks]{transformation_sinks()}}
to route energy consumed by transformation sinks to \code{Non-energy use industry/transformation/energy}.
}
\examples{
load_tidy_iea_df() \%>\% 
  transformation_sinks()
}
