% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_flows}
\alias{transport_flows}
\title{Transport flows}
\format{
A string vector with 9 entries.
\describe{
\item{world_aviation_bunkers}{The string identifying the World aviation bunkers transport sector}
\item{world_marine_bunkers}{The string identifying the World marine bunkers transport sector}
\item{domestic_aviation}{The string identifying Domestic aviation}
\item{road}{The string identifying the Road transport sector}
\item{rail}{The string identifying the Rail transport sector}
\item{pipeline_transport}{The string identifying the Pipeline transport sector}
\item{domestic_navigation}{The string identifying the Domestic navigation sector}
\item{non_specified_transport}{The string identifying the Non-specified transport sector}
\item{transport_not_elsewhere_specified}{The string identifying the Transport not elsewhere specified sector. This string replaced \code{non_specified_transport} in the IEA's 2020 extended energy balances.}
}
}
\usage{
transport_flows
}
\description{
A string vector containing names of \code{Transport} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
transport_flows
}
\keyword{datasets}
