% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_iea_release_years}
\alias{valid_iea_release_years}
\title{IEA release years supported by this package}
\format{
A numeric vector with 2 entries
}
\usage{
valid_iea_release_years
}
\description{
A numeric vector containing release years for IEA extended energy balance data
supported by this package.
}
\examples{
valid_iea_release_years
}
\keyword{datasets}
