% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_fu_allocation_template}
\alias{write_fu_allocation_template}
\title{Write blank final-to-useful templates}
\usage{
write_fu_allocation_template(.tidy_iea_df, path,
  ledger_side = "Ledger.side", consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  allocations_tab_name = "Allocations", quantity = "Quantity",
  e_dot = "E.dot", e_dot_perc = paste(e_dot, "[\%]"),
  maximum_values = "Maximum.values",
  energy_row_font_color_fd = "#104273",
  energy_row_shading_color_fd = "#B8D8F5",
  energy_row_font_color_eiou = "#918700",
  energy_row_shading_color_eiou = "#FCFCAB", overwrite = FALSE,
  .rownum = ".rownum")
}
\arguments{
\item{.tidy_iea_df}{a data frame as discussed in Details}

\item{path}{the file path into which the blank template file will be written.
Include both folder and file name.
If not present, the ".xlsx" extension is added.}

\item{allocations_tab_name}{the name of the tab on which the template will be written. Default is "\code{Allocations}".}

\item{quantity}{the name of the quantity column to be created on output. Default is "\code{Quantity}".}

\item{e_dot}{the name of the energy flow rate column in \code{.tidy_iea_df} and the name of the energy flow rate rows to be included in the Excel file that is written by this function.
Default is "\code{E.dot}".}

\item{e_dot_perc}{the name of the energy flow rate percentage row to be included in the Excel file that is written by this function.
Default is "\code{E.dot.perc}".}

\item{overwrite}{a boolean that tells whether an existing file at \code{path} will be overwritten. Default is "\code{FALSE}".
If \code{path} already exists and \code{overwrite = FALSE}, an error is given.}

\item{.rownum}{a temporary column created internally. \code{.rownum} must not exist in \code{.tidy_iea_df} when \code{write_fu_allocation_template} is called.
Default is "\code{.rownum}".}

\item{energy_row_font_color}{a hex string representing the font color to be used for \code{e_dot} and \code{e_dot_perc} rows in the Excel file that is written by this function.
Default is "\code{#104273}", a light blue color.}

\item{energy_row_shading_color}{a hex string representing the font color to be used for \code{e_dot} and \code{e_dot_perc} rows in the Excel file that is written by this function.
Default is "\code{#104273}".}
}
\value{
the value of the \code{path} argument, regardless of the value of \code{overwrite}.
}
\description{
The analyst often starts with IEA extended energy balance data covering primary and final energy stages.
This function writes a blank Excel template file that, when filled,
allow the energy conversion chain to be extended to the useful stage.
}
\details{
This function returns a data frame that is the template that needs to be filled.
Use the function \code{write.csv()} or similar to save to disk.
The analyst must supply allocation and efficiency data.
Allocations are to be placed in \code{c} rows and efficiencies are to be placed in \code{eta} rows.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  write_fu_templates()
}
