% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_fu_eta_template}
\alias{write_fu_eta_template}
\title{Write a final-to-useful efficiencies template}
\usage{
write_fu_eta_template(.fu_eta_template, path,
  fu_eta_tab_name = "FU etas", overwrite_file = FALSE,
  overwrite_fu_eta_tab = FALSE)
}
\arguments{
\item{path}{}

\item{fu_eta_tab_name}{}

\item{overwrite_file}{}

\item{overwrite_eta_tab}{}
}
\description{
Write a final-to-useful efficiencies template
}
