% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_fu_templates}
\alias{write_fu_templates}
\title{Write blank final-to-useful templates}
\usage{
write_fu_templates(.iea_df, dir)
}
\arguments{
\item{.iea_df}{a data frame as discussed in Details}

\item{dir}{the directory into which blank templates will be written}
}
\value{
a data frame containing the blank template
}
\description{
The analyst often starts with data covering primary and final energy stages.
This function writes blank template files that, when filled,
allow the energy conversion chain to be extended to the useful stage.
}
\details{
\code{.iea_df} should have columns

This function returns a data frame that is the template that needs to be filled.
Use the function \code{write.csv()} or similar to save to disk.
The analyst must supply allocation and efficiency data.
Allocations are to be placed in \code{c} rows and efficiencies are to be placed in \code{eta} rows.
}
