% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_psut_matnames}
\alias{add_psut_matnames}
\title{Add a column of matrix names to tidy data frame}
\usage{
add_psut_matnames(.tidy_iea_df, ledger_side = "Ledger.side",
  e_dot = "E.dot", flow_aggregation_point = "Flow.aggregation.point",
  flow = "Flow", supply = "Supply", consumption = "Consumption",
  eiou = "Energy industry own use", neg_supply_in_fd = c("Exports",
  "International aviation bunkers", "International marine bunkers",
  "Losses", "Statistical differences", "Stock changes"),
  matname = "matname", U_excl_EIOU = "U_excl_EIOU",
  U_EIOU = "U_EIOU", R = "R", V = "V", Y = "Y", .R = ".R")
}
\arguments{
\item{.tidy_iea_df}{a data frame with \code{ledger_side}, \code{energy}, \code{flow_aggregation_point},
and \code{flow} columns.}

\item{ledger_side}{the name of the column in \code{.DF} that contains ledger side
(a string). Default is "\code{Ledger.side}".}

\item{flow_aggregation_point}{the name of the column in \code{.DF} that contains flow aggregation point information.
Default is "\code{Flow.aggregation.point}".}

\item{flow}{the name of the column in \code{.DF} that contains flow information.
Default is "\code{Flow}".}

\item{eiou}{the identifier for items that are energy industry own use.
Default is "\code{"Energy industry own use"}.}

\item{neg_supply_in_fd}{identifiers for \code{flow} items that, when negative,
are entries in the final demand (\code{Y}) matrix.}

\item{matname}{the name of the output column containing the name of the matrix
to which a row's value belongs (a string). Default is "\code{matname}".}

\item{U_excl_EIOU}{the name for the use matrix that excludes energy industry own use (a string). Default is "\code{U_excl_EIOU}".}

\item{U_EIOU}{the name for the energy industry own use matrix. Default is "\code{U_EIOU}".}

\item{R}{the name for the resource matrix (a string). Default is "\code{R}".}

\item{V}{the name for the make matrix (a string). Default is "\code{V}".}

\item{Y}{the name for the final demand matrix (a string). Default is "\code{Y}".}

\item{energy}{the name of the column in \code{.DF} that contains energy and exergy values
(a string). Default is "\code{EX.ktoe}".}

\item{supply_side}{the identifier for items on the supply side of the ledger (a string).
Default is "\code{Supply}".}

\item{consumption_side}{the identifier for items on the consumption side
of the ledger (a string). Default is "\code{Consumption}".}

\item{use_R}{tells whether to use a separate resources matrix (\code{R}). Default is "\code{FALSE}", for now.}
}
\value{
\code{.tidy_iea_df} with an added column \code{matname}.
}
\description{
This function adds a column of matrix names to a tidy data frame
wherein each row of \code{.tidy_iea_df} is a single value in an energy conversion chain.
The default argument values assume that \code{.tidy_iea_df} uses IEA-style nomenclature
and terminology, although \code{.tidy_iea_df} does not necessarily need to contain IEA data.
}
\examples{
library(dplyr)
UKEnergy2000tidy \%>\%
  group_by(Country, Year, Energy.type, Last.stage) \%>\%
  add_matnames_iea() \%>\%
  glimpse()
}
