% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_psut_matnames}
\alias{add_psut_matnames}
\title{Add a column of matrix names to tidy data frame}
\usage{
add_psut_matnames(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  production = IEATools::tpes_flows$production,
  resources = IEATools::tpes_flows$resources,
  eiou = "Energy industry own use",
  neg_supply_in_fd = c("Exports", "International aviation bunkers",
    "International marine bunkers", "Losses", "Statistical differences", "Stock changes"),
  matnames = IEATools::mat_meta_cols$matnames,
  R = IEATools::psut_cols$R,
  U_excl_EIOU = IEATools::psut_cols$U_excl_eiou,
  U_EIOU = IEATools::psut_cols$U_eiou,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y
)
}
\arguments{
\item{.tidy_iea_df}{a data frame with \code{ledger_side}, \code{flow_aggregation_point}, \code{flow}, and \code{e_dot} columns.}

\item{ledger_side, flow_aggregation_point, flow, product, e_dot}{See \code{IEATools::iea_cols}.}

\item{supply, consumption}{See \code{IEATools::ledger_sides}.}

\item{production, resources}{See \code{IEATools::tpes_flows}.}

\item{eiou}{the identifier for items that are energy industry own use.
Default is "Energy industry own use".}

\item{neg_supply_in_fd}{identifiers for flow items that, when negative,
are entries in the final demand (\code{Y}) matrix.}

\item{matnames}{the name of the output column containing the name of the matrix
to which a row's value belongs (a string). Default is "matnames".}

\item{R, U_excl_EIOU, U_EIOU, V, Y}{See \code{IEATools::psut_matnames}.}
}
\value{
\code{.tidy_iea_df} with an added column \code{matnames}.
}
\description{
This function adds a column of matrix names to a tidy data frame
wherein each row of \code{.tidy_iea_df} is a single value in an energy conversion chain.
The default argument values assume that \code{.tidy_iea_df} uses IEA-style nomenclature
and terminology, although \code{.tidy_iea_df} does not necessarily need to contain IEA data.
}
\details{
In a reasonable workflow, this function would be followed by a call to
\code{add_row_col_meta()} and \code{matsindf::collapse_to_matrices()}.

This function respects groups when identifying entries in the resource matrix (\code{R}).
So be sure to group \code{.tidy_iea_df} before calling this function.

Internally, this function adds a temporary column to \code{.tidy_iea_df} called ".R".
An error will occur if \code{.tidy_iea_df} already has a column named ".R".
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  add_psut_matnames() \%>\%
  glimpse()
}
