% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_row_col_meta}
\alias{add_row_col_meta}
\title{Add row, column, row type, and column type metadata}
\usage{
add_row_col_meta(.tidy_iea_df, matname = "matname",
  product = "Product", flow = "Flow", U = "U", U_EIOU = "U_EIOU",
  R = "R", V = "V", Y = "Y", industry_type = "Industry",
  product_type = "Product", sector_type = "Industry",
  resource_type = "Industry", rowname = "rowname",
  colname = "colname", rowtype = "rowtype", coltype = "coltype")
}
\arguments{
\item{.tidy_iea_df}{a data frame containing \code{matname}.}

\item{matname}{the name of the column in \code{.tidy_iea_df} that contains names of matrices
(a string).  Default is "\code{matname}".}

\item{product}{the name of the column in \code{.tidy_iea_df} where Product names
is found (a string). Default is "\code{Product}".}

\item{flow}{the name of the column in \code{.tidy_iea_df} where Flow names
is found (a string). Default is "\code{Flow}".}

\item{U}{the name for use matrices (a string). Default is "\code{U}".}

\item{U_EIOU}{the name for energy industry own use matrices (a string). Default is "\code{U_EIOU}".}

\item{R}{the name for resource matrices (a string). Default is "\code{R}".}

\item{V}{the name for make matrices (a string). Default is "\code{V}".}

\item{Y}{the name for final demand matrices (a string). Default is "\code{Y}".}

\item{industry_type}{the name that identifies production industries and
and transformation processes (a string). Default is "\code{Industry}".}

\item{product_type}{the name that identifies energy carriers (a string).
Default is "\code{Product}".}

\item{sector_type}{the name that identifies final demand sectors (a string).
Default is "\code{Industry}".}

\item{resource_type}{the name that identifies resource sectors (a string).
Default is "\code{Industry}".}

\item{rowname}{the name of the output column that contains row names for matrices
(a string). Default is "\code{rowname}".}

\item{colname}{the name of the output column that contains column names for matrices
(a string). Default is "\code{colname}".}

\item{rowtype}{the name of the output column that contains row types for matrices
(a string). Default is "\code{rowtype}".}

\item{coltype}{the name of the output column that contains column types for matrices
(a string). Default is "\code{coltype}".}
}
\value{
\code{.tidy_iea_df} with additional columns named
\code{rowname}, \code{colname},
\code{rowtype}, and \code{coltype}.
}
\description{
Add row, column, row type, and column type metadata
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  add_psut_matnames() \%>\%
  add_row_col_meta()
}
