% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_row_col_meta}
\alias{add_row_col_meta}
\title{Add row, column, row type, and column type metadata}
\usage{
add_row_col_meta(
  .tidy_iea_df,
  product = IEATools::iea_cols$product,
  flow = IEATools::iea_cols$flow,
  matnames = IEATools::mat_meta_cols$matnames,
  U = IEATools::psut_cols$U,
  U_EIOU = IEATools::psut_cols$U_eiou,
  R = IEATools::psut_cols$R,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  Epsilon = IEATools::psut_cols$Epsilon,
  industry_type = IEATools::row_col_types$industry,
  product_type = IEATools::row_col_types$product,
  sector_type = IEATools::row_col_types$sector,
  resource_type = IEATools::row_col_types$resource,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes
)
}
\arguments{
\item{.tidy_iea_df}{a data frame containing column \code{matnames}}

\item{flow, product}{See \code{IEATools::iea_cols}.}

\item{matnames}{the name of the column in \code{.tidy_iea_df} that contains names of matrices
(a string).  Default is "matnames".}

\item{R, U, U_EIOU, V, Y, Epsilon}{See \code{IEATools::psut_cols}.}

\item{industry_type, product_type, sector_type, resource_type}{See \code{IEATools::row_col_types}.}

\item{rownames, colnames, rowtypes, coltypes}{See \code{IEATools::mat_meta_cols}.}
}
\value{
\code{.tidy_iea_df} with additional columns named
\code{rowname}, \code{colname},
\code{rowtype}, and \code{coltype}.
}
\description{
After calling \code{add_psut_matnames()}, call this function
to add \code{rownames}, \code{colnames}, \code{rowtypes}, and \code{coltypes} columns to \code{.tidy_iea_df}.
}
\details{
If \code{.tidy_iea_df} already contains all of \code{rownames}, \code{colnames}, \code{rowtypes}, and \code{coltypes},
\code{.tidy_iea_df} is returned without modification.
If \code{.tidy_iea_df} contains some but not all of \code{rownames}, \code{colnames}, \code{rowtypes}, or \code{coltypes},
an error is returned.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  add_psut_matnames() \%>\%
  add_row_col_meta()
}
