% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{add_row_col_meta}
\alias{add_row_col_meta}
\title{Add row, column, row type, and column type metadata}
\usage{
add_row_col_meta(
  .tidy_iea_df,
  cols = IEATools::iea_cols,
  product = cols$product,
  flow = cols$flow,
  matnames = "matnames",
  U = "U",
  U_EIOU = "U_EIOU",
  R = "R",
  V = "V",
  Y = "Y",
  industry_type = "Industry",
  product_type = "Product",
  sector_type = "Industry",
  resource_type = "Industry",
  rownames = "rownames",
  colnames = "colnames",
  rowtypes = "rowtypes",
  coltypes = "coltypes"
)
}
\arguments{
\item{.tidy_iea_df}{a data frame containing column \code{matnames}}

\item{cols}{a list containing names of columns. Default is \code{IEATools::iea_cols}.}

\item{product}{the name of the column in \code{.tidy_iea_df} where Product names
is found (a string). Default is \code{cols$product}.}

\item{flow}{the name of the column in \code{.tidy_iea_df} where Flow names
is found (a string). Default is \code{cols$flow}.}

\item{matnames}{the name of the column in \code{.tidy_iea_df} that contains names of matrices
(a string).  Default is "matnames".}

\item{U}{the name for use matrices (a string). Default is "U".}

\item{U_EIOU}{the name for energy industry own use matrices (a string). Default is "U_EIOU".}

\item{R}{the name for resource matrices (a string). Default is "R".}

\item{V}{the name for make matrices (a string). Default is "V".}

\item{Y}{the name for final demand matrices (a string). Default is "Y".}

\item{industry_type}{the name that identifies production industries and
and transformation processes (a string). Default is "Industry".}

\item{product_type}{the name that identifies energy carriers (a string).
Default is "Product".}

\item{sector_type}{the name that identifies final demand sectors (a string).
Default is "Industry".}

\item{resource_type}{the name that identifies resource sectors (a string).
Default is "Industry".}

\item{rownames}{the name of the output column that contains row names for matrices
(a string). Default is "rowname".}

\item{colnames}{the name of the output column that contains column names for matrices
(a string). Default is "colname".}

\item{rowtypes}{the name of the output column that contains row types for matrices
(a string). Default is "rowtype".}

\item{coltypes}{the name of the output column that contains column types for matrices
(a string). Default is "coltype".}
}
\value{
\code{.tidy_iea_df} with additional columns named
\code{rowname}, \code{colname},
\code{rowtype}, and \code{coltype}.
}
\description{
After calling \code{\link[=add_psut_matnames]{add_psut_matnames()}}, call this function
to add row, column, row type, and column type
information to \code{.tidy_iea_df}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  add_psut_matnames() \%>\%
  add_row_col_meta()
}
