% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{aggregate_regions}
\alias{aggregate_regions}
\title{Aggregates IEA regions based on a user-defined concordance matrix}
\usage{
aggregate_regions(
  .tidy_iea_df,
  file_path = "data-raw/default_mapping_exiobase.csv",
  net_trade = FALSE,
  country = IEATools::iea_cols$country,
  destination_regions = "Destination_regions"
)
}
\arguments{
\item{.tidy_iea_df}{The \code{.tidy_iea_df} data frame that needs to be aggregated by regions. The \code{.tidy_iea_df} is likely
to have been obtained with the \code{load_tidy_iea_df()} function.}

\item{file_path}{The path of the file to be loaded. By default, a concordance table converting IEA regions into Exiobase regions
is loaded.}

\item{net_trade}{The boolean that defines whether imports and exports by aggregation region should be converted
into net imports / exports or not. Default is FALSE.}
}
\value{
A \code{.tidy_iea_df} that contains the data of the input \code{.tidy_iea_df} aggregated by regions as specified in the user-defined
country concordance table provided.
}
\description{
Takes as input a tidy dataframe, the file of a country concordance table, and aggregates flows per regions following the
user-defined concordance table. The boolean argument "net_trade" enables to perform the aggregation by keeping only net imports
and/or net exports or by keeping gross imports and exports.
}
