% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation_flows}
\alias{aggregation_flows}
\title{Aggregation flows}
\format{
A string vector with 10 entries.
\describe{
\item{aggregate_primary}{The name of the column containing aggregated primary energy.}
\item{net_aggregate_demand}{The name of the column containing aggregated net final demand energy.}
\item{gross_aggregate_demand}{The name of the column containing aggregated gross final demand energy.}
}
}
\usage{
aggregation_flows
}
\description{
A string vector containing names of industries whose purpose in IEA extended energy balance \code{Flow}s is to provide aggregations.
These items appear in \code{Flow.aggregation.point} columns.
}
\examples{
aggregation_flows
}
\keyword{datasets}
