% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_stages}
\alias{all_stages}
\title{All energy stages}
\format{
A string list with 4
\describe{
\item{primary}{The string identifier for the Primary stage of the energy conversion chain.}
\item{final}{The string identifier for the Final stage of the energy conversion chain.}
\item{useful}{The string identifier for the Useful stage of the energy conversion chain.}
\item{services}{The string identifier for the Services stage of the energy conversion chain.}
}
}
\usage{
all_stages
}
\description{
A string list containing options for the all stages of energy conversion chain analysis.
}
\examples{
all_stages
}
\keyword{datasets}
