% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{arrange_iea_fu_allocation_template_rows}
\alias{arrange_iea_fu_allocation_template_rows}
\title{Arrange the rows of an fu_allocation template}
\usage{
arrange_iea_fu_allocation_template_rows(.fu_allocation_template,
  flow_aggregation_point = "Flow.aggregation.point",
  destionation = "Destionation",
  fap_dest_order = IEATools::fap_flow_iea_order, product = "Product",
  product_order = IEATools::product_iea_order,
  ef_product = "Ef.product", machine = "Machine",
  eu_product = "Eu.product", destination = "Destination",
  quantity = "Quantity", maximum_values = "Maximum.values",
  .temp_sort = ".fap_flow")
}
\arguments{
\item{product}{}

\item{product_order}{}

\item{ef_product}{}

\item{machine}{}

\item{eu_product}{}

\item{destination}{}

\item{quantity}{}

\item{maximum_values}{}

\item{.temp_sort}{}

\item{.iea_data}{}

\item{flow_allocation_point}{}

\item{flow}{}

\item{fap_flow_order}{}
}
\description{
It is helpful to sort rows of a final-to-useful allocation template
to match the row order in the IEA's extended energy balance data to assist analysts with final-to-useful
allocations.
This function sorts rows (top-to-bottom) by default in the same order as appears in the IEA extended energy balance data.
Other orderings can be specified with the \code{fap_flow_order} and \code{product_order} arguments.
}
\details{
The \code{flow_allocation_point}, \code{flow}, and \code{product} columns are considered for the ordering.

Internally, this function figures out which columns are metadata columns
and groups on those columns before applying the row ordering.
If you want to preserve ordering by metadata, be sure to set factor levels on metadata columns prior to calling this function.
}
\examples{
Allocation_template <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\%
  fu_allocation_template() \%>\% 
  arrange_iea_fu_allocation_template_rows()
}
