% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{augment_iea_df}
\alias{augment_iea_df}
\title{Augment an IEA data frame}
\usage{
augment_iea_df(
  .iea_df,
  country = "Country",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  flow = "Flow",
  product = "Product",
  energy_type = "Energy.type",
  energy_type_val = "E",
  method = "Method",
  method_val = "PCM",
  last_stage = "Last.stage",
  last_stage_val = "Final",
  unit = "Unit",
  unit_val = "TJ",
  supply = "Supply",
  consumption = "Consumption",
  tpes = IEATools::tfc_compare_flows$total_primary_energy_supply,
  tpes_flows = IEATools::tpes_flows,
  tfc_compare = "TFC compare",
  tfc_compare_flows = IEATools::tfc_compare_flows,
  transfers = "Transfers",
  statistical_differences = IEATools::tfc_compare_flows$statistical_differences,
  losses = IEATools::tfc_compare_flows$losses,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  tp_flows_suffix = "(transf.)",
  nstp_flows_suffix = "(transformation)",
  mapep = IEATools::transformation_processes$main_activity_producer_electricity_plants,
  eiou = "Energy industry own use",
  eiou_flows_suffix = "(energy)",
  coal_mines = "Coal mines",
  non_specified = "Non-specified",
  tfc = "Total final consumption",
  tfc_flows = IEATools::tfc_flows,
  industry = "Industry",
  industry_flows = IEATools::industry_flows,
  iron_and_steel = "Iron and steel",
  mining_and_quarrying = "Mining and quarrying",
  transport = "Transport",
  transport_flows = IEATools::transport_flows,
  other = "Other",
  other_flows = IEATools::other_flows,
  non_energy = "Non-energy use",
  non_energy_flows = IEATools::non_energy_flows,
  memo_non_energy_flows = IEATools::memo_non_energy_flows,
  electricity_output = "Electricity output (GWh)",
  electricity_output_flows_prefix = "Electricity output (GWh)-",
  heat_output = "Heat output",
  heat_output_flows_prefix = "Heat output-",
  .rownum = ".rownum"
)
}
\arguments{
\item{.iea_df}{A data frame produced by the \code{iea_df()} function.}

\item{country}{The name of the country column in \code{.iea_df}. Default is "Country".}

\item{ledger_side}{The name of the ledger side column to be added to \code{.iea_df}. Default is "Ledger.side".}

\item{flow_aggregation_point}{The name of the flow aggregation point column to be added to \code{.iea_df}. Default is "Flow.aggregation.point".}

\item{flow}{The name of the flow column in \code{.iea_df}.  Default is "Flow".}

\item{product}{The name of the product column in \code{.iea_df}.  Default is "Product".}

\item{energy_type}{The name of the energy type column to be added to \code{.iea_df}. Default is "Energy.type.}

\item{energy_type_val}{The value to put in the \code{energy_type} column. Default is "E".}

\item{method}{The name of the method column to be added to \code{.iea_df}. Default is "Method".}

\item{method_val}{The value to put in the \code{method} column. Default is "PCM" (Physical Content Method, which is used by the IEA).}

\item{last_stage}{The name of the last stage column to be added to \code{.iea_df}. Default is "Last.stage".}

\item{last_stage_val}{The value to put in the \code{last_stage} column. Default is "Final" (which is the last stage supplied by the IEA).}

\item{unit}{The name of the unit column to be added to \code{.iea_df}. Default is "Unit".}

\item{unit_val}{The value to put in the \code{unit} column. Default is "TJ" for terajoule.}

\item{supply}{The string that identifies supply \code{Ledger.side}. Default is "Supply".}

\item{consumption}{The string that identifies consumption \code{Ledger.side}. Default is "Consumption".}

\item{tpes}{The string that identifies total primary energy supply \code{Flow.aggregation.point}. Default is "Total primary energy supply".}

\item{tpes_flows}{A vector of strings that give flows that are aggregated to \verb{Total primary energy supply}.}

\item{tfc_compare}{A string that identifies the \verb{TFC compare} flow aggregation point. Default is "TFC compare".}

\item{tfc_compare_flows}{A vector of strings that give \code{Flow}s that are aggregated to \verb{TFC compare}.}

\item{transfers}{= A string that identifies transfers in the flow column. Default is "Transfers".}

\item{statistical_differences}{A string that identifies statistical differences in flow column. Default is "Statistical differences".}

\item{losses}{The string that indicates losses in the \code{Flow} column. Default is "Losses".}

\item{transformation_processes}{The string that indicates transformation processes in the \code{Flow} column. Default is "Transformation processes".}

\item{tp_flows_suffix}{The suffix for transformation processes in the \code{Flow} column. Default is "(transf.)".}

\item{nstp_flows_suffix}{The suffix for non-specified transformation processes in the \code{Flow} column. Default is "(transformation)".}

\item{mapep}{The string that identifies main activity producer electricity plants in the \code{Flow} column. Default is "Main activity producer electricity plants".}

\item{eiou}{The string that identifies energy industry own use in the \code{Flow} column. Default is "Energy industry own use".}

\item{eiou_flows_suffix}{The suffix for energy industry own use in the \code{Flow} column. Default is "(energy)".}

\item{coal_mines}{The string that identifies coal mines in the \code{Flow} column. Default is "Coal mines".}

\item{non_specified}{The string that identifies non-specified flows in the \code{Flow} column. Default is "Non-specified".}

\item{tfc}{The string that identifies total final consumption in the \code{Flow} column. Default is "Total final consumption".}

\item{tfc_flows}{A vector of strings that give total final consumption in the \code{Flow} column.}

\item{industry}{A string that names the industry \code{Flow.aggregation.point}. Default is "Industry".}

\item{industry_flows}{A vector of strings representing \code{Flow}s to be aggregated in the \code{Industry} \code{Flow.aggregation.point}.}

\item{iron_and_steel}{A string that identifies the iron and steel industry. Default is "Iron and steel".}

\item{mining_and_quarrying}{A string that identifies the mining and quarrying industry. Default is "Mining and quarrying".}

\item{transport}{A string that names the transport \code{Flow.aggregation.point}. Default is "Transport".}

\item{transport_flows}{A vector of strings representing \code{Flow}s to be aggregated in the \code{Transport} \code{Flow.aggregation.point}.}

\item{other}{A string that names the other \code{Flow.aggregation.point}. Default is "Other".}

\item{other_flows}{A vector of strings representing \code{Flow}s to be aggregated in the \code{Other} \code{Flow.aggregation.point}.}

\item{non_energy}{A string that names the non-energy \code{Flow.aggregation.point}. Default is "Non-energy use".}

\item{non_energy_flows}{A list of \code{Flow}s to be aggregated to the \verb{Non-energy use} \code{Flow.aggregation.point}.}

\item{memo_non_energy_flows}{A list of \code{Flow}s to be aggregated to "Memo: Non-energy use in industry".
Default is \code{IEATools::memo_non_energy_flows}.}

\item{electricity_output}{A string that names the electricity output \code{Flow}. Default is "Electricity output (GWh)".}

\item{electricity_output_flows_prefix}{A string prefix for \code{Flow}s to be aggregated in electricity output. Default is "Electricity output (GWh)-".}

\item{heat_output}{A string that names the heat output \code{Flow}. Default is "Heat output".}

\item{heat_output_flows_prefix}{A string prefix for \code{Flow}s to be aggregated in heat output. Default is "Heat output-".}

\item{.rownum}{The name of a column created (and destroyed) internally by this function.
The \code{.rownum} column temporarily holds row numbers for internal calculations.
The \code{.rownum} column is deleted before returning.}
}
\value{
\code{.iea_df} with additional columns named \code{ledger_side}, \code{flow_aggregation_point}, \code{energy_type}, and \code{unit}.
}
\description{
This function prepares an IEA data frame created by \code{iea_df()} for use in R.
It works on IEA data from the 2018 and 2019 releases
of the IEA's extended energy balances.
}
\details{
This function solves several problems.
The first problem is that metadata in the \code{COUNTRY}, \code{FLOW}, and \code{PRODUCT}
columns of an IEA data table are not unique.
A second problem is that the \code{FLOW} column contains both industries to which energy is flowing \emph{and}
the type of flow that is involved.
(E.g., the suffix "(energy)" means that the flow is an own use by the energy industry.
The "(transf.)" suffix means that a flow is involved in a transformation process
between primary and final energy.

To solve these problems, two additional columns are added: \code{Ledger.side} and \code{Flow.aggregation.point}.
\code{Ledger.side} can be one of "Supply" or "Consumption", corresponding to the top or bottom of the IEA's tables, respectively.
\code{Flow.aggregation.point} indicates the next level of aggregation for these data.
\code{Flow.aggregation.point} can be one of
"Total primary energy supply", "Transformation processes", "Energy industry own use", or "TFC compare"
on the \code{Supply} side of the ledger.
On the \code{Consumption} side of the ledger, \code{Flow.aggregation.point} can be one of
"Industry", "Transport", "Other", "Non-energy use", or "Memo: Non-energy use in industry".
When the \code{Flow.aggregation.point} column is present,
the need for the "(energy)" and "(transf.)" suffixes is eliminated,
so they are deleted.

The third problem this function solves is that energy type and units are not specified in IEA data.
An \code{Energy.type} column is added with the value of \code{energy_type_val}.
(Default is \code{E}, for energy, as opposed to \code{X}, which would be exergy.)
A \code{Unit} column is added with the value of \code{unit_val}.
(Default is "TJ", although any string can be specified in \code{unit_val}.)

Note that this function decides where to divide \code{Supply} from \code{Consumption}.
To do so, it first looks for rows in which \code{Flow} is "Losses".
The last "Losses" row is the last row of the "Supply" side of the ledger.
If "Losses" rows are not found, the function looks for rows in which \code{Flow} is "Total final consumption".
The first "Total final consumption" row is the first row of the \code{Consumption} side of the ledger.
If neither "Losses" nor "Total final consumption" \code{Flow}s are present,
an error is generated.
}
\examples{
iea_df(text = paste0(",,TIME,1960,1961\n",
                     "COUNTRY,FLOW,PRODUCT\n",
                     "World,Production,Hard coal (if no detail),42,43\n",
                     "World,Statistical differences,Hard coal (if no detail),7,8\n",
                     "World,Main activity producer electricity plants,",
                       "Hard coal (if no detail),9,10\n",
                     "World,Non-specified,Hard coal (if no detail),11,12\n",
                     "World,Coal mines,Hard coal (if no detail),13,14\n",
                     "World,Non-specified,Hard coal (if no detail),11,12\n",
                     "World,Losses,Hard coal (if no detail),1,2\n",
                     "World,Iron and steel,Hard coal (if no detail),5,6\n")) \%>\% 
  rename_iea_df_cols() \%>\% 
  augment_iea_df()
}
