% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{augment_iea_df}
\alias{augment_iea_df}
\title{Augment an IEA data frame}
\usage{
augment_iea_df(.iea_df, country = "Country",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point", flow = "Flow",
  energy_type = "Energy.type", energy_type_val = "E",
  units = "Units", units_val = "ktoe", supply = "Supply",
  consumption = "Consumption", tpes = "Total primary energy supply",
  tpes_flows = c("Production", "Imports", "Exports",
  "International marine bunkers", "International aviation bunkers",
  "Stock changes"), tfc_compare = "TFC compare",
  tfc_compare_flows = c("Total primary energy supply", "Transfers",
  "Statistical differences", "Transformation processes",
  "Energy industry own use", "Losses"), losses = "Losses",
  transformation_processes = "Transformation processes",
  tp_flows_suffix = "(transf.)",
  nstp_flows_suffix = "(transformation)",
  eiou = "Energy industry own use", eiou_flows_suffix = "(energy)",
  tfc = "Total final consumption", tfc_flows = c("Industry",
  "Transport", "Other", "Non-energy use"), industry = "Industry",
  industry_flows = c("Iron and steel", "Chemical and petrochemical",
  "Non-ferrous metals", "Non-metallic minerals", "Transport equipment",
  "Machinery", "Mining and quarrying", "Food and tobacco",
  "Paper, pulp and print", "Wood and wood products", "Construction",
  "Textile and leather", "Non-specified (industry)"),
  transport = "Transport",
  transport_flows = c("World aviation bunkers", "Domestic aviation",
  "Road", "Rail", "Pipeline transport", "World marine bunkers",
  "Domestic navigation", "Non-specified (transport)"), other = "Other",
  other_flows = c("Residential", "Commercial and public services",
  "Agriculture/forestry", "Fishing", "Non-specified (other)"),
  non_energy = "Non-energy use", non_energy_prefix = "Non-energy use",
  electricity_output = "Electricity output (GWh)",
  electricity_output_flows_prefix = "Electricity output (GWh)-",
  heat_output = "Heat output",
  heat_output_flows_prefix = "Heat output-", .rownum = ".rownum")
}
\arguments{
\item{.iea_df}{a data frame produced by the \code{\link[=iea_df]{iea_df()}} function}

\item{country}{the name of the country column. Default is "\code{Country}".}

\item{ledger_side}{the name of the ledger side column. Default is "\code{Ledger.side}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column. Default is "\code{Flow.aggregation.point}".}

\item{energy_type}{the name of the energy type column to be added to \code{.iea_df}. Default is "\code{Energy.type}.}

\item{energy_type_val}{the value to put in the \code{energy_type} column. Default is "\code{E}".}

\item{units}{the name of the units column to be added to \code{.iea_df}. Default is "\code{Units}".}

\item{units_val}{the value to put in the \code{units} column. Default is "\code{ktoe}" for kilotons of oil equivalent.}

\item{supply}{the string that identifies supply \code{Ledger.side}. Default is "\code{Supply}".}

\item{consumption}{the string that identifies consumption \code{Ledger.side}. Default is "\code{Consumption}".}

\item{tpes}{the string that identifies total primary energy supply \code{Flow.aggregation.point}. Default is "\code{Total primary energy supply}.}

\item{tpes_flows}{a vector of strings that give flows that are aggregated to \code{Total primary energy supply}.}

\item{tfc_compare}{a string that identifies the \code{TFC compare} flow aggregation point. Default is \code{TFC compare}.}

\item{tfc_compare_flows}{a vector of strings that give \code{Flow}s that are aggregated to \code{TFC compare}.}

\item{losses}{the string that indicates losses in the \code{Flow} column. Default is "\code{Losses}".}

\item{transformation_processes}{the string that indicates transformation processes in the \code{Flow} column. Default is "\code{Transformation processes}".}

\item{tp_flows_suffix}{the suffix for transformation processes in the \code{Flow} column. Default is "\code{(transf.)}".}

\item{nstp_flows_suffix}{the suffix for non-specified transformation processes in the \code{Flow} column. Default is "\code{(transformation)}".}

\item{eiou_flows_suffix}{the suffix for energy industry own use in the \code{Flow} column. Default is "\code{(energy)}".}

\item{tfc}{the string that identifies total final consumption in the \code{Flow} column. Default is "\code{Total final consumption}".}

\item{tfc_flows}{a vector of strings that give total final consumption in the \code{Flow} column.}

\item{industry}{a string that names the industry \code{Flow.aggregation.point}. Default is "\code{Industry}".}

\item{industry_flows}{a vector of strings representing \code{Flow}s to be aggregated in the \code{Industry} \code{Flow.aggregation.point}.}

\item{transport}{a string that names the transport \code{Flow.aggregation.point}. Default is "\code{Transport}".}

\item{transport_flows}{a vector of strings representing \code{Flow}s to be aggregated in the \code{Transport} \code{Flow.aggregation.point}.}

\item{other}{a string that names the other \code{Flow.aggregation.point}. Default is "\code{Other}".}

\item{other_flows}{a vector of strings representing \code{Flow}s to be aggregated in the \code{Other} \code{Flow.aggregation.point}.}

\item{non_energy}{a string that names the non-energy \code{Flow.aggregation.point}. Default is "\code{Non-energy use}".}

\item{non_energy_prefix}{a string prefix for \code{Flow}s to be aggregated in the \code{Non-energy use} \code{Flow.aggregation.point}.}

\item{electricity_output}{a string that names the electricity output \code{Flow}. Default is "\code{Electricity output (GWh)}".}

\item{electricity_output_flows_prefix}{a string prefix for \code{Flow}s to be aggregated in electricity output. Default is "\code{Electricity output (GWh)-}".}

\item{heat_output}{a string that names the heat output \code{Flow}. Default is "\code{Heat output}".}

\item{.rownum}{the name of a column created (and destroyed) internally by this function.
The \code{.rownum} column temporarily holds row numbers for internal calculations.
The \code{.rownum} column is deleted before returning.}

\item{eoiu}{the string that identifies energy industry own use in the \code{Flow} column. Default is "\code{Energy industry own use}".}

\item{heat_output_prefix}{a string prefix for \code{Flow}s to be aggregated in heat output. Default is "\code{Heat output-}".}
}
\value{
\code{.iea_df} with additional columns named \code{ledger_side} and \code{flow_aggregation_point}
}
\description{
This function prepares an IEA data frame created by \code{\link[=iea_df]{iea_df()}} for use in R.
}
\details{
This function solves several problems.
The first problem is that metadata in the \code{COUNTRY}, \code{FLOW}, and \code{PRODUCT}
columns of an IEA data table are not unique.
To solve this problem, two additional columns are added: \code{Ledger.side} and \code{Flow.aggregation.point}.
\code{Ledger.side} can be one of "\code{Supply}" or "\code{Consumption}", corresponding to the top or bottom of the IEA's tables, respectively.
\code{Flow.aggregation.point} indicates the next level of aggregation for these data.
\code{Flow.aggregation.point} can be one of
"\code{Total primary energy supply}", "\code{Transformation processes}", "\code{Energy industry own use}", or "\code{TFC compare}"
on the \code{Supply} side of the ledger.
On the \code{Consumption} side of the ledger, \code{Flow.aggregation.point} can be one of
"\code{Industry}", "\code{Transport}", "\code{Other}", or "\code{Non-energy use}".

The second problem this function solves is that energy type and units are not specified in IEA data.
An \code{Energy.type} column is added with the value of \code{energy_type_val}.
(Default is \code{E}, for energy, as opposed to \code{X}, which would be exergy.)
A \code{Units} column is added with the value of \code{units_val}.
(Default is \code{ktoe}, although any string can be specified in \code{units_val}.)

Note that this function decides where to divide \code{Supply} from \code{Consumption}.
To do so, it first looks for rows in which \code{Flow} is "\code{Losses}".
The last "\code{Losses}" row is the last row of the \code{Supply} side of the ledger.
If "\code{Losses}" rows are not found, the function looks for rows in which \code{Flow} is "\code{Total final consumption}".
The first "\code{Total final consumption}" row is the first row of the \code{Consumption} side of the ledger.
If neither "\code{Losses}" nor "\code{Total final consumption}" \code{Flow}s are present,
an error is generated.
}
\examples{
iea_df(text = paste0(",,TIME,1960,1961\\n",
                       "COUNTRY,FLOW,PRODUCT\\n",
                       "World,Production,Hard coal (if no detail),42,43\\n",
                       "World,Losses,Hard coal (if no detail),1,2")) \%>\% 
  rename_iea_df_cols() \%>\% 
  augment_iea_df()
}
