% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{augment_iea_df}
\alias{augment_iea_df}
\title{Augment an IEA data frame}
\usage{
augment_iea_df(
  .iea_df,
  country = "Country",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  flow = "Flow",
  product = "Product",
  energy_type = "Energy.type",
  energy_type_val = "E",
  method = "Method",
  method_val = "PCM",
  last_stage = "Last.stage",
  last_stage_val = "Final",
  unit = "Unit",
  unit_val = "ktoe",
  supply = "Supply",
  consumption = "Consumption",
  tpes = IEATools::tfc_compare_flows$total_primary_energy_supply,
  tpes_flows = IEATools::tpes_flows,
  tfc_compare = "TFC compare",
  tfc_compare_flows = IEATools::tfc_compare_flows,
  transfers = "Transfers",
  statistical_differences = IEATools::tfc_compare_flows$statistical_differences,
  losses = IEATools::tfc_compare_flows$losses,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  tp_flows_suffix = "(transf.)",
  nstp_flows_suffix = "(transformation)",
  mapep = IEATools::transformation_processes$main_activity_producer_electricity_plants,
  eiou = "Energy industry own use",
  eiou_flows_suffix = "(energy)",
  coal_mines = "Coal mines",
  non_specified = "Non-specified",
  tfc = "Total final consumption",
  tfc_flows = IEATools::tfc_flows,
  industry = "Industry",
  industry_flows = IEATools::industry_flows,
  iron_and_steel = "Iron and steel",
  mining_and_quarrying = "Mining and quarrying",
  transport = "Transport",
  transport_flows = IEATools::transport_flows,
  other = "Other",
  other_flows = IEATools::other_flows,
  non_energy = "Non-energy use",
  non_energy_prefix = "Non-energy use",
  electricity_output = "Electricity output (GWh)",
  electricity_output_flows_prefix = "Electricity output (GWh)-",
  heat_output = "Heat output",
  heat_output_flows_prefix = "Heat output-",
  .rownum = ".rownum"
)
}
\arguments{
\item{.iea_df}{a data frame produced by the \code{\link[=iea_df]{iea_df()}} function}

\item{country}{the name of the country column in \code{.iea_df}. Default is "Country".}

\item{ledger_side}{the name of the ledger side column to be added to \code{.iea_df}. Default is "Ledger.side".}

\item{flow_aggregation_point}{the name of the flow aggregation point column to be added to \code{.iea_df}. Default is "Flow.aggregation.point".}

\item{flow}{the name of the flow column in \code{.iea_df}.  Default is "Flow".}

\item{product}{the name of the product column in \code{.iea_df}.  Default is "Product".}

\item{energy_type}{the name of the energy type column to be added to \code{.iea_df}. Default is "Energy.type.}

\item{energy_type_val}{the value to put in the \code{energy_type} column. Default is "E".}

\item{method}{the name of the method column to be added to \code{.iea_df}. Default is "Method".}

\item{method_val}{the value to put in the \code{method} column. Default is "PCM" (Physical Content Method, which is used by the IEA).}

\item{last_stage}{the name of the last stage column to be added to \code{.iea_df}. Default is "Last.stage".}

\item{last_stage_val}{the value to put in the \code{last_stage} column. Default is "Final" (which is the last stage supplied by the IEA).}

\item{unit}{the name of the unit column to be added to \code{.iea_df}. Default is "Unit".}

\item{unit_val}{the value to put in the \code{unit} column. Default is "\code{ktoe}" for kilotons of oil equivalent.}

\item{supply}{the string that identifies supply \code{Ledger.side}. Default is "Supply".}

\item{consumption}{the string that identifies consumption \code{Ledger.side}. Default is "Consumption".}

\item{tpes}{the string that identifies total primary energy supply \code{Flow.aggregation.point}. Default is "Total primary energy supply".}

\item{tpes_flows}{a vector of strings that give flows that are aggregated to \verb{Total primary energy supply}.}

\item{tfc_compare}{a string that identifies the \verb{TFC compare} flow aggregation point. Default is "TFC compare".}

\item{tfc_compare_flows}{a vector of strings that give \code{Flow}s that are aggregated to \verb{TFC compare}.}

\item{transfers}{= a string that identifies transfers in the flow column. Default is "Transfers".}

\item{statistical_differences}{a string that identifies statistical differences in flow column. Default is "Statistical differences".}

\item{losses}{the string that indicates losses in the \code{Flow} column. Default is "Losses".}

\item{transformation_processes}{the string that indicates transformation processes in the \code{Flow} column. Default is "Transformation processes".}

\item{tp_flows_suffix}{the suffix for transformation processes in the \code{Flow} column. Default is "(transf.)".}

\item{nstp_flows_suffix}{the suffix for non-specified transformation processes in the \code{Flow} column. Default is "(transformation)".}

\item{mapep}{the string that identifies main activity producer electricity plants in the \code{Flow} column. Default is "Main activity producer electricity plants".}

\item{eiou}{the string that identifies energy industry own use in the \code{Flow} column. Default is "Energy industry own use".}

\item{eiou_flows_suffix}{the suffix for energy industry own use in the \code{Flow} column. Default is "(energy)".}

\item{coal_mines}{the string that identifies coal mines in the \code{Flow} column. Default is "Coal mines".}

\item{non_specified}{the string that identifies non-specified flows in the \code{Flow} column. Default is "Non-specified".}

\item{tfc}{the string that identifies total final consumption in the \code{Flow} column. Default is "Total final consumption".}

\item{tfc_flows}{a vector of strings that give total final consumption in the \code{Flow} column.}

\item{industry}{a string that names the industry \code{Flow.aggregation.point}. Default is "Industry".}

\item{industry_flows}{a vector of strings representing \code{Flow}s to be aggregated in the \code{Industry} \code{Flow.aggregation.point}.}

\item{iron_and_steel}{a string that identifies the iron and steel industry. Default is "Iron and steel".}

\item{mining_and_quarrying}{a string that identifies the mining and quarrying industry. Default is "Mining and quarrying".}

\item{transport}{a string that names the transport \code{Flow.aggregation.point}. Default is "Transport".}

\item{transport_flows}{a vector of strings representing \code{Flow}s to be aggregated in the \code{Transport} \code{Flow.aggregation.point}.}

\item{other}{a string that names the other \code{Flow.aggregation.point}. Default is "Other".}

\item{other_flows}{a vector of strings representing \code{Flow}s to be aggregated in the \code{Other} \code{Flow.aggregation.point}.}

\item{non_energy}{a string that names the non-energy \code{Flow.aggregation.point}. Default is "Non-energy use".}

\item{non_energy_prefix}{a string prefix for \code{Flow}s to be aggregated in the \verb{Non-energy use} \code{Flow.aggregation.point}.}

\item{electricity_output}{a string that names the electricity output \code{Flow}. Default is "Electricity output (GWh)".}

\item{electricity_output_flows_prefix}{a string prefix for \code{Flow}s to be aggregated in electricity output. Default is "Electricity output (GWh)-".}

\item{heat_output}{a string that names the heat output \code{Flow}. Default is "Heat output".}

\item{heat_output_flows_prefix}{a string prefix for \code{Flow}s to be aggregated in heat output. Default is "Heat output-".}

\item{.rownum}{the name of a column created (and destroyed) internally by this function.
The \code{.rownum} column temporarily holds row numbers for internal calculations.
The \code{.rownum} column is deleted before returning.}
}
\value{
\code{.iea_df} with additional columns named \code{ledger_side}, \code{flow_aggregation_point}, \code{energy_type}, and \code{unit}.
}
\description{
This function prepares an IEA data frame created by \code{\link[=iea_df]{iea_df()}} for use in R.
It works on IEA data from the 2018 and 2019 releases
of the IEA's extended energy balances.
}
\details{
This function solves several problems.
The first problem is that metadata in the \code{COUNTRY}, \code{FLOW}, and \code{PRODUCT}
columns of an IEA data table are not unique.
A second problem is that the \code{FLOW} column contains both industries to which energy is flowing \emph{and}
the type of flow that is involved.
(E.g., the suffix "\code{(energy)}" means that the flow is an own use by the energy industry.
The "\code{(transf.)}" suffix means that a flow is involved in a transformation process
between primary and final energy.

To solve these problems, two additional columns are added: \code{Ledger.side} and \code{Flow.aggregation.point}.
\code{Ledger.side} can be one of "\code{Supply}" or "\code{Consumption}", corresponding to the top or bottom of the IEA's tables, respectively.
\code{Flow.aggregation.point} indicates the next level of aggregation for these data.
\code{Flow.aggregation.point} can be one of
"\verb{Total primary energy supply}", "\verb{Transformation processes}", "\verb{Energy industry own use}", or "\verb{TFC compare}"
on the \code{Supply} side of the ledger.
On the \code{Consumption} side of the ledger, \code{Flow.aggregation.point} can be one of
"\code{Industry}", "\code{Transport}", "\code{Other}", or "\verb{Non-energy use}".
When the \code{Flow.aggregation.point} column is present,
the need for the "\code{(energy)}" and "\code{(transf.)}" suffixes is eliminated,
so they are deleted.

The third problem this function solves is that energy type and units are not specified in IEA data.
An \code{Energy.type} column is added with the value of \code{energy_type_val}.
(Default is \code{E}, for energy, as opposed to \code{X}, which would be exergy.)
A \code{Unit} column is added with the value of \code{unit_val}.
(Default is \code{ktoe}, although any string can be specified in \code{unit_val}.)

Note that this function decides where to divide \code{Supply} from \code{Consumption}.
To do so, it first looks for rows in which \code{Flow} is "\code{Losses}".
The last "\code{Losses}" row is the last row of the \code{Supply} side of the ledger.
If "\code{Losses}" rows are not found, the function looks for rows in which \code{Flow} is "\verb{Total final consumption}".
The first "\verb{Total final consumption}" row is the first row of the \code{Consumption} side of the ledger.
If neither "\code{Losses}" nor "\verb{Total final consumption}" \code{Flow}s are present,
an error is generated.
}
\examples{
iea_df(text = paste0(",,TIME,1960,1961\n",
                     "COUNTRY,FLOW,PRODUCT\n",
                     "World,Production,Hard coal (if no detail),42,43\n",
                     "World,Statistical differences,Hard coal (if no detail),7,8\n",
                     "World,Main activity producer electricity plants,",
                       "Hard coal (if no detail),9,10\n",
                     "World,Non-specified,Hard coal (if no detail),11,12\n",
                     "World,Coal mines,Hard coal (if no detail),13,14\n",
                     "World,Non-specified,Hard coal (if no detail),11,12\n",
                     "World,Losses,Hard coal (if no detail),1,2\n",
                     "World,Iron and steel,Hard coal (if no detail),5,6\n")) \%>\% 
  rename_iea_df_cols() \%>\% 
  augment_iea_df()
}
