% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biofuels_and_waste_products}
\alias{biofuels_and_waste_products}
\title{Biofuel and waste products}
\format{A string vector with 10 entries}
\usage{
biofuels_and_waste_products
}
\description{
A string vector containing names of products classified by the IEA as biofuel and waste products.
}
\keyword{datasets}
