% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vif.R
\name{calc_vif}
\alias{calc_vif}
\title{Calculate the variance inflation factors}
\usage{
calc_vif(x)
}
\arguments{
\item{x}{matrix or data frame containing only numerical variables.}
}
\value{
The function returns a matrix with 3 columns:
\describe{
  \item{\code{Rsquared}}{}
  \item{\code{Tolerance}}{}
  \item{\code{VIF}}{}
 }
}
\description{
Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy
eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam
voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet
clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
}
\examples{
x <- data.frame(a = 1:20, b = 1:20*2 + rnorm(20) )
calc_vif(x)
}
