% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coal_and_coal_products}
\alias{coal_and_coal_products}
\title{Coal and coal products}
\format{
A string vector with 16 entries.
}
\usage{
coal_and_coal_products
}
\description{
A string vector containing names of products classified by the IEA as coal and coal products.
}
\examples{
coal_and_coal_products
}
\keyword{datasets}
