% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{collapse_to_tidy_psut}
\alias{collapse_to_tidy_psut}
\title{Collapse a tidy data frame of IEA data to a tidy PSUT data frame}
\usage{
collapse_to_tidy_psut(.tidy_iea_df, matname = "matname",
  e_dot = "E.dot", rowname = "rowname", colname = "colname",
  rowtype = "rowtype", coltype = "coltype", matval = "matval",
  grouping_vars = c("Method", "Energy.type", "Last.stage", "Country",
  "Year"))
}
\arguments{
\item{.tidy_iea_df}{a data frame containing \code{matname} and several other columns}

\item{matname}{the name of a column in \code{.tidy_iea_df} containing matrix names. Default is "\code{matname}".}

\item{e_dot}{the name of a column in \code{.tidy_iea_df} containing energy flow rates. Default is "\code{E.dot}".}

\item{rowname}{the name of a column to be added to \code{.tidy_iea_df} for row names. Default is "\code{rowname}".}

\item{colname}{the name of a column to be added to \code{.tidy_iea_df} for column names. Default is "\code{colname}".}

\item{rowtype}{the name of a column to be added to \code{.tidy_iea_df} for row types. Default is "\code{rowtype}".}

\item{coltype}{the name of a column to be added to \code{.tidy_iea_df} for column types. Default is "\code{coltype}".}

\item{matval}{the name of a column to be added to \code{.tidy_iea_df} for matrices. Default is "\code{matval}".}

\item{grouping_vars}{the columns in \code{.tidy_iea_df} by which you want to group matrices.
Default is \code{c("Method", "Last.stage", "Country", "Year", "Energy.type")}.}
}
\value{
\code{.tidy_iea_df} with all values converted to matrices in the \code{matval} column
}
\description{
Call this function after calling \code{\link[=add_row_col_meta]{add_row_col_meta()}}
to collapse \code{.tidy_iea_df} into a tidy PSUT data frame.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  add_psut_matnames() \%>\% 
  add_row_col_meta() \%>\% 
  collapse_to_tidy_psut()
}
