% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{collapse_to_tidy_psut}
\alias{collapse_to_tidy_psut}
\title{Collapse a tidy data frame of IEA data to a tidy PSUT data frame}
\usage{
collapse_to_tidy_psut(.tidy_iea_df, matnames = "matnames",
  e_dot = "E.dot", rownames = "rownames", colnames = "colnames",
  rowtypes = "rowtypes", coltypes = "coltypes",
  ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point", unit = "Unit",
  flow = "Flow", product = "Product", matvals = "matvals")
}
\arguments{
\item{.tidy_iea_df}{a data frame containing \code{matnames} and several other columns}

\item{matnames}{the name of a column in \code{.tidy_iea_df} containing matrix names. Default is "\code{matname}".}

\item{e_dot}{the name of a column in \code{.tidy_iea_df} containing energy flow rates. Default is "\code{E.dot}".}

\item{rownames}{the name of a column to be added to \code{.tidy_iea_df} for row names. Default is "\code{rownames}".}

\item{colnames}{the name of a column to be added to \code{.tidy_iea_df} for column names. Default is "\code{colnames}".}

\item{rowtypes}{the name of a column to be added to \code{.tidy_iea_df} for row types. Default is "\code{rowtypes}".}

\item{coltypes}{the name of a column to be added to \code{.tidy_iea_df} for column types. Default is "\code{coltypes}".}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is "\code{Ledger.side}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{unit}{the name of the unit column in \code{.tidy_iea_df}. Default is "\code{Unit}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{matvals}{the name of a column to be added to \code{.tidy_iea_df} for matrices. Default is "\code{matvals}".}
}
\value{
\code{.tidy_iea_df} with all values converted to matrices in the \code{matvals} column
}
\description{
Call this function after calling \code{\link[=add_row_col_meta]{add_row_col_meta()}}
to collapse \code{.tidy_iea_df} into a tidy PSUT data frame.
}
\details{
This function ensures that all energy flow numbers are positive
before creating the matrices.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  add_psut_matnames() \%>\% 
  add_row_col_meta() \%>\% 
  collapse_to_tidy_psut()
}
