% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{collapse_to_tidy_psut}
\alias{collapse_to_tidy_psut}
\title{Collapse a tidy data frame of IEA data to a tidy PSUT data frame}
\usage{
collapse_to_tidy_psut(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  Epsilon = IEATools::psut_cols$epsilon,
  matnames = IEATools::mat_meta_cols$matnames,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  matvals = IEATools::psut_cols$matvals
)
}
\arguments{
\item{.tidy_iea_df}{a data frame containing \code{matnames} and several other columns}

\item{ledger_side, flow_aggregation_point, flow, product, e_dot, unit}{See \code{IEATools::iea_cols}.}

\item{Epsilon}{Name of the Epsilon matrix. See \code{IEATools::psut_cols}.}

\item{matnames, rownames, colnames, rowtypes, coltypes}{See \code{IEATools::mat_meta_cols}.}

\item{matvals}{See \code{IEATools::psut_cols}.}
}
\value{
\code{.tidy_iea_df} with all values converted to matrices in the \code{matvals} column
}
\description{
Call this function after calling \verb{add_row_col_meta()`` to collapse }.tidy_iea_df` into a tidy PSUT data frame.
}
\details{
This function ensures that all energy flow numbers are positive
before creating the matrices.

Note that the \code{.tidy_iea_df} is ungrouped using the function
\code{dplyr::ungroup()} prior to undergoing any modification.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  add_psut_matnames() \%>\% 
  add_row_col_meta() \%>\% 
  collapse_to_tidy_psut()
}
