% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eiou_flows}
\alias{eiou_flows}
\title{Energy Industry Own Use flows}
\format{
A string list with 17 entries.
}
\usage{
eiou_flows
}
\description{
A string list containing named names of EIOU flows in the IEA World Energy Extended Balances data.
Items in the list provide default values for EIOU flows function arguments
throughout the \code{IEATools} package.
}
\examples{
eiou_flows
}
\keyword{datasets}
