% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{eiou_fu_template}
\alias{eiou_fu_template}
\title{Create an energy industry own use final-to-useful template}
\usage{
eiou_fu_template(.tidy_iea_df, energy_type = "Energy.type",
  energy = "E", last_stage = "Last.stage", final = "Final",
  year = "Year", ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  tpes = "Total primary energy supply", flow = "Flow",
  product = "Product", destination = "Destination",
  quantity = "Quantity", e_dot = "E.dot", e_dot_total = paste0(e_dot,
  ".total"), e_dot_perc = paste0(e_dot, ".perc"),
  maximum_values = "Maximum values", year_for_maximum_values = 0,
  ef_product = "Ef product", allocation_var = "C_",
  n_allocation_rows = 3, machine = "Machine",
  eu_product = "Eu product", .value = ".value")
}
\arguments{
\item{.tidy_iea_df}{}
}
\value{
a data frame containing the EIOU template
}
\description{
Create an energy industry own use final-to-useful template
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  eiou_fu_template()
}
