% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{eiou_fu_template}
\alias{eiou_fu_template}
\title{Create an energy industry own use final-to-useful template}
\usage{
eiou_fu_template(.tidy_iea_df, energy_type = "Energy.type",
  energy = "E", last_stage = "Last.stage", final = "Final",
  year = "Year", ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  tpes = "Total primary energy supply", flow = "Flow",
  destination = "Destination", e_dot = "E.dot",
  e_dot_total = paste0(e_dot, ".total"), e_dot_perc = paste0(e_dot,
  ".perc"), e_dot_max = paste0(e_dot, ".max"),
  e_dot_perc_max = paste0(e_dot_perc, ".max"), allocation_var = "C_",
  n_allocation_rows = 3, val = ".value", grouping_vars = c("Method",
  "Last.stage", "Country", "Year", "Unit"))
}
\arguments{
\item{.tidy_iea_df}{}
}
\value{
a data frame containing the EIOU template
}
\description{
Create an energy industry own use final-to-useful template
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  eiou_fu_template()
}
