% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_types}
\alias{energy_types}
\title{Energy type order}
\format{
A string list with 2 entries.
}
\usage{
energy_types
}
\description{
A string list containing the order for energy types.
}
\examples{
energy_types
}
\keyword{datasets}
