% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{eta_fu_table_completed}
\alias{eta_fu_table_completed}
\title{Tell whether a final-to-useful efficiency table has been completed}
\usage{
eta_fu_table_completed(
  eta_fu_table = NULL,
  machines_that_need_quantities,
  which_quantity = c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u),
  e_dot = IEATools::iea_cols$e_dot,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  unit = IEATools::iea_cols$unit,
  ef_product = IEATools::template_cols$ef_product,
  quantity = IEATools::template_cols$quantity,
  e_dot_perc = IEATools::template_cols$e_dot_perc,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  destination = IEATools::template_cols$destination,
  eta_fu = IEATools::template_cols$eta_fu,
  phi_u = IEATools::template_cols$phi_u,
  .values = IEATools::template_cols$.values
)
}
\arguments{
\item{eta_fu_table}{The final-to-useful efficiency table whose completeness is to be determined.
If \code{NULL} (the default), all rows in \code{fu_allocation_table} that need efficiencies are returned
as the "unallocated_rows" attribute of \code{FALSE}.}

\item{machines_that_need_quantities}{The final-to-useful allocation table
whose final-to-useful machines must be assigned efficiencies (eta_fu) or
exergy-to-energy ratio (phi_u) values.}

\item{which_quantity}{A vector of quantities to be completed in the eta_FU table.
Default is \code{c(IEATools::template_cols$eta_fu, IEATools::template_cols$phi_u)}.}

\item{e_dot, year, method, ledger_side, flow_aggregation_point, unit}{See \code{IEATools::iea_cols}.}

\item{ef_product, quantity, e_dot_perc, e_dot_machine, e_dot_machine_perc, maximum_values, destination, eta_fu, phi_u, .values}{See \code{IEATools::template_cols}.}
}
\value{
A boolean telling whether \code{eta_fu_table} is complete.
If \code{FALSE}, a data frame of fu_allocation_table machines that lack efficiencies
is stored in the "unallocated_rows" attribute of the return value.
Retrieve with \code{attr(done, "unallocated_rows")}
if the result of this function is assigned to the variable \code{done}.
}
\description{
A final-to-useful efficiency table is complete iff all of the machines in a final-to-useful allocation table
have been assigned efficiencies in each year for each country.
}
\details{
This function should really be named \verb{eta_fu_table_completed?}, because it answers a question.
}
\examples{
fu_allocations <- load_fu_allocation_data()
fu_efficiencies <- load_eta_fu_data()
eta_fu_table_completed(fu_efficiencies, fu_allocations)
}
