% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{eta_fu_template}
\alias{eta_fu_template}
\title{Final-to-useful efficiency template}
\usage{
eta_fu_template(.fu_allocations, ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point",
  ef_product = "Ef.product", country = "Country",
  machine = "Machine", eu_product = "Eu.product", eta_fu = "eta.fu",
  phi_u = "phi.u", destination = "Destination",
  quantity = "Quantity", maximum_values = "Maximum.values",
  perc = "[\%]", e_dot = "E.dot", c_ = "C_",
  c_perc = paste(substr(c_, 1, 1), perc),
  eu_order = IEATools::eu_product_order, .year = ".year",
  .value = ".value", e_dot_dest = paste0(e_dot, "_dest"),
  e_dot_machine = paste0(e_dot, "_machine"),
  e_dot_machine_tot = paste0(e_dot_machine, "_tot"),
  e_dot_machine_perc = paste(e_dot_machine, perc),
  e_dot_machine_perc_max = paste0(e_dot_machine, "_max", " ", perc))
}
\arguments{
\item{.fu_allocation_template}{a data frame containing a completed final-to-useful allocation template for final demand.}

\item{fu_allocations_tab_name}{the name of the tab on which the template will be written. Default is "\code{FU Allocations}".}
}
\value{
a data frame containing row-ordered blank template for final-to-useful machine efficiencies.
}
\description{
Using a final-to-useful allocation table,
this function generates a blank template for final-to-useful machine efficiencies.
}
\examples{
load_fu_allocation_data() \%>\% 
  eta_template()
}
