% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{eta_template}
\alias{eta_template}
\title{Final-to-useful efficiency template}
\usage{
eta_template(.fd_fu_allocations, .eiou_fu_allocations)
}
\arguments{
\item{.fd_fu_allocations}{a data frame containing a completed final-to-useful allocation template for final demand.}

\item{.eiou_fu_allocations}{a data frame containing a completed final-to-useful allocation template for energy industry own use.}
}
\value{
a data frame containing a blank template for final-to-useful machine efficiencies.
}
\description{
Using a final-to-useful allocation table,
this function generates a blank template for final-to-useful machine efficiencies.
}
