% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{eta_template}
\alias{eta_template}
\title{Final-to-useful efficiency template}
\usage{
eta_template(.fu_allocation_template)
}
\arguments{
\item{.fu_allocation_template}{a data frame containing a completed final-to-useful allocation template for final demand.}
}
\value{
a data frame containing row-ordered blank template for final-to-useful machine efficiencies.
}
\description{
Using a final-to-useful allocation table,
this function generates a blank template for final-to-useful machine efficiencies.
}
