% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{extract_S_units_from_tidy}
\alias{extract_S_units_from_tidy}
\title{Extract a unit summation matrix from a tidy data frame}
\usage{
extract_S_units_from_tidy(.tidy_iea_df, ledger_side = "Ledger.side",
  flow_aggregation_point = "Flow.aggregation.point", flow = "Flow",
  product = "Product", e_dot = "E.dot", unit = "Unit",
  s_units = "S_units", val = ".val", rowtype = ".rowtype",
  coltype = ".coltype")
}
\arguments{
\item{.tidy_iea_df}{the tidy data frame from which a unit summation \code{S_units} matrix is to be formed.}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{unit}{the name of the unit column in "\code{.tidy_iea_df}". Default is "\code{Unit}".}

\item{s_units}{the name of the unit summation column to be added to \code{.tidy_iea_df}. Default is "\code{S_unit}".}

\item{val}{the name of a temporary column to be created in \code{.tidy_iea_df}. Deafult is "\code{.val}".}

\item{rowtype}{the name of a temporary rowype column created in \code{.tidy_iea_df}. Default is "\code{rowtype}".}

\item{coltype}{the name of a temporary colype column created in \code{.tidy_iea_df}. Default is "\code{coltype}".}

\item{grouping_vars}{a string vector of column names in \code{.tidy_iea_df} by which S_units matrix extraction should be performed.
Default is \code{c("Method", "Last.stage", "Country", "Year", "Energy.type")}.}
}
\value{
a data frame containing grouping variables and a new column of unit summation matrices called \code{s_unit}.
}
\description{
Unit summation matrices have products in rows and units in columns, with
\code{1}s where a product is expressed in the unit and \code{0}s otherwise.
}
\details{
\code{.tidy_iea_df} should be grouped as needed, typically on
\code{Country}, \code{Year}, \code{Energy.type}, \code{Last.stage}, etc., but
\emph{not} on \code{Unit}, \code{Flow} or \code{Product}.
\code{.tidy_iea_df} is typically obtained from \code{\link[=tidy_iea_df]{tidy_iea_df()}}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  extract_S_units_from_tidy()
}
