% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{extract_S_units_from_tidy}
\alias{extract_S_units_from_tidy}
\title{Extract a unit summation matrix from a tidy data frame}
\usage{
extract_S_units_from_tidy(
  .tidy_iea_df,
  matrix_class = c("matrix", "Matrix"),
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  matnames = IEATools::mat_meta_cols$matnames,
  product_type = IEATools::row_col_types$product,
  unit_type = IEATools::row_col_types$unit,
  s_units = IEATools::psut_cols$s_units,
  .val = ".val",
  .rowtype = ".rowtype",
  .coltype = ".coltype"
)
}
\arguments{
\item{.tidy_iea_df}{the tidy data frame from which a unit summation \code{S_units} matrix is to be formed.}

\item{matrix_class}{The type of matrix to be created, one of "matrix" or "Matrix".
Default is "matrix".}

\item{ledger_side, flow_aggregation_point, flow, product, e_dot, unit, matnames}{See \code{IEATools::iea_cols}.}

\item{product_type, unit_type}{See \code{IEATools::row_col_types}.}

\item{s_units}{See \code{IEATools::psut_cols}.}

\item{.val}{the name of a temporary value column to be created in \code{.tidy_iea_df}. Default is ".val".}

\item{.rowtype}{the name of a temporary rowtype column created in \code{.tidy_iea_df}. Default is ".rowtype".}

\item{.coltype}{the name of a temporary coltype column created in \code{.tidy_iea_df}. Default is ".coltype".}
}
\value{
a data frame containing grouping variables and a new column of unit summation matrices called \code{s_unit}.
}
\description{
Unit summation matrices have products in rows and units in columns, with
\code{1}s where a product is expressed in the unit and \code{0}s otherwise.
}
\details{
\code{.tidy_iea_df} should be grouped as needed, typically on
\code{Country}, \code{Year}, \code{EnergyType}, \code{Last.stage}, etc., but
\emph{not} on \code{Unit}, \code{Flow} or \code{Product}.
\code{.tidy_iea_df} is typically obtained from \code{tidy_iea_df()}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  extract_S_units_from_tidy()
}
