% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_splits.R
\name{find_eiou_start}
\alias{find_eiou_start}
\title{Find the point where Energy industry own use starts}
\usage{
find_eiou_start(.ctry_tbl, flow, non_specified, eiou, coal_mines)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{the name of the flow column}

\item{non_specified}{the name of non-specified in the flow column}

\item{eiou}{the name for energy industry own use in the flow column}

\item{coal_mines}{the name for coal mines in the flow column}
}
\value{
a pair of integers representing the rows that straddle
the split between Transformation processes and Energy industry own use
The second integer is the first row of Energy industry own use.
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the Transformation processes and Energy industry own use.
This function simply calls \code{find_transformation_end()}, because
the dividing line between Transformation processes and Energy industry own use
is also the point where Energy industry own use starts.
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between
Transformation processes and Energy industry own use.
}
