% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_supply_consumption_split}
\alias{find_supply_consumption_split}
\title{Find the split point between Supply and Consumption}
\usage{
find_supply_consumption_split(
  .ctry_tbl,
  flow,
  losses,
  iron_and_steel,
  mining_and_quarrying,
  tfc,
  industry
)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{the name of the flow column}

\item{losses}{the name for losses in the flow column}

\item{iron_and_steel}{the name for the iron and steel industry in the flow column}

\item{mining_and_quarrying}{the name for the mining and quarrying industry in the flow column}

\item{tfc}{the name for total final consumption in the flow column}

\item{industry}{the name for industry in the flow column}
}
\value{
a pair of integers representing the rows that straddle the split between
the supply and consumption side of the ledger
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the Supply and Consumption sides of the ledger.
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between supply and consumption.
}
