% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iea_splits.R
\name{find_transformation_start}
\alias{find_transformation_start}
\title{Find the point where Transformation processes begin}
\usage{
find_transformation_start(
  .ctry_tbl,
  flow,
  statistical_differences,
  transformation_processes,
  mapep
)
}
\arguments{
\item{.ctry_tbl}{a country's IEA data frame}

\item{flow}{the name of the flow column}

\item{statistical_differences}{the name for statistical difference in the flow column}

\item{transformation_processes}{the name for transformation processes in the flow column}

\item{mapep}{the name for main activity producer electricity plants in the flow column}
}
\value{
a pair of integers representing the rows that straddle the split between
the TFC compare and Transformation processes flows.
The second integer is the first row of Transformation processes.
}
\description{
Given a country's IEA extended energy balance data frame,
find the row numbers that represent the transition between
the first TFC Compare section and Transformation processes.
}
\details{
Arguments should be supplied by the calling function.

An error is given if this function fails to find the location of the split between
TFC compare and Transformation processes.
}
