% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_GHA_industry_electricity}
\alias{fix_GHA_industry_electricity}
\title{Improve IEA Ghana Industry Electricity data from 1974 through 2017}
\usage{
fix_GHA_industry_electricity(
  .tidy_iea_df,
  col_names = IEATools::iea_cols,
  country = col_names$country,
  year = col_names$year,
  e_dot = col_names$e_dot
)
}
\arguments{
\item{.tidy_iea_df}{a tidy IEA data frame produced by \code{\link[=load_tidy_iea_df]{load_tidy_iea_df()}}}

\item{col_names}{a list of column names in IEA data frames. Default is \code{IEATools::iea_cols}.}

\item{country}{the name of the country column in \code{.tidy_iea_df}. Default is \code{col_names$country}.}

\item{year}{the name of the year column in \code{.tidy_iea_df}. Default is \code{col_names$year}.}

\item{e_dot}{the name of the energy flow rate column in \code{.tidy_iea_df}. Default is \code{colnames$e_dot}.}
}
\value{
\code{.tidy_iea_df} with improved Ghana Industry Electricity, if warranted
}
\description{
Ghana's Industry Electricity data are specified for
Mining and quarrying,
Non-ferrous metals, and
Textiles and leather
for the year 1971--1973 only.
However, data are available from
the Ghana Grid Company (GridCo) and
the Volta River Authority
that fill most remaining years to 2017.
The details to fill the extra years are in the object \code{Fixed_GHA_Industry_Electricity}.
This function applies the fixes to years that are available in \code{.tidy_iea_data}.
}
\details{
See the Supplemental information to
M. K. Heun and P. E. Brockway.
Meeting 2030 primary energy and economic growth goals: Mission impossible?
Applied Energy, 251(112697):1–24, May 2019
for additional details.

If \code{.tidy_iea_df} does not contain data from Ghana for the years in question,
no fixing is performed, and \code{.tidy_iea_df} is returned unmodified.

Also see the file named "GHA-IndustryElectricity.xlsx" for the actual calculations.
}
\examples{
library(dplyr)
example_tidy_iea_df <- load_tidy_iea_df() \%>\% 
  filter(Country == "GHA")
example_tidy_iea_df
fixed <- example_tidy_iea_df \%>\% 
  fix_GHA_industry_electricity()
# Compare changed values
example_tidy_iea_df \%>\% 
  filter(Flow \%in\% c("Mining and quarrying",
                     "Non-ferrous metals",
                     "Textile and leather", 
                     "Industry not elsewhere specified"), 
         Product == "Electricity") \%>\% 
  select("Year", "Flow", "E.dot", "Unit")
fixed \%>\% 
  filter(Flow \%in\% c("Mining and quarrying",
                     "Non-ferrous metals",
                     "Textile and leather", 
                     "Industry not elsewhere specified"), 
         Product == "Electricity") \%>\% 
  select("Year", "Flow", "E.dot", "Unit")
# Show that new data are still in balance
example_tidy_iea_df \%>\% 
  filter(Year == 2000, 
         Product == "Electricity", 
         Flow == "Industry not elsewhere specified") \%>\% 
  select(E.dot) \%>\% 
  as.numeric()
fixed \%>\% 
  filter(Year == 2000, 
         Product == "Electricity", 
         Flow \%in\% c("Mining and quarrying",
                     "Non-ferrous metals",
                     "Textile and leather", 
                     "Industry not elsewhere specified")) \%>\% 
  select(E.dot) \%>\% 
  sum()
}
