% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_GHA_psb}
\alias{fix_GHA_psb}
\title{Improve Ghana Primary solid biofuels data from 1991 through 1999}
\usage{
fix_GHA_psb(
  .tidy_iea_df,
  col_names = IEATools::iea_cols,
  country = col_names$country,
  method = col_names$method,
  energy_type = col_names$energy_type,
  last_stage = col_names$last_stage,
  ledger_side = col_names$ledger_side,
  flow_aggregation_point = col_names$flow_aggregation_point,
  flow = col_names$flow,
  product = col_names$product,
  unit = col_names$unit,
  year = col_names$year
)
}
\arguments{
\item{.tidy_iea_df}{a tidy IEA data frame produced by \code{\link[=load_tidy_iea_df]{load_tidy_iea_df()}}}

\item{col_names}{a list of column names in IEA data frames. Default is \code{IEATools::iea_cols}.}

\item{country}{the name of the country column in \code{.tidy_iea_df}. Default is \code{col_names$country}.}

\item{method}{the name of the method column in \code{.tidy_iea_df}. Default is \code{col_names$method}.}

\item{energy_type}{the name of the energy type column in \code{.tidy_iea_df}. Default is \code{col_names$energy_type}.}

\item{last_stage}{the name of the last stage column in \code{.tidy_iea_df}. Default is \code{col_names$last_stage}.}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is \code{col_names$ledger_side}.}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is \code{col_names$flow_aggregation_point}.}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is \code{col_names$flow}.}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is \code{col_names$product}.}

\item{unit}{the name of the unit column in \code{.tidy_iea_df}. Default is \code{col_names$unit}.}

\item{year}{the name of the year column in \code{.tidy_iea_df}. Default is \code{col_names$year}.}
}
\value{
\code{.tidy_iea_df} with smoothed Ghana Primary solid biofuels data
}
\description{
Ghana's Primary solid biofuels data show a very large and dramatic decline from 1999 to 2000.
This decline is due to new survey data being used for the 2000 data.
When we look at the PSB data on a per-capita basis, it is clear that
a near-constant PSB/capita value was used to extrapolate per-capita usage in the late 1990s.
When new survey data became available for the 2000 reporting year,
the per-capita consumption of PSB obviously changed.
Our approach to this problem is to smooth out the really big peak in PSB consumption
by reducing the per-capita consumption of PSB, starting in 1991.
}
\details{
See the Supplemental information to
M. K. Heun and P. E. Brockway.
Meeting 2030 primary energy and economic growth goals: Mission impossible?
Applied Energy, 251(112697):1–24, May 2019
for additional details.

Also see the file named "GHA-PSB.xlsx" for the actual calculations.
}
\examples{
library(dplyr)
# Build an example tidy IEA data frame in which Ghana's Primary solid biofuels can be fixed.
example_tidy_iea_df <- load_tidy_iea_df() \%>\% 
  filter(Country == "GHA") \%>\% 
  filter(Product == "Primary solid biofuels") \%>\% 
  # The example data frame has PSB for Ghana for 1971 and 2000.  
  # Let's pretend that 1971 is 1991 and 2000 is 1992.
  mutate(
    Year := dplyr::case_when(
      Year == 1971 ~ 1991,
      Year == 2000 ~ 1992
    )
  ) 
example_tidy_iea_df
fixed <- example_tidy_iea_df \%>\% 
  fix_GHA_psb()
# Compare production of Primary solid biofuels in 1991
example_tidy_iea_df \%>\% 
  filter(Year == 1991, Flow == "Production") \%>\% 
  select("E.dot", "Unit")
fixed \%>\% 
  filter(Year == 1991, Flow == "Production") \%>\% 
  select("E.dot", "Unit")
}
