% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fixed_OAMR_cpp}
\alias{Fixed_OAMR_cpp}
\title{Fixed Other non-OECD Americas Charcoal production 1971--2010}
\format{
A data frame with 11 columns.
}
\usage{
Fixed_OAMR_cpp
}
\description{
Other Non-OECD Americas has several years (1971--2010)
in which Charcoal is produced
but no Primary solid biofuels are consumed to
create the Charcoal.
This object contains (presumably) correct data.
In particular, Charcoal production plants
now consume Primary solid biofuels in all years, and
Primary solid biofuels production is boosted accordingly.
The efficiency of Charcoal production plants in 2011
was used to create the filled data.
The function \code{fix_OAMR_cpp()} makes use of these data.
}
\examples{
Fixed_OAMR_cpp
}
\keyword{datasets}
