% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{fu_allocation_template}
\alias{fu_allocation_template}
\title{Create a template for analysis of final-to-useful transformation processes}
\usage{
fu_allocation_template(
  .tidy_iea_df,
  energy_type = "EnergyType",
  energy = "E",
  last_stage = "LastStage",
  final = "Final",
  year = "Year",
  ledger_side = "LedgerSide",
  consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point",
  eiou = "Energy industry own use",
  non_energy_use = "Non-energy use",
  tfc = "Total final consumption",
  tpes = "Total primary energy supply",
  flow = "Flow",
  product = "Product",
  destination = "Destination",
  quantity = "Quantity",
  e_dot = "E.dot",
  e_dot_total = paste0(e_dot, ".total"),
  perc_unit_string = "[\%]",
  e_dot_perc = paste(e_dot, perc_unit_string),
  maximum_values = "Maximum.values",
  year_for_maximum_values = 0,
  ef_product = "Ef.product",
  allocation_var = "C_",
  n_allocation_rows = 4,
  machine = "Machine",
  eu_product = "Eu.product",
  arrange = TRUE,
  .value = ".value"
)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{energy_type}{the name of the energy type column. Default is "EnergyType".}

\item{energy}{the string identifier for energy (as opposed to exergy) in the \code{energy_type} column. Default is "\code{E}".}

\item{last_stage}{the name of the last stage column. Default is "LastStage".}

\item{final}{the string identifier for final energy (as \code{Last.stage}). Default is "Final".}

\item{year}{the name of the year column. Default is "Year".}

\item{ledger_side}{the name of the ledger side column. Default is "LedgerSide".}

\item{consumption}{the string identifier for the consumption side of the ledger. Default is "Consumption".}

\item{flow_aggregation_point}{the name of the flow aggregation point column. Default is "Flow.aggregation.point".}

\item{eiou}{the string identifier for energy industry own use in \code{flow_aggregation_point}. Default is "Energy industry own use".}

\item{non_energy_use}{string identifier for non-energy use in \code{flow_aggregation_point}. Default is "Non-energy use".}

\item{tfc}{the string identifier for total final consumption. Default is "Total final consumption".}

\item{tpes}{the string identifier for total primary energy supply. Default is "Total primary energy supply".}

\item{flow}{the name of the flow column. Default is "Flow".}

\item{product}{the name of the product column. Default is "Product".}

\item{destination}{the name for the destination column. Default is "Destination".}

\item{quantity}{the name of the quantity column. Default is "Quantity".}

\item{e_dot}{the name of the energy flow rate column. Default is "E.dot".}

\item{e_dot_total}{the string identifier for total energy. Default is "E.dot.total".}

\item{perc_unit_string}{the string used to indicate percentages. Default is "\verb{[\%]}".}

\item{e_dot_perc}{the string identifier for energy percentage. Default is "E.dot.perc".}

\item{maximum_values}{the name for the maximum energy values column. Default is "Maximum values".}

\item{year_for_maximum_values}{an integer for the first year (in which maximum values will be stored before renaming the column to \code{maximum_values}).
Default is \code{0}.}

\item{ef_product}{the name of the final energy carrier column. Default is "Ef product".}

\item{allocation_var}{the string identifier for the allocation percentage column. Default is "C_".}

\item{n_allocation_rows}{an integer stating how many allocation rows are desired. Default is \code{3}.}

\item{machine}{the name of the column of final-to-useful transformation process machines. Default is "Machine".}

\item{eu_product}{the name of the useful energy carrier column. Default is "Eu product".}

\item{arrange}{a boolean telling whether to arranged the rows and columns
(using \code{arrange_iea_fu_allocation_template()}) before returning.
Default is \code{TRUE}.}

\item{.value}{the name of a temporary value column added to \code{.tidy_iea_df}.
A \code{.value} column must not be present in \code{.tidy_iea_df}. Default is ".value".}
}
\value{
a data frame containing the EIOU template
}
\description{
This function creates a blank template for final-to-useful energy transformation process analysis.
The template is a data frame derived from IEA extended energy balance data (\code{.tidy_iea_df}),
which gives both energy industry own use and final demand for final energy carriers.
From the IEA's final energy information, templates are created for conversion to
useful energy carriers.
Final energy consumed by each final demand sector or transformation process machine
is allocated to final-to-useful machines and converted to useful energy at some efficiency.
The allocation fractions are to be supplied by the analyst by filling blanks in
the template.
}
\details{
Non-energy use is removed from \code{.tidy_iea_df} before creating the template.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  fu_allocation_template()
}
