% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iea_cols}
\alias{iea_cols}
\title{IEA data frame columns}
\format{
A string list with 10 entries
}
\usage{
iea_cols
}
\description{
A string list containing named names of columns in IEA data frames.
These data frames can be
tidy (one row per data point) or
wide (years spread to the right).
Names of the items in the list are the typical function arguments
for column names.
}
\examples{
iea_cols
}
\keyword{datasets}
