% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{iea_df}
\alias{iea_df}
\title{Load IEA data from csv file}
\usage{
iea_df(iea_file = NULL, text = NULL,
  expected_1st_line_start = ",,TIME",
  expected_2nd_line_start = "COUNTRY,FLOW,PRODUCT")
}
\arguments{
\item{iea_file}{a string containing the path to a .csv file of extended energy balances from the IEA}

\item{text}{a character string that can be parsed as IEA extended energy balances. 
(This argument is useful for testing.)}

\item{expected_1st_line_start}{the expected start of the first line of \code{iea_file}. Default is "\code{,,TIME}".}

\item{expected_2nd_line}{the expected second line of \code{iea_file}. Default is "\code{COUNTRY,FLOW,PRODUCT}".
Note that no additional commas are expected at the end of this line.}
}
\value{
a data frame containing the as-read IEA data
}
\description{
This function reads an IEA extended energy balances file and
converts it to a data frame with appropriately-labelled columns.
One of \code{iea_file} or \code{text} must be specified, but not both.
The first line of \code{iea_file} or \code{text} 
is expected to start with \code{expected_start_1st_line}, and
the second line is expected to be \code{expected_2nd_line}.
If those conditions are not met, execution is halted.
}
\details{
This function is expected to work even as more years are added
as columns at the right of \code{iea_file}, 
because column names are constructed from the first line of \code{iea_file} (which contains years).

The data frame returned from this function is not ready to be used in R, 
because rows are not unique, and 
several empty cells are filled with ".." or "x".
To further prepre the data frame for use, call \code{augment_iea_data()}
with the output of this function.
}
\examples{
iea_df(text = ",,TIME,1960,1961\\nCOUNTRY,FLOW,PRODUCT\\nWorld,Production,Hard coal,42,43")
}
