% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interface_industries}
\alias{interface_industries}
\title{Interface industries}
\format{
A string vector with 5 entries
}
\usage{
interface_industries
}
\description{
A string vector containing names of \code{Flow}s that interface with the world outside of the economy.
}
\examples{
interface_industries
}
\keyword{datasets}
