% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_stages}
\alias{last_stages}
\title{Last stage order}
\format{
A string vector
}
\usage{
last_stages
}
\description{
A string vector containing options for the last stage of energy conversion chain analysis.
}
\examples{
last_stages
}
\keyword{datasets}
