% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{load_fu_allocation_data}
\alias{load_fu_allocation_data}
\title{Load final-to-useful allocation data}
\usage{
load_fu_allocation_data(
  path = sample_fu_allocation_table_path(),
  fu_allocations_tab_name = IEATools::fu_analysis_file_info$fu_allocation_tab_name,
  country = IEATools::iea_cols$country,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  ef_product = IEATools::template_cols$ef_product,
  machine = IEATools::template_cols$machine,
  eu_product = IEATools::template_cols$eu_product,
  destination = IEATools::template_cols$destination,
  non_energy_machine = "Non-energy"
)
}
\arguments{
\item{path}{The path from which final-to-useful allocation data will be loaded. Default is the path to allocation data supplied with this package.}

\item{fu_allocations_tab_name}{See \code{IEATools::fu_analysis_file_info}.}

\item{country, flow_aggregation_point}{See IEATools::iea_cols}

\item{ef_product, machine, eu_product, destination}{See \code{IEATools::template_cols}.}

\item{non_energy_machine}{The string that identifies a Non-energy machine. Default is "Non-energy".}
}
\value{
the \code{fu_allocations_tab_name} tab in \code{path} as a data frame.
}
\description{
When performing extending an energy conversion chain from useful energy to final energy,
allocations of final energy consumption to useful energy categories are defined by the analyst.
The Excel file at \code{path} contains those allocations.
}
\details{
A final-to-useful allocation template can be
generated using \code{fu_allocation_template()} and \code{write_fu_allocation_template()}.

A filled example can be loaded with the default value of \code{path}.

Note that any machine named \code{non_energy_machine} is required to have
identical values for \code{Ef.product} and \code{Eu.product}.
Violations of this requirement cause errors to be thrown.
}
\examples{
# Loads final-to-useful allocation data supplied with the package
load_fu_allocation_data()
}
