% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_to_x.R
\name{load_phi_constants_table}
\alias{load_phi_constants_table}
\title{Read a table of constant phi (exergy-to-energy ratio) values from a file.}
\usage{
load_phi_constants_table(
  phi_constants_table_path = sample_phi_constants_path(),
  phi_constants_tab_name = IEATools::phi_constants_names$phi_constants_tab_name,
  product_colname = IEATools::phi_constants_names$product_colname,
  phi_colname = IEATools::phi_constants_names$phi_colname
)
}
\arguments{
\item{phi_constants_table_path}{The path to the Excel file containing a table of constant phi values.
Default is the value of \code{sample_phi_constants_path()}.}

\item{phi_constants_tab_name, product_colname, phi_colname}{See \code{IEATools::phi_constants_names}.}
}
\value{

}
\description{
Read a table of constant phi (exergy-to-energy ratio) values from a file.
}
\examples{
load_phi_constants_table()
}
