% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{load_tidy_iea_df}
\alias{load_tidy_iea_df}
\title{Load IEA extended energy balance data into tidy format}
\usage{
load_tidy_iea_df(.iea_file = sample_iea_data_path(), remove_zeroes = TRUE)
}
\arguments{
\item{.iea_file}{the path of the file to be loaded. Default loads example data bundled with the package via \code{\link[=sample_iea_data_path]{sample_iea_data_path()}}.}

\item{remove_zeroes}{a logical indicating whether data points with the value \code{0} are to be removed from the output. (Default is \code{TRUE}.)}
}
\value{
a tidy, augmented data frame of IEA extended energy balance data.
}
\description{
Loads an IEA extended energy balance data file in \code{.csv} format from disk and converts to a tidy format.
This function bundles several others:
\enumerate{
\item \code{\link[=iea_df]{iea_df()}},
\item \code{\link[=rename_iea_df_cols]{rename_iea_df_cols()}},
\item \code{\link[=use_iso_countries]{use_iso_countries()}},
\item \code{\link[=remove_agg_memo_flows]{remove_agg_memo_flows()}},
\item \code{\link[=augment_iea_df]{augment_iea_df()}}, and
\item \code{\link[=tidy_iea_df]{tidy_iea_df()}}.
}
}
\details{
Each bundled function is called in turn using default arguments.
See examples for two ways to achieve the same result.
}
\examples{
# Check the file first
iea_file_OK(sample_iea_data_path())
# Take a simple approach
simple <- load_tidy_iea_df()
# Take the complicated approach
complicated <- sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  clean_iea_whitespace() \%>\% 
  remove_agg_memo_flows() \%>\% 
  use_iso_countries() \%>\% 
  augment_iea_df() \%>\% 
  tidy_iea_df()
# simple and complicated should be exactly the same
all(simple == complicated)
}
