% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{load_tidy_iea_df}
\alias{load_tidy_iea_df}
\title{Load IEA extended energy balance data into tidy format}
\usage{
load_tidy_iea_df(file_path = file.path("extdata",
  "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") \%>\%
  system.file(package = "IEATools"))
}
\arguments{
\item{file_path}{the path of the file to be loaded. Default loads example data bundled with the package.}
}
\value{
a tidy, augmented data frame of IEA extended energy balance data.
}
\description{
Loads an IEA extended energy balance data file in \code{.csv} format from disk and converts to a tidy format.
This function bundles several others:
\code{\link[=iea_df]{iea_df()}},
\code{\link[=rename_iea_df_cols]{rename_iea_df_cols()}},
\code{\link[=use_iso_countries]{use_iso_countries()}},
\code{\link[=remove_agg_memo_flows]{remove_agg_memo_flows()}},
\code{\link[=augment_iea_df]{augment_iea_df()}}, and
\code{\link[=tidy_iea_df]{tidy_iea_df()}}.
Each is called in turn using default arguments.
See examples for two ways to achieve the same result.
}
\examples{
simple <- load_tidy_iea_df()
complicated <- file.path("extdata", "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") \%>\% 
  system.file(package = "IEATools") \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  remove_agg_memo_flows() \%>\% 
  use_iso_countries() \%>\% 
  augment_iea_df() \%>\% 
  tidy_iea_df()
all(simple == complicated)
}
