% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_act_plants}
\alias{main_act_plants}
\title{Main activity producer plants}
\format{
A string vector with 6 entries.
}
\usage{
main_act_plants
}
\description{
A string vector containing names of Main activity producer plants.
}
\examples{
main_act_plants
}
\keyword{datasets}
