% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_to_useful.R
\name{make_C_mats}
\alias{make_C_mats}
\title{Create an allocation matrix (\code{C}) from an allocation table}
\usage{
make_C_mats(
  .fu_allocations,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  e_dot = IEATools::iea_cols$e_dot,
  year = IEATools::iea_cols$year,
  supply = IEATools::ledger_sides$supply,
  consumption = IEATools::ledger_sides$consumption,
  quantity = IEATools::template_cols$quantity,
  machine = IEATools::template_cols$machine,
  ef_product = IEATools::template_cols$ef_product,
  eu_product = IEATools::template_cols$eu_product,
  destination = IEATools::template_cols$destination,
  e_dot_perc = IEATools::template_cols$e_dot_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  matnames = IEATools::mat_meta_cols$matnames,
  matvals = IEATools::mat_meta_cols$matvals,
  rownames = IEATools::mat_meta_cols$rownames,
  colnames = IEATools::mat_meta_cols$colnames,
  rowtypes = IEATools::mat_meta_cols$rowtypes,
  coltypes = IEATools::mat_meta_cols$coltypes,
  product = IEATools::row_col_types$product,
  industry = IEATools::row_col_types$industry,
  sep = " -> ",
  C_eiou = "C_EIOU",
  C_Y = "C_Y",
  .should_be_1_vector = ".should_be_1_vector",
  .is_1_vector = ".is_1_vector"
)
}
\arguments{
\item{.fu_allocations}{}
}
\value{
a tidy data frame with metadata columns (and year) along with a \code{C} column.
}
\description{
To create an allocation matrix (\code{C}),
rownames are taken from the \code{Ef.product} and \code{Destination} columns, and
colnames are taken from the \code{Machine} and \code{Eu.product} columns.
}
\details{
The \code{.fu_alocations} data frame is first made tidy by gathering all years.
}
