% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manufacturing_flows}
\alias{manufacturing_flows}
\title{Manufacturing flows}
\format{A string vector with 15 entries}
\usage{
manufacturing_flows
}
\description{
A string vector containing names of \code{Manufacturing} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
manufacturing_flows
}
\keyword{datasets}
