% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods}
\alias{methods}
\title{Method order}
\format{
A string list with 3 entries
\describe{
\item{PCM}{Physical content method (used by the IEA).}
\item{RCM}{Resource content method.}
\item{PSM}{Partial substitution method (used by EIA and BP)..}
}
}
\usage{
methods
}
\description{
A string list containing types of methods for quantifying the primary equivalent of renewable electricity.
}
\examples{
methods
}
\keyword{datasets}
