% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_energy_flows}
\alias{non_energy_flows}
\title{Non-energy flows}
\format{
A string vector with 3 entries.
\describe{
\item{non_energy_use_industry_transformation_energy}{The string identifier for the Non-energy use industry/transformation/energy sector.}
\item{non_energy_use_in_transport}{The string identifier for the Non-energy use in transport sector.}
\item{non_energy_use_in_other}{The string identifier for the Non energy use in other sector.}
}
}
\usage{
non_energy_flows
}
\description{
A string vector containing names of \code{Non-energy} \code{Flow}s in the IEA extended energy balances database.
}
\examples{
non_energy_flows
}
\keyword{datasets}
