% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_specified_flows}
\alias{non_specified_flows}
\title{IEA non-specified flows}
\format{
A string list with 6 entries.
\describe{
\item{non_specified_transformation}{The string identifying non-specified transformation process flows.}
\item{non_specified_energy}{The string identifying non-specified energy industry own use.}
\item{non_specified_transport}{The string identifying non-specified transport.}
\item{non_specified_industry}{The string identifying non-specified industry for 2018 and earlier versions of the IEA's extended energy balance data.}
\item{industry_not_elsewhere_specified}{The string identifying non-specified industry flows for 2019 and later versions of the IEA's extended energy balance data.}
\item{non_specified}{The string identifying generic non-specified flows.}
}
}
\usage{
non_specified_flows
}
\description{
A string vector containing non-specified flows in IEA order.
}
\examples{
non_specified_flows
}
\keyword{datasets}
