% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{other_flows}
\alias{other_flows}
\title{Other flows}
\format{A string vector with 6 entries}
\usage{
other_flows
}
\description{
A string vector containing names of \code{Other} \code{Flow}s in the IEA extended energy balances database.
}
\keyword{datasets}
