% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{override_iso_codes_df}
\alias{override_iso_codes_df}
\title{Country code overrides}
\format{
A data frame with four rows and two columns.
\describe{
\item{pfu_code}{The column of 3-letter country codes that to override those found in \code{countrycode::codelist}.}
\item{iea_name}{The column containing IEA country names..}
}
}
\usage{
override_iso_codes_df
}
\description{
A data frame containing 3-letter country codes and IEA country names.
These code-country pairs are used as a default set of overrides (relative to \code{countrycode::codelist})
in the function \code{use_iso_countries()}.
}
\examples{
override_iso_codes_df
}
\keyword{datasets}
