% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{prep_psut}
\alias{prep_psut}
\title{Prepare for PSUT analysis}
\usage{
prep_psut(.tidy_iea_df)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}
}
\value{
an enhanced and corrected version of \code{.tidy_iea_df}
that is ready for physical supply-use table (PSUT) analysis.
}
\description{
This is a convenience function.
This function bundles several others:
}
\details{
\enumerate{
\item \code{\link[=specify_primary_production]{specify_primary_production()}}
\item \code{\link[=production_to_resources]{production_to_resources()}}
\item \code{\link[=specify_interface_industries]{specify_interface_industries()}}
}

Each bundled function is called in turn using default arguments.
See examples for two ways to achieve the same result.
}
\examples{
# Simple
load_tidy_iea_df() \%>\% 
  prep_psut()
# Complicated
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  production_to_resources()
}
