% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{prep_psut}
\alias{prep_psut}
\title{Prepare for PSUT analysis}
\usage{
prep_psut(.tidy_iea_df)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame that has been specified with \code{\link[=specify_all]{specify_all()}}.}
}
\value{
a tidy PSUT data frame
}
\description{
Converts a tidy IEA data frame into a PSUT data frame
by collapsing the IEA data into PSUT matrices (R, U, V, and Y).
}
\examples{
Simple <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  prep_psut()
Complicated <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  add_psut_matnames() \%>\% 
  add_row_col_meta() \%>\% 
  collapse_to_tidy_psut()
all(Simple == Complicated)
}
