% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psut.R
\name{prep_psut}
\alias{prep_psut}
\title{Prepare for PSUT analysis}
\usage{
prep_psut(.tidy_iea_df, ledger_side = "Ledger.side", supply = "Supply",
  consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point", flow = "Flow",
  product = "Product", unit = "Unit", e_dot = "E.dot",
  matnames = "matnames", matvals = "matvals", rownames = "rownames",
  colnames = "colnames", rowtypes = "rowtypes",
  coltypes = "coltypes")
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame that has been specified with \code{\link[=specify_all]{specify_all()}}.}

\item{ledger_side}{the name of the ledger side column. Default is "\code{Ledger.side}".}

\item{supply}{the string identifying the supply side of the ledger. Default is "\code{Supply}".}

\item{consumption}{the string identifying the consumption side of the ledger. Default is "\code{Consumption}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{flow}{the name of the flow column. Default is "\code{Flow}".}

\item{product}{the name of the product column. Default is "\code{Product}".}

\item{unit}{the name of the unit column. Default is "\code{Unit}".}

\item{e_dot}{the name of the energy rate column. Default is "E.dot".}

\item{matnames}{the name of the matrix names column added by this function. Default is "\code{matnames}".}

\item{matvals}{the name of the matrix value column added by this function. Default is "\code{matvals}".}

\item{rownames}{the name of the rownames column passed to \code{collapse_to_tidy_psut()}. Default is "\code{rownames}".}

\item{colnames}{the name of the colnames column passed to \code{collapse_to_tidy_psut()}. Default is "\code{colnames}".}

\item{rowtypes}{the name of the rowtypes column passed to \code{collapse_to_tidy_psut()}. Default is "\code{rowtypes}".}

\item{coltypes}{the name of the coltypes column passed to \code{collapse_to_tidy_psut()}. Default is "\code{coltypes}".}
}
\value{
a tidy PSUT data frame
}
\description{
Converts a tidy IEA data frame into a PSUT data frame
by collapsing the IEA data into PSUT matrices (R, U, V, and Y).
}
\details{
This function bundles several others:
\enumerate{
\item \code{\link[=add_psut_matnames]{add_psut_matnames()}}
\item \code{\link[=add_row_col_meta]{add_row_col_meta()}}
\item \code{\link[=collapse_to_tidy_psut]{collapse_to_tidy_psut()}}
}

Furthermore, it extracts \code{S_units} matrices using \code{\link[=extract_S_units_from_tidy]{extract_S_units_from_tidy()}}
and adds those matrices to the data frame.
}
\examples{
library(dplyr)
library(tidyr)
Simple <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  prep_psut() \%>\% 
  rename(matval_simple = matvals)
S_units <- load_tidy_iea_df() \%>\% 
  extract_S_units_from_tidy()
Complicated <- load_tidy_iea_df() \%>\% 
  specify_all() \%>\% 
  add_psut_matnames() \%>\% 
  add_row_col_meta() \%>\% 
  collapse_to_tidy_psut() \%>\% 
  spread(key = matnames, value = matvals) \%>\% 
  full_join(S_units, by = c("Method", "Energy.type", "Last.stage", 
                            "Country", "Year")) \%>\% 
  gather(key = matnames, value = matvals, R, U_EIOU, U_excl_EIOU, 
                                        V, Y, S_units) \%>\% 
  rename(matval_complicated = matvals)
# Simple and Complicated are same.
full_join(Simple, Complicated, by = c("Method", "Energy.type", 
                                      "Last.stage", "Country", 
                                      "Year", "matnames")) \%>\% 
  dplyr::mutate(
    same = matsbyname::equal_byname(matval_simple, matval_complicated)
  ) \%>\% 
  magrittr::extract2("same") \%>\% 
  as.logical() \%>\% 
  all()
}
