% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_peat_products}
\alias{primary_peat_products}
\title{Primary peat products}
\format{A string vector with 1 entry}
\usage{
primary_peat_products
}
\description{
A string vector containing names of products classified by the IEA as primary peat products.
}
\keyword{datasets}
