% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prod_tp_eiou_energy_carriers}
\alias{prod_tp_eiou_energy_carriers}
\title{Identify all energy types supplied or consumed by Production, Transformation processes, or Energy industry own use}
\usage{
prod_tp_eiou_energy_carriers(
  file_path = file.path("extdata", "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv")
    \%>\% system.file(package = "IEATools"),
  iea_df = IEATools::load_tidy_iea_df(file_path),
  side = c("Consumption", "Supply"),
  flow_aggregation_point = "Flow.aggregation.point",
  production = "Production",
  transformation_processes = "Transformation processes",
  eiou = "Energy industry own use",
  stage = c(production, transformation_processes, eiou),
  e_dot = "E.dot",
  flow = "Flow",
  product = "Product"
)
}
\arguments{
\item{file_path}{the path to the IEA data file (optional)}

\item{iea_df}{a data frame containing IEA data. Default is \code{IEATools::load_tidy_iea_df(file_path)}.}

\item{side}{refers to the "\code{Consumptiion}" or "\code{Supply}" side of Production, Transformation processes, or Energy industry own use.
One of "\code{Consumption}" or "\code{Supply}". Default is "\code{Consumption}".}

\item{flow_aggregation_point}{the flow aggregation point column in \code{iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{production}{the string indicating the production flow. Default is "\code{Production}".}

\item{transformation_processes}{the string indicating the transformation process stage. Default is "\verb{Transformation processes}".}

\item{eiou}{the string indicating the energy industry own use flow. Default is "\verb{Energy industry own use}".}

\item{stage}{the string indicating the stage for the analysis. One of \code{production}, \code{transformation_processes}, or \code{eiou}.
Default is \code{production}.}

\item{e_dot}{the energy flow rate column in \code{iea_df}.  Default is "\code{E.dot}".}

\item{flow}{the flow column in \code{iea_df}. Default is "\code{Flow}".}

\item{product}{the product column in \code{iea_df}.  Default is "\code{Product}".}
}
\value{
a list of string vectors
}
\description{
Sometimes, it is helpful to know all types of energy supplied or consumed by Production, Transformation processes, or Energy industry own use.
This function (optionally) reads an IEA data file or loads an IEA data frame and builds a named list of energy types
supplied or consumed by Production, Transformation processes, or Energy industry own use.
}
\details{
The names in the returned list are the Production, Transformation processes, or Energy industry own use industries in \code{iea_df}.
The items in the returned list are vectors of energy types produced or consumed by the corresponding industries.
}
\examples{
prod_tp_eiou_energy_carriers()
}
