% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocate.R
\name{reallocate_statistical_differences}
\alias{reallocate_statistical_differences}
\title{Reallocate Statistical differences}
\usage{
reallocate_statistical_differences(
  .sutmats = NULL,
  stat_diffs = IEATools::tfc_compare_flows$statistical_differences,
  R = IEATools::psut_cols$R,
  U = IEATools::psut_cols$U,
  U_feed = IEATools::psut_cols$U_feed,
  U_eiou = IEATools::psut_cols$U_eiou,
  r_eiou = IEATools::psut_cols$r_eiou,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  R_colname = IEATools::psut_cols$R,
  U_colname = IEATools::psut_cols$U,
  U_feed_colname = IEATools::psut_cols$U_feed,
  U_eiou_colname = IEATools::psut_cols$U_eiou,
  r_eiou_colname = IEATools::psut_cols$r_eiou,
  V_colname = IEATools::psut_cols$V,
  Y_colname = IEATools::psut_cols$Y,
  prime_suffix = "_prime",
  tol = 1e-06
)
}
\arguments{
\item{.sutmats}{A data frame of PSUT matrices,
most likely the result of \code{\link[=prep_psut]{prep_psut()}}.}

\item{stat_diffs}{The name of the row in \strong{R} and columns in \strong{U_feed} and \strong{Y}
for Statistical differences.
Default is \code{IEATools::tfc_compare_flows$statistical_differences}.}

\item{R, U, U_feed, U_eiou, r_eiou, V, Y}{Matrices or names of columns of matrices in \code{.sutmats}.
See \link{psut_cols}.}

\item{R_colname, U_colname, U_feed_colname, U_eiou_colname, r_eiou_colname, V_colname, Y_colname}{Names of columns of matrices in \code{.sutmats}.
See \link{psut_cols}.}

\item{prime_suffix}{The string suffix for new versions of matrices with reallocated
statistical differences.
Default is "_prime".}

\item{tol}{The tolerance for energy imbalance.
Default is \code{1e-6}.}
}
\value{
A version of \code{.sutmats} in which energy consumption by "Statistical differences"
is reallocated to other Industries in proportion to their energy consumption.
}
\description{
The IEA data include "Statistical differences".
In some cases, it is desirable to reallocate Statistical differences
in proportion to the non-zero consumption of each
energy carrier in other industries.
This function performs that reallocation
and should be called after the PSUT matrices have been formed,
most likely by calling \code{\link[=prep_psut]{prep_psut()}}.
}
\details{
Statistical differences can be found in either the \strong{R} or \strong{Y} matrix.
Both are reallocated to the \strong{Y} and \strong{U} matrices in proportion.
The steps are:
\enumerate{
\item For those rows with no other consumption in \strong{U} or \strong{Y},
move \strong{Y} Statistical differences to \strong{R}  by subtraction.
\item Reallocate negative Statistical differences in \strong{R} to
\strong{R} and \strong{V} using \code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
\item Move remaining (positive) Statistical differences found in
\strong{R} to the \strong{Y} matrix by subtraction.
\item Reallocate Statistical differences in the \strong{Y} matrix to the \strong{Y} and \strong{U}
matrices using \code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
}

Internally, the \strong{R} and \strong{V} matrices are added before calling
\code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
\strong{R} and \strong{V} are split again prior to returning.
Similarly, the \strong{Y} and \strong{U} matrices are added before calling
\code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
\strong{Y} and \strong{U} are split again prior to returning.

Energy balance is checked both
prior to reallocating statistical differences
and after reallocating statistical differences.
Imbalances beyond \code{tol} cause an error.

Note that most functions in \code{IEATools} operate on
tabular IEA data.
However, this function assumes IEA data have already been
converted to matrix (PSUT) format,
most likely with \code{\link[=prep_psut]{prep_psut()}}.
}
\examples{
R <- matrix(c(98, 0, 
               0, 50, 
               2, 0), 
            byrow = TRUE, nrow = 3, ncol = 2, 
            dimnames = list(c("Resources [of Coal]", 
                              "Resources [of Prod C]",
                              "Statistical differences"), 
                            c("Coal [from Resources]", "Prod C"))) |> 
  matsbyname::setrowtype("Industry") |> matsbyname::setcoltype("Product")
R
U <- matrix(c(100, 
              2), 
            byrow = TRUE, nrow = 2, ncol = 1, 
            dimnames = list(c("Coal [from Resources]", "Electricity"), 
                            c("Mapep"))) |> 
  matsbyname::setrowtype("Product") |> matsbyname::setcoltype("Industry")
U
V <- matrix(40, 
            byrow = TRUE, nrow = 1, ncol = 1, 
            dimnames = list(c("Mapep"), c("Electricity"))) |> 
  matsbyname::setrowtype("Industry") |> matsbyname::setcoltype("Product")
V
Y <- matrix(c(20, 10, 8, 
               0, 0, 50), 
            byrow = TRUE, nrow = 2, ncol = 3, 
            dimnames = list(c("Electricity", "Prod C"), 
                            c("Industry 1", "Industry 2", 
                            "Statistical differences"))) |> 
  matsbyname::setrowtype("Product") |> matsbyname::setcoltype("Industry")
Y
r_eiou <- matrix(1, 
                 byrow = TRUE, nrow = 1, ncol = 1, 
                 dimnames = list("Electricity", "Mapep")) |> 
  matsbyname::setrowtype("Product") |> matsbyname::setcoltype("Industry")
r_eiou
U_EIOU <- matsbyname::hadamardproduct_byname(U, r_eiou)
U_EIOU
U_feed <- matsbyname::difference_byname(U, U_EIOU)
U_feed
reallocate_statistical_differences(R = R, 
                                   U = U, 
                                   U_feed = U_feed, 
                                   U_eiou = U_EIOU, 
                                   r_eiou = r_eiou, 
                                   V = V,
                                   Y = Y)
}
