% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocate.R
\name{reallocate_statistical_differences}
\alias{reallocate_statistical_differences}
\title{Reallocate Statistical differences}
\usage{
reallocate_statistical_differences(
  .sutmats,
  stat_diffs = IEATools::tfc_compare_flows$statistical_differences,
  R = IEATools::psut_cols$R,
  U_feed = IEATools::psut_cols$U_feed,
  U_eiou = IEATools::psut_cols$U_eiou,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  prime_suffix = "_prime",
  R_prime = paste0(R, prime_suffix),
  U_feed_prime = paste0(U_feed, prime_suffix),
  U_eiou_prime = paste0(U_eiou, prime_suffix),
  Y_prime = paste0(Y, prime_suffix)
)
}
\arguments{
\item{.sutmats}{A data frame of PSUT matrices,
most likely the result of \code{\link[=prep_psut]{prep_psut()}}.}

\item{stat_diffs}{}
}
\value{
A version of \code{.sutmats} in which energy consumption by "Statistical differences"
is reallocated to other Industries in proportion to their energy consumption.
}
\description{
The IEA data include "Statistical differences".
In some cases, it is desirable to reallocate Statistical differences
in proportion to the non-zero consumption of each
energy carrier in other industries.
This function performs that reallocation
and should be called after the PSUT matrices have been formed,
most likely by calling \code{\link[=prep_psut]{prep_psut()}}.
}
\details{
Statistical differences can be found in either the \strong{R} or \strong{Y} matrix.
Both are reallocated to the \strong{U_EIOU} matrix.
The steps are:
\itemize{
\item Move Statistical differences found in the \strong{R} matrix to the \strong{Y} matrix by subtraction.
\item Reallocate Statistical differences in the \strong{Y} matrix to the \strong{U_EIOU} matrix using \code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}.
}

This function calls \code{\link[matsbyname:reallocate_byname]{matsbyname::reallocate_byname()}}
internally to perform the reallocation.
}
