% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_to_tidy.R
\name{remove_agg_memo_flows}
\alias{remove_agg_memo_flows}
\title{Title}
\usage{
remove_agg_memo_flows(.iea_df, flow = "Flow", product = "Product",
  agg_flows = c("Total primary energy supply", "Total final consumption",
  "Transformation processes", "Energy industry own use", "Industry",
  "Transport", "Other", "Non-energy use"),
  memo_flow_prefixes = c("Memo: ", "Electricity output (GWh)",
  "Heat output"), memo_product_prefix = "Memo: ")
}
\arguments{
\item{.iea_df}{}

\item{flow}{}

\item{agg_flows}{}

\item{memo_flow_prefixes}{}
}
\description{
Title
}
\examples{
file.path("extdata", "GH-ZA-ktoe-Extended-Energy-Balances-sample.csv") \%>\% 
  system.file(package = "IEAData") \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  remove_agg_memo_flows()
}
