% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renewable_products}
\alias{renewable_products}
\title{Renewable products}
\format{A string vector with 23 entries}
\usage{
renewable_products
}
\description{
A string vector containing names of products classified by the IEA as renewables.
}
\keyword{datasets}
