% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secondary_coal_products}
\alias{secondary_coal_products}
\title{Secondary coal products}
\format{
A string vector with 9 entries
}
\usage{
secondary_coal_products
}
\description{
A string vector containing names of products classified by the IEA as \code{coal_and_coal_products} that are not \code{primary_coal_products}.
}
\examples{
secondary_coal_products
}
\keyword{datasets}
