% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_all}
\alias{specify_all}
\title{Specify all industries}
\usage{
specify_all(.tidy_iea_df)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}
}
\value{
an enhanced and corrected version of \code{.tidy_iea_df}
that is ready for physical supply-use table (PSUT) analysis.
}
\description{
This is a convenience function.
It bundles several others:
\enumerate{
\item \code{\link[=specify_primary_production]{specify_primary_production()}}
\item \code{\link[=specify_production_to_resources]{specify_production_to_resources()}}
\item \code{\link[=specify_tp_eiou]{specify_tp_eiou()}}
\item \code{\link[=specify_interface_industries]{specify_interface_industries()}}
\item \code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}}
}
}
\details{
Each bundled function is called in turn using default arguments.
See examples for two ways to achieve the same result.
}
\examples{
# Simple
load_tidy_iea_df() \%>\% 
  specify_all()
# Complicated
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_production_to_resources() \%>\% 
  specify_tp_eiou() \%>\% 
  specify_interface_industries() \%>\% 
  tp_sinks_to_nonenergy()
}
