% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_all}
\alias{specify_all}
\title{Specify all industries}
\usage{
specify_all(
  .tidy_iea_df,
  split_own_use_elect_chp_heat_using_shares_of = c("input", "output"),
  route_non_specified_eiou = TRUE,
  route_non_specified_tp = TRUE
)
}
\arguments{
\item{.tidy_iea_df}{A tidy data frame containing IEA extended energy balance data}

\item{split_own_use_elect_chp_heat_using_shares_of}{Indicates whether the input or outputs to
Main activity producer plants should be use for
splitting the Own use in electricity, CHP and heat plants
EIOU flow. Default is "input".}

\item{route_non_specified_eiou}{Boolean stating whether non-specified EIOU flows should be routed to existing industries
Default is TRUE.}

\item{route_non_specified_tp}{Boolean stating whether non-specified transformation processes flows should be routed to existing industries
Default is TRUE.}
}
\value{
An enhanced and corrected version of \code{.tidy_iea_df}
That is ready for physical supply-use table (PSUT) analysis.
}
\description{
This is a convenience function.
It bundles several others:
\enumerate{
\item \code{\link[=specify_primary_production]{specify_primary_production()}}
\item \code{\link[=specify_tp_eiou]{specify_tp_eiou()}}
\item \code{\link[=specify_interface_industries]{specify_interface_industries()}}
\item \code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}}
}
}
\details{
Each bundled function is called in turn using default arguments.
See examples for two ways to achieve the same result.
}
\examples{
# Simple
load_tidy_iea_df() \%>\% 
  specify_all()
# Complicated
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_tp_eiou() \%>\% 
  specify_interface_industries() \%>\% 
  tp_sinks_to_nonenergy()
}
