% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_interface_industries}
\alias{specify_interface_industries}
\title{Specify interface industries}
\usage{
specify_interface_industries(
  .tidy_iea_df,
  flow = "Flow",
  int_industries = list(IEATools::interface_industries, "Resources", "Manufacture"),
  product = "Product",
  notation = RCLabels::of_notation
)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}.  Default is "\code{Flow}".}

\item{int_industries}{a string vector of industries involved in exchanges with other countries,
bunkers, or stock changes. Default is \code{IEATools::interface_industries}.}

\item{product}{the name of the product column in \code{.tidy_iea_df}.  Default is "\code{Product}".}

\item{notation}{a list of specification notations. Default is \code{RCLabels::of_notation}.}
}
\value{
a modified version of \code{.tidy_iea_df} with specified interface industries
}
\description{
An interface industry is one that moves energy carriers into or out of a country.
When \code{Flow} is any of the interface industries, we need to be more specific.
If we don't separate these \code{Flow}s by \code{Product}, we run into trouble with
upstream swims (e.g., all \code{Product}s are produced even if only one is needed) and
embodied energy calculations (many types of energy are embodied, even if only one should be).
This function adds a suffix \verb{[of Product]} to each of these interface industries.
}
\details{
Note that "\code{Production}" also needs to be specified,
but that is accomplished in the \code{specify_primary_production()} and
\code{specify_production_to_resources()} functions.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_interface_industries()
}
