% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_interface_industries}
\alias{specify_interface_industries}
\title{Specify interface industries}
\usage{
specify_interface_industries(.tidy_iea_df, flow = "Flow",
  int_industries = IEATools::interface_industries, product = "Product")
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}.  Default is "\code{Flow}".}

\item{int_industries}{a string vector of industries involved in exchanges with other countries,
bunkers, or stock changes. Default is \link{interface_industries}.}

\item{product}{the name of the product column in \code{.tidy_iea_df}.  Default is "\code{Product}".}
}
\value{
a modified version of \code{.tidy_iea_df} with specified interface industries
}
\description{
An interface industry is one that moves energy resources in or out of a country.
When \code{Flow} is any of the interface industries, we need to be more specific.
If we don't separate these Flows, we run into trouble with
upstream swims (e.g., all \code{Product}s are produced even if only one is needed) and
embodied energy calculations (many types of energy are embodied, even if only one sould be).
This function adds a suffix \code{ (Product)} to each of these interface industries.
}
\details{
Note that "\code{Production}" also needs to be specified,
but that is accomplished in the \code{\link[=specify_primary_production]{specify_primary_production()}} and
\code{\link[=production_to_resources]{production_to_resources()}} functions.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_interface_industries()
}
