% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{specify_non_energy_use}
\alias{specify_non_energy_use}
\title{Specify Non-energy use when possible}
\usage{
specify_non_energy_use(
  .iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  unit = IEATools::iea_cols$unit,
  ledger_side = IEATools::iea_cols$ledger_side,
  last_stage = IEATools::iea_cols$last_stage,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  non_energy_use = IEATools::aggregation_flows$non_energy_use,
  non_energy_flows_industry_transformation_energy =
    IEATools::non_energy_flows$non_energy_use_industry_transformation_energy,
  memo_non_energy_use_in_industry_not_elsewhere_specified =
    IEATools::memo_non_energy_flows$memo_non_energy_use_in_industry_not_elsewhere_specified,
  non_energy_use_in_industry_not_elsewhere_specified = sub(pattern = paste0("^", memo),
    replacement = "", memo_non_energy_use_in_industry_not_elsewhere_specified),
  memo = IEATools::memo_aggregation_flow_prefixes$memo,
  memo_non_energy_flows_industry =
    IEATools::memo_non_energy_flows$memo_non_energy_use_in_industry,
  memo_non_energy_use_in = "Memo: Non-energy use in ",
  non_energy_use_in = sub(pattern = paste0("^", memo), replacement = "", x =
    memo_non_energy_use_in),
  total = IEATools::memo_aggregation_product_prefixes$total,
  .values = IEATools::template_cols$.values,
  .values_summarised = paste0(.values, "_summarised"),
  .diff = ".diff",
  tol = 1e-06
)
}
\arguments{
\item{.iea_df}{A data frame of IEA data,
created by \code{augment_iea_df()}.}

\item{country, year, method, energy_type, unit, ledger_side, last_stage, flow_aggregation_point, flow, product}{See \code{IEATools::iea_cols}.}

\item{non_energy_use}{See \code{IEATools::aggregation_flows}.}

\item{non_energy_flows_industry_transformation_energy}{See \code{IEATools::non_energy_flows}.}

\item{memo_non_energy_use_in_industry_not_elsewhere_specified}{See \code{IEATools::memo_non_energy_flows}.}

\item{non_energy_use_in_industry_not_elsewhere_specified}{Same as \code{memo_non_energy_use_in_industry_not_elsewhere_specified} without the \code{memo} prefix.}

\item{memo}{A string prefix for memo flows.
Default is \code{IEATools::memo_aggregation_flow_prefixes$memo}.}

\item{memo_non_energy_flows_industry}{See \code{IEATools::memo_non_energy_flows}.}

\item{memo_non_energy_use_in}{A prefix for specific Non-energy use flows.
Default is "Memo: Non-energy use in ".}

\item{non_energy_use_in}{Same as \code{memo_non_energy_use_in} without the \code{memo} prefix.}

\item{total}{See \code{IEATools::memo_aggregation_product_prefixes}.}

\item{.values}{An internal column name.
Default is \code{IEATools::template_cols$.values}.}

\item{.values_summarised}{An internal column name.
Default is \code{paste0(.values, "_summarised")}.}

\item{.diff}{An internal column name.
Default is ".diff".}

\item{tol}{The tolerance for differences from \code{0}.
Default is \code{1e-6}.}
}
\value{
\code{.iea_df} with specified \verb{Non-energy use}s, where possible.
}
\description{
For some countries and years, Non-energy use details are supplied
by "Memo:" fields.
We use those details if they exist.
This function assess the energy balance of
replacing "Non-energy use industry/transformation/energy" by
"Non-energy use in <\if{html}{\out{<specific industry>}}>"
rows.
If there is an imbalance, an attempt is made to adjust the imbalance.
}
\details{
Note that energy balance checks are \emph{not} performed on the incoming \code{.iea_df},
but that the specification process itself
ensures that the modifications, themselves, are balanced.
In other words, specifying Non-energy use flows
will not change the energy balance of the incoming \code{.iea_df} data frame.

This function does \emph{not} promise the resulting data frame
is internally consistent.
Rather, this function should be considered a step along the path to
creating a coherent data frame of IEA data.
Specifically,
this function does not specify aggregation or Memo: flows, so
coherence will be achieved again only after
\code{remove_agg_memo_flows()}
is called.
Note that \code{load_tidy_iea_df(specify_non_energy_flows = TRUE)}
correctly calls \code{remove_agg_memo_flows()} internally, so
using \code{load_tidy_iea_df()} is preferred to calling \code{specify_non_energy_use()} directly.
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  clean_iea_whitespace() \%>\%
  augment_iea_df() \%>\%
  specify_non_energy_use()
}
