% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_primary_production}
\alias{specify_primary_production}
\title{Specify primary production industries}
\usage{
specify_primary_production(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot,
  list_primary_coal_products = IEATools::primary_coal_products,
  list_primary_oil_products = IEATools::primary_oil_products,
  list_primary_gas_products = IEATools::primary_gas_products,
  production = IEATools::tpes_flows$production,
  coal_mines = IEATools::industry_flows$coal_mines,
  oil_extraction = IEATools::industry_flows$oil_extraction,
  gas_extraction = IEATools::industry_flows$natural_gas_extraction,
  liquefaction_regas = "Liquefaction (LNG) / regasification plants",
  liquefaction_regas_reassign = IEATools::industry_flows$natural_gas_extraction,
  transformation_processes = IEATools::aggregation_flows$transformation_processes,
  resources = IEATools::tpes_flows$resources,
  resource_products_notation = RCLabels::from_notation,
  resources_flow_notation = RCLabels::of_notation,
  manufacture = "Manufacture",
  manufacture_flow_notation = RCLabels::of_notation
)
}
\arguments{
\item{.tidy_iea_df}{An IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}.}

\item{ledger_side, flow, product, flow_aggregation_point}{See \code{IEATools::iea_cols}.}

\item{e_dot}{The name of the energy column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}

\item{list_primary_coal_products}{The list of primary coal products for which the production industry needs to be changed.
Default is \code{IEATools::primary_coal_products}.}

\item{list_primary_oil_products}{The list of primary oil products for which the production industry needs to be changed.
Default is \code{IEATools::primary_oil_products}.}

\item{list_primary_gas_products}{The list of primary gas products for which the production industry needs to be changed.
Default is \code{IEATools::primary_gas_products}.}

\item{production}{A string identifying production in the flow column. Default is "\code{Production}".}

\item{coal_mines}{The name of the new industry that produces primary coal products.
Default is \code{IEATools::industry_flows$coal_mines}.}

\item{oil_extraction}{The name of the new industry that produces primary oil and gas products.
Default is \code{IEATools::industry_flows$oil_extraction}.}

\item{gas_extraction}{The name of the new industry that produces primary oil and gas products.
Default is \code{IEATools::industry_flows$natural_gas_extraction}.}

\item{liquefaction_regas}{A string identifying liquefaction and regasification plants.
Default is "Liquefaction (LNG) / regasification plants".}

\item{liquefaction_regas_reassign}{A string identifying the industry to which EIOU into \code{liquefaction_regas} will be reassigned.
Default is "Natural gas extraction".}

\item{transformation_processes}{A string identifying transformation processes in the flow column of \code{.tidy_iea_df}.
Default is "Transformation processes".}

\item{resources}{A string identifying resource industries to be added to \code{.tidy_iea_df}.
Default is "\code{Resources}".}

\item{resource_products_notation}{The notation to be used for defining products coming from the new resource industries.
E.g., the Crude oil product will be called "Crude oil [from Resources]".
Default is \code{RCLabels::from_notation}.}

\item{resources_flow_notation}{The notation to be used for defining the new resource industries.
E.g., the Crude oil resource will be called "Resources [of Crude oil]".
Default is \code{RCLabels::of_notation}.}

\item{manufacture}{The name of the industries that convert resource-products \(inputs\) into actual products \(outputs\),
when a corresponding a corresponding industry does not exist by default in IEA data.
Default is "Manufacture".}

\item{manufacture_flow_notation}{The notation to be used for the newly created manufacture industries
\(each manufacturing industry is specified\) by the product it manufactures.
Default is \code{RCLabels::of_notation}.}
}
\value{
A \code{.tidy_iea_df} with adjusted production information for primary energy
for both coal and coal products and oil and gas extraction
}
\description{
The IEA extended energy balances include
some \code{Flow}s that identify \verb{Energy industry own use} (EIOU),
the consumption of energy by energy-producing industries.
But some primary production industries that receive EIOU do not produce anything.
For example, \verb{Coal mines} receive electricity
but there are no \verb{Coal mines} that produce coal.
Rather, the generic \code{Production} industry produces coal.
This function solves that problem by
replacing the generic \code{Production} industry with
specific industries.
}
\details{
By default, the following changes are made to \code{.tidy_iea_df}:
\enumerate{
\item Energy industry own use for \verb{Liquefaction (LNG) / regasification plants} is
reassigned to \verb{Oil and gas extraction}.
\item Each \code{Production} flow is replaced by a \verb{Resources [of Product]} flow,
which produces \verb{Product [from Resources]}.
\item For each \code{Production} flow, a manufacturing flow, that takes as input
the \verb{Product [from Resources]} supplied by the new \verb{Resources [of Product]},
and that produces the given \code{Product}, is added. The name of the manufacturing industry
is \verb{Coal mines} for \code{coal_and_coal_products}, \verb{Oil and gas extraction} for \code{oil_and_gas_products},
and \verb{Manufacture [of Product]} for all other products.
}

Users can specify other changes by adjusting the default argument values.

Be sure to call this function \emph{after} calling \code{augment_iea_df()} or
\code{load_tidy_iea_df()}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  add_psut_matnames() \%>\% 
  dplyr::filter(Flow == "Coal mines" | stringr::str_detect(Flow, "Resources")) \%>\% 
  select(-Method, -Last.stage, -Ledger.side, -Unit)
# EIOU by "Liquefaction (LNG) / regasification plants" is reassigned to "Oil and gas extraction"
data.frame(
  Flow.aggregation.point = c("Energy industry own use"),
  Flow = c("Liquefaction (LNG) / regasification plants"), 
  Product = c("Natural gas"),
  E.dot = c(-42), 
  stringsAsFactors = FALSE
) \%>\% 
  specify_primary_production()
}
