% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiou.R
\name{specify_production}
\alias{specify_production}
\title{Specify production industries}
\usage{
specify_production(.tidy_iea_data, flow = "Flow", product = "Product",
  production = "Production", coal_mines = "Coal mines",
  coals = coal_and_coal_products)
}
\arguments{
\item{.tidy_iea_data}{}

\item{flow}{}

\item{product}{}

\item{production}{}

\item{coal_mines}{}

\item{coals}{}
}
\description{
Specify production industries
}
