% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_production_to_resources}
\alias{specify_production_to_resources}
\title{Convert Production Flows to Resource Flows}
\usage{
specify_production_to_resources(
  .tidy_iea_df,
  flow = IEATools::iea_cols$flow,
  product = IEATools::iea_cols$product,
  production = IEATools::tpes_flows$production,
  resources = IEATools::tpes_flows$resources,
  notation = IEATools::from_notation
)
}
\arguments{
\item{.tidy_iea_df}{An IEA data frame whose columns have been renamed by \code{rename_iea_df_cols()}}

\item{flow}{The name of the flow column in \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$flow}.}

\item{product}{The name of the product column in \code{.tidy_iea_df}.
Default is \code{IEATools::iea_cols$product}.}

\item{production}{A string identifying production in the flow column.
Default is \code{IEATools::tpes_flows$production}.}

\item{resources}{A string identifying resource industries to be added to \code{.tidy_iea_df}.
Default is \code{IEATools::tpes_flows$resources}.}

\item{notation}{A list of specification notations.
Default is \code{IEATools::bracket_notation}.}
}
\value{
A \code{.tidy_iea_df} with \code{Production} changed to \verb{resources .resources_open product .resources_close} in the \code{flow} column
}
\description{
The IEA gives resource extraction in rows where the \code{Flow} is "\code{Production}".
This function changes the "\code{Production}" string
to "\verb{Resources [of Product]}",
where \code{product} is the name of the energy carrier for this resource.
}
\details{
This function should be called \emph{after} \code{specify_primary_production()},
which adjusts for energy industry own use
of some primary energy producing industries.
If this function is called first,
EIOU will not be accounted correctly.
}
\examples{
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_production_to_resources()
}
