% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{switch-notation}
\alias{switch-notation}
\alias{from_to_arrow_byname}
\alias{arrow_to_bracket_byname}
\alias{bracket_to_arrow_byname}
\title{Change specification notation}
\usage{
from_to_arrow_byname(m, margin = c(1, 2))

arrow_to_bracket_byname(m, margin = c(1, 2))

bracket_to_arrow_byname(m, margin = c(1, 2))
}
\arguments{
\item{m}{a single matrix or a list of matrices.}

\item{margin}{the margin over which the notation switch should be made:
\code{1} for rows, \code{2} for columns, or \code{c(1, 2)} (the default) for both rows and columns}
}
\value{
An object of same length as \code{m} with switched notation.
}
\description{
IEA data can be "specified" by adding additional information.
Three notations are useful: arrow, bracket, and from.
Arrow notation is from -> to, and
bracket notation is
destination [source].
These functions change matrix row/column names
among the notations.
\itemize{
\item \code{arrow_to_from_byname()} switches from arrow notation to from notation.
\item \code{from_to_arrow_byname()} switches from from notation to arrow notation.
\item \code{arrow_to_bracket_byname()} switches from arrow notation to bracket notation.
\item \code{bracket_to_arrow_byname()} switches from bracket notation to arrow notation.
}

Functions with the \verb{_byname} suffix behave like functions in the \code{matsbyname} package.
Specifically, they can work with a single matrix or a list of matrices supplied to the \code{m} argument.
}
\examples{
m <- matrix(c(1, 2,  
              3, 4), nrow = 2, ncol = 2, byrow = TRUE, 
            dimnames = list(c("a -> b", "c -> d"), c("f [e]", "h [g]")))
m
arrow_to_from_byname(m)
arrow_to_from_byname(m, margin = 2) # No changes expected.
from_to_arrow_byname(m)
arrow_to_bracket_byname(m)
bracket_to_arrow_byname(m)

}
