% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{switch-notation}
\alias{switch-notation}
\alias{switch_notation}
\alias{arrow_to_paren}
\alias{paren_to_arrow}
\title{Change specification notation}
\usage{
switch_notation(x, old_start, old_end, new_start, new_end)

arrow_to_paren(
  x,
  old_start = specify_notation$arrow,
  old_end = "",
  new_start = specify_notation$open,
  new_end = specify_notation$close
)

paren_to_arrow(
  x,
  old_start = specify_notation$open,
  old_end = specify_notation$close,
  new_start = specify_notation$arrow,
  new_end = ""
)
}
\arguments{
\item{x}{a vector of strings for which notation is to be switched}

\item{old_start}{the start of specification notation in \code{strings}}

\item{old_end}{the end of specification notation in \code{strings}}

\item{new_start}{the start of specification notation in output}

\item{new_end}{the end of specification notation in output}
}
\value{
An object of same length as \code{strings} with switched notation.
}
\description{
IEA data can be "specified" by adding additional information.
Two notations are useful: arrow and parenthetical.
Arrow notation is "from -> to",
and parenthetical notation is "to [from]".
These functions change a vector of strings between the specification notations.
\itemize{
\item \code{switch_notation()} is the worker function.
\item \code{arrow_to_paren()} switches from arrow to parenthetical specification notations.
\item \code{paren_to_arrow()} switches from parenthetical to arrow specification notations.
}
}
