% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_cols}
\alias{template_cols}
\title{Final-to-useful template column names}
\format{
A string list with 17 entries
\describe{
\item{ef_product}{The name of the final energy product column in final-to-useful templates.}
\item{machine}{The name of the machine column in final-to-useful templates.}
\item{eu_product}{The name of the useful energy product column in final-to-useful templates.}
\item{destination}{The name of the destination column in final-to-useful templates.}
\item{quantity}{The name of the quantity column in final-to-useful templates.}
\item{maximum_values}{The name of the maximum values column in final-to-useful templates.}
\item{c_}{The name of the allocation (\code{C_1}, \code{C_2}, \code{C_3}, etc.) columns in final-to-useful templates.}
\item{C}{The name of the allocation (\code{C}) matrix column.}
\item{e_dot_perc}{The name of the energy flow percentage column in final-to-useful templates.}
\item{e_dot_machine}{The name of the column representing energy flow into a machine in final-to-useful templates.}
\item{e_dot_machine_perc}{The name of the machine energy flow percentage column in final-to-useful templates.}
\item{eta_fu}{The name of the final-to-useful machine efficiency column in final-to-useful templates.}
\item{phi_u}{The name of the exergy-to-energy ratio column in final-to-useful templates.}
\item{c_source}{The name of a column containing country identifiers for the source of final-to-useful allocation information.}
\item{eta_fu_source}{The name of a column containing country identifiers for the source of final-to-useful efficiency information.}
\item{phi_source}{The name of a column containing country identifiers for the source of exergy-to-energy ratios for useful-stage energy.}
}
}
\usage{
template_cols
}
\description{
A string list containing named names of columns in templates
(either final-to-useful allocations or machine efficiency)
for final-to-useful analysis.
Items in the list provide default values for column and row name function arguments
throughout the \code{IEATools} package.
}
\details{
Some of these names may represent rows (instead of columns) at some stages of the final-to-useful analysis.
}
\examples{
template_cols
}
\keyword{datasets}
