% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_cols}
\alias{template_cols}
\title{Final-to-useful template column names}
\format{
A string list with 24 entries
\describe{
\item{ef_product}{The name of the final energy product column in final-to-useful templates.}
\item{machine}{The name of the machine column in final-to-useful templates.}
\item{eu_product}{The name of the useful energy product column in final-to-useful templates.}
\item{destination}{The name of the destination column in final-to-useful templates.}
\item{quantity}{The name of the quantity column in final-to-useful templates.}
\item{maximum_values}{The name of the maximum values column in final-to-useful templates.}
\item{C_eiou}{The name of the EIOU allocation rows in final-to-useful templates.}
\item{C_Y}{The name of the final demand allocation rows in final-to-useful templates.}
\item{C_perc}{The name of the percentage allocation rows in final-to-useful templates.}
\item{e_dot_max}{The name of the maximum E.dot column in final-to-useful templates.}
\item{e_dot_dest}{The name of the column representing the destination for energy flows in final-to-useful templates.}
\item{e_dot_perc}{The name of the energy flow percentage column in final-to-useful templates.}
\item{e_dot_machine}{The name of the column representing energy flow into a machine in final-to-useful templates.}
\item{e_dot_machine_tot}{The name of the column representing total energy flow into a machine in final-to-useful templates.}
\item{e_dot_machine_perc}{The name of the machine energy flow percentage column in final-to-useful templates.}
\item{e_dot_machine_max_perc}{The name of the machine energy flow maximum percentage column in final-to-useful templates.}
\item{eta_fu}{The name of the final-to-useful machine efficiency column in final-to-useful templates.}
\item{phi_pf}{The name of the exergy-to-energy ratio column various data frames templates.}
\item{phi_u}{The name of the useful exergy-to-energy ratio column in final-to-useful templates.}
\item{phi}{The name of the exergy-to-energy ratio column for primary, final, and useful stages.}
\item{c_source}{The name of a column containing the source of final-to-useful allocation information.}
\item{eta_fu__source}{The name of a column containing the source of final-to-useful efficiency data.}
\item{phi_source}{The name of a column containing the source of exergy-to-energy ratios.}
\item{.values}{The name of a temporary column that holds values.}
}
}
\usage{
template_cols
}
\description{
A string list containing named names of columns in templates
(either final-to-useful allocations or machine efficiency)
for final-to-useful analysis.
Items in the list provide default values for column and row name function arguments
throughout the \code{IEATools} package.
}
\details{
Some of these names may represent rows (instead of columns) at some stages of the final-to-useful analysis.
}
\examples{
template_cols
}
\keyword{datasets}
