% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{tp_sinks_to_nonenergy}
\alias{tp_sinks_to_nonenergy}
\title{Reassign Transformation process sinks to Non-energy use}
\usage{
tp_sinks_to_nonenergy(.tidy_iea_df, ledger_side = "Ledger.side",
  consumption = "Consumption",
  flow_aggregation_point = "Flow.aggregation.point",
  non_energy_flow_agg_point = "Non-energy use",
  transformation_processes = "Transformation processes",
  eiou = "Energy industry own use", flow = "Flow",
  non_energy_flow = "Non-energy use industry/transformation/energy",
  product = "Product", e_dot = "E.dot", grouping_vars = c("Method",
  "Last.stage", "Country", "Year", "Energy.type"))
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is "\code{Ledger.side}".}

\item{consumption}{a string identifying the consumption side of the ledger. Default is "\code{Consumption}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{non_energy_flow_agg_point}{the name of the aggregation point where transforamtion process sinks will be reassigned. Default is "\code{Non-energy use}".}

\item{transformation_processes}{a string that identifies transformation processes in the \code{flow_aggregation_point} column. Default is "\code{Transformation processes}".}

\item{eiou}{a string that identifies energy industry own use in the \code{flow_aggregation_point} column. Default is "\code{Energy industry own use}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{non_energy_flow}{a sting identifying non-energy flows. Default is "\code{Non-energy use industry/transformation/energy}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{e_dot}{the name of the energy rate column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}

\item{grouping_vars}{a string vector of column names by which \code{.tidy_iea_df} will be grouped before finding transformation sinks. Default is \code{c("Method", "Last.stage", "Country", "Year", "Energy.type")}.}
}
\value{
\code{.tidy_iea_df} with energy sunk in Transformation processes sinks reassigned to Non-energy use
}
\description{
Transformation processes that consume energy without producing any energy are called
"transformation process sinks".
See \code{\link[=tp_sinks_sources]{tp_sinks_sources()}} for information about why transformation process sinks are problematic.
This function reclassifies energy flowing into transformation process sinks
as \code{non_energy_flow}, by default "Non-energy use in industry/transformation/energy".
}
\examples{
library(dplyr)
DF <- data.frame(
  Ledger.side = c("Supply", "Supply", "Supply", "Consumption"),
  Flow.aggregation.point = c("Transformation processes", 
                             "Transformation processes", 
                             "Transformation processes", 
                             "Non-energy use"), 
  Flow = c("Automobiles", "Automobiles", "Furnaces", 
           "Non-energy use industry/transformation/energy"),
  Product = c("Petrol", "MD", "Coal", "Coal"),
  E.dot = c(-1, 1, -2, 8), 
  stringsAsFactors = FALSE
) \%>\% 
  mutate(
    Method = "PCM", 
    Last.stage = "Final",
    Energy.type = "E",
    Country = "Bogus",
    Year = 1971
  )
DF
DF \%>\% 
  tp_sinks_to_nonenergy()
}
