% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{tp_sinks_to_nonenergy}
\alias{tp_sinks_to_nonenergy}
\title{Reassign Transformation process sinks to Non-energy use}
\usage{
tp_sinks_to_nonenergy(
  .tidy_iea_df,
  ledger_side = IEATools::iea_cols$ledger_side,
  consumption = IEATools::ledger_sides$consumption,
  flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
  non_energy_flow_agg_point = IEATools::tfc_flows$non_energy_use,
  transformation_processes = IEATools::tfc_compare_flows$transformation_processes,
  eiou = IEATools::tfc_compare_flows$energy_industry_own_use,
  flow = IEATools::iea_cols$flow,
  non_energy_flow =
    IEATools::non_energy_flows$non_energy_use_industry_transformation_energy,
  product = IEATools::iea_cols$product,
  e_dot = IEATools::iea_cols$e_dot
)
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{ledger_side}{the name of the ledger side column in \code{.tidy_iea_df}. Default is "\code{Ledger.side}".}

\item{consumption}{a string identifying the consumption side of the ledger. Default is "\code{Consumption}".}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{non_energy_flow_agg_point}{the name of the aggregation point where transformation process sinks will be reassigned. Default is "\verb{Non-energy use}".}

\item{transformation_processes}{a string that identifies transformation processes in the \code{flow_aggregation_point} column. Default is "\verb{Transformation processes}".}

\item{eiou}{a string that identifies energy industry own use in the \code{flow_aggregation_point} column. Default is "\verb{Energy industry own use}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{non_energy_flow}{a sting identifying non-energy flows. Default is "\verb{Non-energy use industry/transformation/energy}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{e_dot}{the name of the energy rate column in \code{.tidy_iea_df}. Default is "\code{Edot}".}
}
\value{
\code{.tidy_iea_df} with energy sunk in Transformation processes sinks reassigned to Non-energy use
}
\description{
Transformation processes that consume energy without producing any energy are called
"transformation process sinks".
See \code{\link[=tp_sinks_sources]{tp_sinks_sources()}} for information about why transformation process sinks are problematic.
This function reclassifies energy flowing into transformation process sinks
as \code{non_energy_flow}, by default "Non-energy use in industry/transformation/energy".
}
\examples{
library(dplyr)
DF <- data.frame(
  LedgerSide = c("Supply", "Supply", "Supply", "Consumption"),
  FlowAggregationPoint = c("Transformation processes", 
                             "Transformation processes", 
                             "Transformation processes", 
                             "Non-energy use"), 
  Flow = c("Automobiles", "Automobiles", "Furnaces", 
           "Non-energy use industry/transformation/energy"),
  Product = c("Petrol", "MD", "Coal", "Coal"),
  Edot = c(-1, 1, -2, 8), 
  stringsAsFactors = FALSE
) \%>\% 
  mutate(
    Method = "PCM", 
    LastStage = "Final",
    EnergyType = "E",
    Country = "Bogus",
    Year = 1971
  )
DF
DF \%>\% 
  tp_sinks_to_nonenergy()
}
