% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpes_flows}
\alias{tpes_flows}
\title{Total primary energy supply flows}
\format{A string vector}
\usage{
tpes_flows
}
\description{
A string vector containing names of Total primary energy supply flows in the IEA extended energy balances database.
}
\keyword{datasets}
