% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{transformation_sinks}
\alias{transformation_sinks}
\title{Find transformation sinks}
\usage{
transformation_sinks(.tidy_iea_df,
  flow_aggregation_point = "Flow.aggregation.point",
  transformation_processes = "Transformation processes",
  eiou = "Energy industry own use", flow = "Flow",
  product = "Product", e_dot = "E.dot")
}
\arguments{
\item{.tidy_iea_df}{a tidy data frame containing IEA extended energy balance data}

\item{flow_aggregation_point}{the name of the flow aggregation point column in \code{.tidy_iea_df}. Default is "\code{Flow.aggregation.point}".}

\item{transformation_processes}{a string that identifies transformation processes in the \code{flow_aggregation_point} column. Default is "\code{Transformation processes}".}

\item{eiou}{a string that identifies energy industry own use in the \code{flow_aggregation_point} column. Default is "\code{Energy industry own use}".}

\item{flow}{the name of the flow column in \code{.tidy_iea_df}. Default is "\code{Flow}".}

\item{product}{the name of the product column in \code{.tidy_iea_df}. Default is "\code{Product}".}

\item{e_dot}{the name of the energy rate column in \code{.tidy_iea_df}. Default is "\code{E.dot}".}
}
\value{
the grouping columns of \code{.tidy_iea_df} plus the \code{flow} column,
with one row for each industry that is a transformation sink.
Industries that are transformation sinks are named in the \code{flow} column.
}
\description{
In the IEA extended energy balance data,
transformation processes ought to both consume and produce energy.
But some transformation processes consume energy without producing any energy.
Those transformation processes can be called "transformation sinks."
This function finds and identifies industries that act as transformation sinks.
}
\details{
It is important to identify transformation sinks,
because they cause two problems for physical supply-use table (PSUT) analysis.
First, when swimming upstream, a PSUT analysis cannot "see" the sunk energy carriers,
because they have no downstream effects.
Thus, upstream swims cannot conserve energy.
Second, when calculating embodied energy for each downstream energy carrier,
the sunk energy carriers cannot be embodied in any final demand energy carriers.
Thus, embodied energy calculations cannot conserve energy.

Transformation sinks are identified by the following algorithm:
\enumerate{
\item Identify (per group in \code{.tidy_iea_df}) all \code{Transformation processes} that consume energy (negative value for \code{E.dot}).
Energy consumption can be for the transformation process itself or for Energy industry own use.
\item Identify (per group in \code{.tidy_iea_df}) all \code{Transformation processes} that produce energy (positive value for \code{E.dot}).
\item Take the set difference between the two (producers less consumers).
The set difference is the list of transformation sinks.
}

\code{\link[=transformation_sinks]{transformation_sinks()}} is a function not unlike \code{\link[dplyr:summarise]{dplyr::summarise()}};
it returns a summary containing grouping variables and industries that are transformation sinks.
So be sure to \code{\link[dplyr:group_by]{dplyr::group_by()}} the \code{.tidy_iea_df} data frame \emph{before} calling this function.
Typical grouping variables are \code{Method}, \code{Last.stage}, \code{Country}, \code{Year}, \code{Energy.type}.
Don't group on \code{Flow.aggregation.point}, because energy from different aggregation points
(\code{Energy industry own use} and \code{Transformation processes}) flows into each machine.
Don't group on \code{Flow}, \code{Product}, or \code{E.dot}, either.
If groups are not set,
\code{flow}s will be analyzed together, possibly leading to missed transformation sinks.

The various \code{specify_*()} functions should also be called \emph{before} calling \code{\link[=transformation_sinks]{transformation_sinks()}}.
The \code{specify_*()} functions clean up the IEA data, ensuring that energy is routed to the right places.

Note that this function only identifies the problem of transformation sinks;
it does not fix the problem.
To solve the problem of transformation sinks,
see the \code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}} function.
\code{\link[=tp_sinks_to_nonenergy]{tp_sinks_to_nonenergy()}} uses the output of \code{\link[=transformation_sinks]{transformation_sinks()}}
to route energy consumed by transformation sinks to \code{Non-energy use industry/transformation/energy}.
}
\examples{
library(dplyr)
load_tidy_iea_df() \%>\% 
  specify_primary_production() \%>\% 
  specify_interface_industries() \%>\% 
  specify_tp_eiou() \%>\% 
  group_by(Method, Last.stage, Country, Year, Energy.type) \%>\% 
  transformation_sinks()
}
