% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_flows}
\alias{transport_flows}
\title{Transport flows}
\format{A string vector with 8 entries}
\usage{
transport_flows
}
\description{
A string vector containing names of \code{Transport} \code{Flow}s in the IEA extended energy balances database.
}
\keyword{datasets}
