% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{use_iso_countries}
\alias{use_iso_countries}
\title{Replace country names with 3-letter ISO abbreviations}
\usage{
use_iso_countries(
  .iea_df,
  country = "Country",
  iea_china = "People's Republic of China",
  ieatools_china = "CHN",
  iea_hk = "Hong Kong (China)",
  ieatools_hk = "HKG",
  iea_world_marine_bunkers = "World marine bunkers",
  ieatools_wmb = "WMB",
  iea_world_aviation_bunkers = "World aviation bunkers",
  ieatools_wab = "WAB"
)
}
\arguments{
\item{.iea_df}{A data frame containing a \code{country} column}

\item{country}{The name of the country column in \code{.iea_df}. Default is "Country".}

\item{ieatools_china}{The 3-letter string that replaces \code{iea_china_HK}.
Default is "CHN".}

\item{iea_china_HK}{The IEA string for China and Hong Kong.
Default is "China (P.R. of China and Hong Kong, China)".}

\item{world_marine_bunkers}{A string specifying the full name for the "country" called World marine bunkers.
Default is "World marine bunkers".}

\item{wmb}{A 3-letter code for World marine bunkers. Default is "WMB".}

\item{world_aviation_bunkers}{A string specifying the full name for the "country" called World aviation bunkers.
Default is "World aviation bunkers".}

\item{wab}{A 3-letter code for World aviation bunkers. Default is "WAB".}
}
\value{
\code{.iea_df} with 3-letter ISO country abbreviations in the \code{country} column.
}
\description{
The IEA uses full country names, but it is more concise to use the 3-letter ISO abbreviations.
This function replaces the full country names with ISO abbreviations where possible.
}
\details{
Special cases are considered where
IEA extended energy balance data country names
differ from
\code{countrycode} package country names.\tabular{lll}{
   IEA name \tab \code{countrycode} name (names) \tab \code{use_iso_countries()} decision \cr
   "People's Republic of China" \tab "China" \tab "CHN" \cr
}
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  use_iso_countries()
}
