% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{use_iso_countries}
\alias{use_iso_countries}
\title{Replace country names with 3-letter ISO abbreviations}
\usage{
use_iso_countries(
  .iea_df,
  override_df = data.frame(a = c("CHN", "HKG", "WMB", "WAB"), b =
    c("People's Republic of China", "Hong Kong (China)", "World marine bunkers",
    "World aviation bunkers")) \%>\% magrittr::set_names(c(pfu_code, iea_name)),
  country = IEATools::iea_cols$country,
  pfu_code = IEATools::country_concordance_cols$pfu_code,
  iea_name = IEATools::country_concordance_cols$iea_name
)
}
\arguments{
\item{.iea_df}{A data frame containing a \code{country} column.}

\item{override_df}{A data frame containing columns named \code{pfu_code} and \code{iea_name}.}

\item{country}{The name of the country column in \code{.iea_df}. Default is "Country".}

\item{pfu_code, iea_name}{See \code{IEATools::countryconcordance_cols}.}
}
\value{
\code{.iea_df} with 3-letter ISO country abbreviations in the \code{country} column.
}
\description{
The IEA uses full country names, but it is more concise to use the 3-letter ISO abbreviations.
This function replaces the full country names with ISO abbreviations where possible.
}
\details{
Internally, the \code{countrycode} package is used for ISO codes where possible.
Optionally, an \code{override_df} can be specified.
\code{override_df} should contain columns \code{pfu_code} (for 3-letter ISO codes) and \code{iea_name} for country names.

If neither the \code{countrycode} package nor \code{override_df} contain a matching \code{country} name,
the \code{country} column is left untouched.

By default, special cases are considered via the \code{override_df} argument.
The default value of \code{override_df} sets codes for China and Hong Kong,
as well as World X bunkers.
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  use_iso_countries()
}
