% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{use_iso_countries}
\alias{use_iso_countries}
\title{Replace country names with 3-letter ISO abbreviations}
\usage{
use_iso_countries(.iea_df, country = "Country")
}
\arguments{
\item{.iea_df}{a data frame containing a \code{country} column}

\item{country}{the name of the country column in \code{.iea_df}. Default is "Country".}
}
\value{
\code{.iea_df} with 3-letter ISO country abbreviations
}
\description{
The IEA uses full country names, but it is more concise to use the 3-letter ISO abbreviations.
This function replaces the full country names with ISO abbreviations where possible.
}
\examples{
sample_iea_data_path() \%>\% 
  iea_df() \%>\%
  rename_iea_df_cols() \%>\% 
  use_iso_countries()
}
