% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_eta_fu_template}
\alias{write_eta_fu_template}
\title{Write a final-to-useful machine efficiencies template to a file}
\usage{
write_eta_fu_template(.eta_fu_template, path,
  eta_fu_tab_name = "FU etas", overwrite_file = FALSE,
  overwrite_fu_eta_tab = FALSE, eta_fu = "eta.fu",
  header_row_font_color = "#FFFFFF",
  header_row_shading_color = "#5A80B8", eta_row_font_color = "#8C87A0",
  eta_row_shading_color = "#E3DFEB", quantity = "Quantity",
  e_dot_machine_max_perc = "E.dot_machine_max [\%]",
  .rownum = ".rownum")
}
\arguments{
\item{.eta_fu_template}{a template for final-to-useful energy efficiency values, generated by \code{eta_fu_template()}.}

\item{path}{the file path where the template will be written}

\item{eta_fu_tab_name}{the name of the final-to-useful efficiency tab. Default is "\code{FU etas}".}

\item{overwrite_file}{a logical telling whether to overwrite a file, if it already exists. Default is \code{FALSE}.}

\item{overwrite_fu_eta_tab}{a logical telling whether to overwrite the final-to-useful efficiency tab, if it already exists. Default is \code{FALSE}.}

\item{eta_fu}{the name of the final-to-useful efficiency rows in \code{.eta_fu_template}. Default is "\code{eta.fu}".}

\item{header_row_font_color}{a hex string representing the font color for the header row in the Excel file that is written by this function.
Default is "\code{#FFFFFF}", white.}

\item{header_row_shading_color}{a hex string representing the shading color for the header row in the Excel file that is written by this function.
Default is "\code{#5A80B8}", medium blue.}

\item{eta_row_font_color}{a hex string representing the font color for \code{eta} rows in the Excel file that is written by this function.
Default is "\code{#8C87A0}", a dark purple color.}

\item{eta_row_shading_color}{a hex string representing the shading color for \code{eta} rows in the Excel file that is written by this function.
Default is "\code{#E3DFEB}", a light purple color.}

\item{quantity}{the name of the quantity column in \code{.eta_fu_template}. Default is "\code{Quantity}".}

\item{e_dot_machine_max_perc}{the name of the rows that give maximum percentages. Default is "\code{E.dot_machine_max [\%]}".}

\item{.rownum}{the name of a temporary column containing row numbers. Default is "\code{.rownum}".}
}
\value{
\code{path}
}
\description{
The template should be created by \code{eta_fu_template()}.
}
\examples{
f <- tempfile(fileext = ".xlsx")
load_fu_allocation_data() \%>\% 
  eta_fu_template() \%>\% 
  write_eta_fu_template(f)
if (file.exists(f)) {
  file.remove(f)
}
}
