% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_eta_fu_template}
\alias{write_eta_fu_template}
\title{Write a final-to-useful efficiencies template}
\usage{
write_eta_fu_template(.eta_fu_template, path,
  eta_fu_tab_name = "FU etas", overwrite_file = FALSE,
  overwrite_fu_eta_tab = FALSE, eta_fu = "eta.fu",
  eta_row_font_color = "#104273", eta_row_shading_color = "#B8D8F5",
  quantity = "Quantity", .rownum = ".rownum")
}
\arguments{
\item{path}{}

\item{eta_fu_tab_name}{}

\item{overwrite_file}{}

\item{eta_row_font_color}{a hex string representing the font color for \code{eta} rows in the Excel file that is written by this function.
Default is "\code{#104273}", a dark blue color.}

\item{eta_row_shading_color}{a hex string representing the shading color for \code{eta} rows in the Excel file that is written by this function.
Default is "\code{#B8D8F5}", a light blue color.}

\item{overwrite_eta_tab}{}
}
\description{
Write a final-to-useful efficiencies template
}
