% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_eta_fu_template}
\alias{write_eta_fu_template}
\title{Write a final-to-useful machine efficiencies template to a file}
\usage{
write_eta_fu_template(
  .eta_fu_template,
  path,
  eta_fu_tab_name = IEATools::fu_analysis_file_info$eta_fu_tab_name,
  overwrite_file = FALSE,
  overwrite_fu_eta_tab = FALSE,
  eta_fu = IEATools::template_cols$eta_fu,
  e_dot_machine = IEATools::template_cols$e_dot_machine,
  e_dot_machine_perc = IEATools::template_cols$e_dot_machine_perc,
  maximum_values = IEATools::template_cols$maximum_values,
  header_row_font_color = "#FFFFFF",
  header_row_shading_color = "#5A80B8",
  e_dot_machine_row_font_color = "#8C87A0",
  e_dot_machine_row_shading_color = "#E3DFEB",
  e_dot_machine_perc_row_font_color = "#8C87A0",
  e_dot_machine_perc_row_shading_color = "#E3DFEB",
  eta_row_font_color = "#000000",
  eta_row_shading_color = "#FFFFFF",
  phi_row_font_color = "#000000",
  phi_row_shading_color = "#FFFFFF",
  blank_shading_color = "#808080",
  quantity = IEATools::template_cols$quantity,
  e_dot_machine_max_perc = IEATools::template_cols$e_dot_machine_max_perc,
  .rownum = ".rownum"
)
}
\arguments{
\item{.eta_fu_template}{a template for final-to-useful energy efficiency values, generated by \code{eta_fu_template()}.}

\item{path}{the file path where the eta_fu template will be written}

\item{eta_fu_tab_name}{the name of the final-to-useful efficiency tab. Default is "FU etas".}

\item{overwrite_file}{a logical telling whether to overwrite a file, if it already exists. Default is \code{FALSE}.}

\item{overwrite_fu_eta_tab}{a logical telling whether to overwrite the final-to-useful efficiency tab, if it already exists. Default is \code{FALSE}.}

\item{eta_fu}{the name of the final-to-useful efficiency rows in \code{.eta_fu_template}. Default is "eta.fu".}

\item{e_dot_machine}{a string identifying energy flow into final-to-useful machines. Default is "Edot_machine".}

\item{e_dot_machine_perc}{a string identifying percentage of total final energy flowing into final-to-useful machines. Default is "E.dot_machine [\%]".}

\item{maximum_values}{a string identifying the maximum values column in the outgoing template. Default is \code{IEATools::template_cols$maximum_values}.}

\item{header_row_font_color}{a hex string representing the font color for the header row in the Excel file that is written by this function.
Default is "#FFFFFF", white.}

\item{header_row_shading_color}{a hex string representing the shading color for the header row in the Excel file that is written by this function.
Default is "#5A80B8", medium blue.}

\item{e_dot_machine_row_font_color}{a hex string representing the font color for rows of energy flows into final-to-useful machines in the Excel file that is written by this function.
Default is "#8C87A0", a dark purple color.}

\item{e_dot_machine_row_shading_color}{a hex string representing the shading color for rows of energy flows into final-to-useful machines in the Excel file that is written by this function.
Default is "#5A80B8", medium purple color.}

\item{e_dot_machine_perc_row_font_color}{a hex string representing the font color for rows of percentage of energy flows into final-to-useful machines in the Excel file that is written by this function.
Default is "#8C87A0", a dark purple color.}

\item{e_dot_machine_perc_row_shading_color}{a hex string representing the shading color for rows of percentage of energy flows into final-to-useful machines in the Excel file that is written by this function.
Default is "#5A80B8", medium purple color.}

\item{eta_row_font_color}{a hex string representing the font color for \code{eta_fu} rows in the Excel file that is written by this function.
Default is "#000000", black.}

\item{eta_row_shading_color}{a hex string representing the shading color for \code{eta_fu} rows in the Excel file that is written by this function.
Default is "#FFFFFF", white.}

\item{phi_row_font_color}{a hex string representing the font color for \code{phi_u} rows in the Excel file that is written by this function.
Default is "#000000", black.}

\item{phi_row_shading_color}{a hex string representing the shading color for \code{phi_u} rows in the Excel file that is written by this function.
Default is "#FFFFFF", white.}

\item{blank_shading_color}{a hex string representing the shading color for blank cells in the \code{maximum_values} column.
Default is "#808080".}

\item{quantity}{the name of the quantity column in \code{.eta_fu_template}. Default is "Quantity".}

\item{e_dot_machine_max_perc}{the name of the rows that give maximum percentages. Default is "E.dot_machine_max [\%]".}

\item{.rownum}{the name of a temporary column containing row numbers. Default is ".rownum".}
}
\value{
the \code{path} argument
}
\description{
The template should be created by \code{eta_fu_template()}.
}
\examples{
f <- tempfile(fileext = ".xlsx")
load_fu_allocation_data() \%>\% 
  eta_fu_template() \%>\% 
  write_eta_fu_template(f)
if (file.exists(f)) {
  file.remove(f)
}
}
